% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/305_exports.R
\docType{methods}
\name{abs,Expression-method}
\alias{abs,Expression-method}
\alias{abs}
\title{Absolute Value}
\usage{
\S4method{abs}{Expression}(x)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}
}
\value{
An \linkS4class{Expression} representing the absolute value of the input.
}
\description{
The elementwise absolute value.
}
\examples{
A <- Variable(2,2)
prob <- Problem(Minimize(sum(abs(A))), list(A <= -2))
result <- solve(prob)
result$value
result$getValue(A)
}
