% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solver.R
\name{accepts,ConicSolver,Problem-method}
\alias{accepts,ConicSolver,Problem-method}
\title{Can the problem be solved with a conic solver?}
\usage{
\S4method{accepts}{ConicSolver,Problem}(object, problem)
}
\arguments{
\item{object}{A \linkS4class{ConicSolver} object.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
Can the problem be solved with a conic solver?
}
