% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_qp_solver.R
\name{accepts,QpSolver,Problem-method}
\alias{accepts,QpSolver,Problem-method}
\title{Is this a QP problem?}
\usage{
\S4method{accepts}{QpSolver,Problem}(object, problem)
}
\arguments{
\item{object}{A \linkS4class{QpSolver} object.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
Is this a QP problem?
}
