% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{accepts}
\alias{accepts}
\title{Reduction Acceptance}
\usage{
accepts(object, problem)
}
\arguments{
\item{object}{A \linkS4class{Reduction} object.}

\item{problem}{A \linkS4class{Problem} to check.}
}
\value{
A logical value indicating whether the reduction can be applied.
}
\description{
Determine whether the reduction accepts a problem.
}
