% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/affine.R
\docType{class}
\name{+,Expression,missing-method}
\alias{+,Expression,missing-method}
\alias{+,Expression,Expression-method}
\alias{+,Expression,ConstVal-method}
\alias{+,ConstVal,Expression-method}
\alias{AddExpression-class}
\alias{AddExpression}
\alias{to_numeric,AddExpression-method}
\alias{size_from_args,AddExpression-method}
\alias{graph_implementation,AddExpression-method}
\title{The AddExpression class.}
\usage{
\S4method{+}{Expression,missing}(e1, e2)

\S4method{+}{Expression,Expression}(e1, e2)

\S4method{+}{Expression,ConstVal}(e1, e2)

\S4method{+}{ConstVal,Expression}(e1, e2)

\S4method{to_numeric}{AddExpression}(object, values)

\S4method{size_from_args}{AddExpression}(object)

\S4method{graph_implementation}{AddExpression}(object, arg_objs, size, data = NA_real_)
}
\arguments{
\item{e1, e2}{The \linkS4class{Expression} objects or numeric constants to add.}

\item{object}{An \linkS4class{AddExpression} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the sum of any number of expressions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: Sum all the values.

\item \code{size_from_args}: The size of the expression.

\item \code{graph_implementation}: The graph implementation of the expression.
}}

\section{Slots}{

\describe{
\item{\code{arg_groups}}{A \code{list} of \linkS4class{Expression}s and numeric data.frame, matrix, or vector objects.}
}}

