% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{AffAtom-class}
\alias{AffAtom-class}
\alias{AffAtom}
\alias{sign_from_args,AffAtom-method}
\alias{is_atom_convex,AffAtom-method}
\alias{is_atom_concave,AffAtom-method}
\alias{is_incr,AffAtom-method}
\alias{is_decr,AffAtom-method}
\alias{is_quadratic,AffAtom-method}
\alias{is_pwl,AffAtom-method}
\title{The AffAtom class.}
\usage{
\S4method{sign_from_args}{AffAtom}(object)

\S4method{is_atom_convex}{AffAtom}(object)

\S4method{is_atom_concave}{AffAtom}(object)

\S4method{is_incr}{AffAtom}(object, idx)

\S4method{is_decr}{AffAtom}(object, idx)

\S4method{is_quadratic}{AffAtom}(object)

\S4method{is_pwl}{AffAtom}(object)
}
\arguments{
\item{object}{An \linkS4class{AffAtom} object.}

\item{idx}{An index into the atom.}
}
\description{
This virtual class represents an affine atomic expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{sign_from_args}: The sign of the atom.

\item \code{is_atom_convex}: The atom is convex.

\item \code{is_atom_concave}: The atom is concave.

\item \code{is_incr}: The atom is weakly increasing in every argument.

\item \code{is_decr}: The atom is not weakly decreasing in any argument.

\item \code{is_quadratic}: Is every argument quadratic?

\item \code{is_pwl}: Is every argument piecewise linear?
}}

