% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{affine_prod}
\alias{affine_prod}
\title{Affine Product}
\usage{
affine_prod(x, y)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric constant representing the left-hand value.}

\item{y}{An \linkS4class{Expression} or numeric constant representing the right-hand value.}
}
\value{
An \linkS4class{Expression} representing the product of \code{x} and \code{y}.
}
\description{
The product of two affine expressions.
}
