% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/019_atom.R
\name{allow_complex,Atom-method}
\alias{allow_complex,Atom-method}
\title{Does the atom handle complex numbers?}
\usage{
\S4method{allow_complex}{Atom}(object)
}
\arguments{
\item{object}{an \linkS4class{Atom} object}
}
\description{
Does the atom handle complex numbers?
}
