% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{are_args_affine}
\alias{are_args_affine}
\title{Are the arguments affine?}
\usage{
are_args_affine(constraints)
}
\arguments{
\item{constraints}{A \linkS4class{Constraint} object.}
}
\value{
All the affine arguments in given constraints.
}
\description{
Are the arguments affine?
}
