% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/283_conic_solver.R
\name{as_block_diag_linear_operator}
\alias{as_block_diag_linear_operator}
\title{A wrapper for linear operators}
\usage{
as_block_diag_linear_operator(matrices)
}
\arguments{
\item{linear_op}{a sparse matrix or a function that takes one argument X}

\item{dim}{the dimensions}
}
\value{
a function that will perform the linear operation
}
\description{
A wrapper for linear operators
}
