% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{atom_domain}
\alias{atom_domain}
\title{Get Atom-Specific Domain Constraints}
\usage{
atom_domain(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
List of Constraint objects.
}
\description{
Get Atom-Specific Domain Constraints
}
\keyword{internal}
