% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{atoms}
\alias{atoms}
\title{Get the Atoms in an Expression}
\usage{
atoms(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
List of atom objects.
}
\description{
Get the Atoms in an Expression
}
\keyword{internal}
