% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{attributes_were_lowered}
\alias{attributes_were_lowered}
\title{Determine if the attributes on a variable were lowered}
\usage{
attributes_were_lowered(object)
}
\arguments{
\item{object}{A \linkS4class{Variable} object.}
}
\value{
A logical value indicating whether the attributes were lowered or not
}
\description{
Determine if the attributes on a variable were lowered
}
