% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{bisect,Problem-method}
\alias{bisect,Problem-method}
\title{Bisection on a one-parameter family of DCP problems.}
\usage{
\S4method{bisect}{Problem}(
  problem,
  solver = NULL,
  low = NA_real_,
  high = NA_real_,
  eps = 1e-06,
  verbose = FALSE,
  max_iters = 100,
  max_iters_interval_search = 100
)
}
\arguments{
\item{problem}{A \linkS4class{Problem} emitted by Dqcp2Dcp.}

\item{solver}{The solver to use for bisection.}

\item{low}{(Optional) Lower bound for bisection.}

\item{high}{(Optional) Upper bound for bisection.}

\item{eps}{Terminate bisection when width of interval is strictly less than eps.}

\item{verbose}{A logical value indicating whether to print verbose output related to bisection.}

\item{max_iters}{The maximum number of iterations to run the bisection.}
}
\value{
A \linkS4class{Solution} object.
}
\description{
Bisects on a one-parameter family of DCP problems emitted by Dqcp2Dcp.
}
