% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonInterface.R
\name{build_lin_op_tree}
\alias{build_lin_op_tree}
\title{Construct C++ LinOp tree from R LinOp tree.}
\usage{
build_lin_op_tree(root_linR, linR_to_linC)
}
\arguments{
\item{linR_to_linC}{A named for memoizing construction and storing
the C++ LinOps.}
}
\description{
Constructed C++ linOps are stored in the \code{linR_to_linC} dict,
which maps Python linOps to their corresponding C++ linOps.
}
