% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{canonical_form}
\alias{canonical_form}
\title{Get the Canonical Form}
\usage{
canonical_form(x, ...)
}
\arguments{
\item{x}{A canonicalizable object.}

\item{...}{Not used.}
}
\value{
List with (expression, constraints).
}
\description{
Get the Canonical Form
}
\keyword{internal}
