% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\docType{class}
\name{Canonicalization-class}
\alias{Canonicalization-class}
\alias{.Canonicalization}
\alias{perform,Canonicalization,Problem-method}
\alias{invert,Canonicalization,Solution,InverseData-method}
\alias{canonicalize_tree,Canonicalization,ANY,ANY-method}
\alias{canonicalize_expr,Canonicalization-method}
\title{The Canonicalization class.}
\usage{
\S4method{perform}{Canonicalization,Problem}(object, problem)

\S4method{invert}{Canonicalization,Solution,InverseData}(object, solution, inverse_data)

\S4method{canonicalize_tree}{Canonicalization,ANY,ANY}(object, expr)

\S4method{canonicalize_expr}{Canonicalization}(object, expr, args)
}
\arguments{
\item{object}{A \linkS4class{Canonicalization} object.}

\item{problem}{A \linkS4class{Problem} object.}

\item{solution}{A \linkS4class{Solution} to a problem that generated the inverse data.}

\item{inverse_data}{An \linkS4class{InverseData} object that contains the data encoding the original problem.}

\item{expr}{An \linkS4class{Expression} object.}

\item{args}{List of arguments to canonicalize the expression.}
}
\description{
This class represents a canonicalization reduction.
}
\details{
This reduction recursively canonicalizes every expression tree in a problem,
visiting each node. At every node, this reduction first canonicalizes its
arguments; it then canonicalizes the node, using the canonicalized arguments.

The attribute canon_methods is a list mapping node types to functions that
canonicalize them; the signature of these canonicalizing functions must be
\code{canon_func(expr, canon_args) --> (new_expr, constraints) }
where expr is the Expression (node) to canonicalize, canon_args is a list of
the canonicalized arguments of this expression, new_expr is a canonicalized
expression, and constraints is a list of constraints introduced while
canonicalizing expr.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perform(object = Canonicalization, problem = Problem)}: Recursively canonicalize the objective and every constraint.

\item \code{invert(
  object = Canonicalization,
  solution = Solution,
  inverse_data = InverseData
)}: Performs the reduction on a problem and returns an equivalent problem.

\item \code{canonicalize_tree(object = Canonicalization, expr = ANY, affine_above = ANY)}: Recursively canonicalize an Expression.

\item \code{canonicalize_expr(Canonicalization)}: Canonicalize an expression, w.r.t. canonicalized arguments.

}}
