% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{Ceil-class}
\alias{Ceil-class}
\alias{.Ceil}
\alias{Ceil}
\alias{to_numeric,Ceil-method}
\alias{sign_from_args,Ceil-method}
\alias{is_atom_convex,Ceil-method}
\alias{is_atom_concave,Ceil-method}
\alias{is_atom_log_log_convex,Ceil-method}
\alias{is_atom_log_log_concave,Ceil-method}
\alias{is_atom_quasiconvex,Ceil-method}
\alias{is_atom_quasiconcave,Ceil-method}
\alias{is_incr,Ceil-method}
\alias{is_decr,Ceil-method}
\alias{.grad,Ceil-method}
\title{The Ceil class.}
\usage{
Ceil(x)

\S4method{to_numeric}{Ceil}(object, values)

\S4method{sign_from_args}{Ceil}(object)

\S4method{is_atom_convex}{Ceil}(object)

\S4method{is_atom_concave}{Ceil}(object)

\S4method{is_atom_log_log_convex}{Ceil}(object)

\S4method{is_atom_log_log_concave}{Ceil}(object)

\S4method{is_atom_quasiconvex}{Ceil}(object)

\S4method{is_atom_quasiconcave}{Ceil}(object)

\S4method{is_incr}{Ceil}(object, idx)

\S4method{is_decr}{Ceil}(object, idx)

\S4method{.grad}{Ceil}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{object}{A \linkS4class{Ceil} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents elementwise ceiling.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(Ceil)}: The numeric value of the expression.

\item \code{sign_from_args(Ceil)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(Ceil)}: Is the atom convex?

\item \code{is_atom_concave(Ceil)}: Is the atom concave?

\item \code{is_atom_log_log_convex(Ceil)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(Ceil)}: Is the atom log-log concave?

\item \code{is_atom_quasiconvex(Ceil)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(Ceil)}: Is the atom quasiconcave?

\item \code{is_incr(Ceil)}: Is the atom weakly increasing in the index?

\item \code{is_decr(Ceil)}: Is the atom weakly decreasing in the index?

\item \code{.grad(Ceil)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}
}}

