% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarabel.R
\name{CLARABEL.dims_to_solver_dict}
\alias{CLARABEL.dims_to_solver_dict}
\title{Utility method for formatting a ConeDims instance into a dictionary
that can be supplied to Clarabel}
\usage{
CLARABEL.dims_to_solver_dict(cone_dims)
}
\arguments{
\item{cone_dims}{A \linkS4class{ConeDims} instance.}
}
\value{
The dimensions of the cones.
}
\description{
Utility method for formatting a ConeDims instance into a dictionary
that can be supplied to Clarabel
}
