% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/284_clarabel_conif.R
\name{CLARABEL.extract_dual_value}
\alias{CLARABEL.extract_dual_value}
\title{Extracts the dual value for constraint starting at offset.}
\usage{
CLARABEL.extract_dual_value(result_vec, offset, constraint)
}
\arguments{
\item{result_vec}{The vector to extract dual values from.}

\item{offset}{The starting point of the vector to extract from.}

\item{constraint}{A \linkS4class{Constraint} object.}
}
\value{
The dual values for the corresponding PSD constraints
}
\description{
Special cases PSD constraints, as per the CLARABEL specification.
}
