% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/161_coeff_extractor.R
\docType{class}
\name{CoeffExtractor-class}
\alias{CoeffExtractor-class}
\alias{.CoeffExtractor}
\alias{affine,CoeffExtractor,list-method}
\title{A coefficient matrix extractor from InverseData}
\usage{
\S4method{affine}{CoeffExtractor,list}(object, expr)
}
\description{
A coefficient matrix extractor from InverseData
}
\section{Methods (by generic)}{
\itemize{
\item \code{affine(object = CoeffExtractor, expr = list)}: Extract problem data tensor from an expression that is reducible to \eqn{A*x + b}. Applying the tensor to a flattened parameter vector and reshaping will recover \eqn{A} and \eqn{b} (see the helpers in canonInterface).

}}
