% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/305_exports.R
\name{complex-atoms}
\alias{complex-atoms}
\alias{Re,Expression-method}
\alias{Im,Expression-method}
\alias{Conj,Expression-method}
\title{Complex Numbers}
\usage{
\S4method{Re}{Expression}(z)

\S4method{Im}{Expression}(z)

\S4method{Conj}{Expression}(z)
}
\arguments{
\item{z}{An \linkS4class{Expression} object.}
}
\value{
An \linkS4class{Expression} object that represents the real, imaginary, or complex conjugate.
}
\description{
Basic atoms that support complex arithmetic.
}
