% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\name{complex-methods}
\alias{complex-methods}
\alias{is_real}
\alias{is_imag}
\alias{is_complex}
\title{Complex Properties}
\usage{
is_real(object)

is_imag(object)

is_complex(object)
}
\arguments{
\item{object}{an atom}
}
\value{
A logical value.

a logical value indicating whether the atom is real-valued
}
\description{
Determine if an expression is real, imaginary, or complex.
}
