% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex2real.R
\name{Complex2Real.add}
\alias{Complex2Real.add}
\title{Helper function to sum arguments.}
\usage{
Complex2Real.add(lh_arg, rh_arg, neg = FALSE)
}
\arguments{
\item{lh_arg}{The arguments for the left-hand side}

\item{rh_arg}{The arguments for the right-hand side}

\item{neg}{Whether to negate the right hand side}
}
\description{
Helper function to sum arguments.
}
