% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex2real.R
\name{Complex2Real.at_least_2D}
\alias{Complex2Real.at_least_2D}
\title{Upcast 0D and 1D to 2D.}
\usage{
Complex2Real.at_least_2D(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}
}
\value{
An expression of dimension at least 2.
}
\description{
Upcast 0D and 1D to 2D.
}
