% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex2real.R
\name{Complex2Real.binary_canon}
\alias{Complex2Real.binary_canon}
\title{Complex canonicalizer for the binary atom}
\usage{
Complex2Real.binary_canon(expr, real_args, imag_args, real2imag)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{real_args}{A list of \linkS4class{Constraint} objects for the real part of the expression}

\item{imag_args}{A list of \linkS4class{Constraint} objects for the imaginary part of the expression}
}
\value{
A canonicalization of a binary atom, where the returned
variables are the real component and the imaginary component.
}
\description{
Complex canonicalizer for the binary atom
}
