% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/166_aff_canon.R
\name{Complex2Real.join}
\alias{Complex2Real.join}
\title{Helper function to combine arguments.}
\usage{
Complex2Real.join(expr, lh_arg, rh_arg)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{lh_arg}{The arguments for the left-hand side}

\item{rh_arg}{The arguments for the right-hand side}
}
\value{
A joined expression of both left and right expressions
}
\description{
Helper function to combine arguments.
}
