% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/169_inequality_canon.R
\name{Complex2Real.nonpos_canon}
\alias{Complex2Real.nonpos_canon}
\title{Complex canonicalizer for the nonpositive constraint}
\usage{
Complex2Real.nonpos_canon(expr, real_args, imag_args, real2imag)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{real_args}{A list of \linkS4class{Constraint} objects for the real part of the expression}

\item{imag_args}{A list of \linkS4class{Constraint} objects for the imaginary part of the expression}

\item{real2imag}{A list mapping the ID of the real part of a complex expression to the ID of its imaginary part.}
}
\value{
A canonicalization of a nonpositive constraint, where the returned variables are the real component and the imaginary component.
}
\description{
Complex canonicalizer for the nonpositive constraint
}
