% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex2real.R
\name{Complex2Real.real_canon}
\alias{Complex2Real.real_canon}
\title{Complex canonicalizer for the real atom}
\usage{
Complex2Real.real_canon(expr, real_args, imag_args, real2imag)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{real_args}{A list of \linkS4class{Constraint} objects for the real part of the expression}

\item{imag_args}{A list of \linkS4class{Constraint} objects for the imaginary part of the expression}

\item{real2imag}{A list mapping the ID of the real part of a complex expression to the ID of its imaginary part.}
}
\value{
A canonicalization of a real atom, where the returned
variables are the real component and NULL for the imaginary component.
}
\description{
Complex canonicalizer for the real atom
}
