% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_atoms_condition_number.R
\name{condition_number}
\alias{condition_number}
\title{Condition number of a PSD matrix}
\usage{
condition_number(A)
}
\arguments{
\item{A}{A square matrix expression (must be PSD)}
}
\value{
An expression representing the condition number of A
}
\description{
Computes the condition number lambda_max(A) / lambda_min(A) for a
positive semidefinite matrix A. This is a quasiconvex atom.
}
