% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/021_condition_number.R
\docType{class}
\name{ConditionNumber-class}
\alias{ConditionNumber-class}
\alias{.ConditionNumber}
\alias{ConditionNumber}
\alias{to_numeric,ConditionNumber-method}
\alias{.domain,ConditionNumber-method}
\alias{.grad,ConditionNumber-method}
\alias{validate_args,ConditionNumber-method}
\alias{dim_from_args,ConditionNumber-method}
\alias{sign_from_args,ConditionNumber-method}
\alias{is_atom_convex,ConditionNumber-method}
\alias{is_atom_concave,ConditionNumber-method}
\alias{is_incr,ConditionNumber-method}
\alias{is_decr,ConditionNumber-method}
\title{The ConditionNumber class.}
\usage{
ConditionNumber(A)

\S4method{to_numeric}{ConditionNumber}(object, values)

\S4method{.domain}{ConditionNumber}(object)

\S4method{.grad}{ConditionNumber}(object, values)

\S4method{validate_args}{ConditionNumber}(object)

\S4method{dim_from_args}{ConditionNumber}(object)

\S4method{sign_from_args}{ConditionNumber}(object)

\S4method{is_atom_convex}{ConditionNumber}(object)

\S4method{is_atom_concave}{ConditionNumber}(object)

\S4method{is_incr}{ConditionNumber}(object, idx)

\S4method{is_decr}{ConditionNumber}(object, idx)
}
\arguments{
\item{A}{An \linkS4class{Expression} or numeric matrix.}

\item{object}{A \linkS4class{ConditionNumber} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents the condition number of a positive definite matrix \eqn{A}, i.e.,
}
\details{
\deqn{\lambda_{\max}(A) / \lambda_{\min}(A)}
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(ConditionNumber)}: The condition number of the matrix.

\item \code{.domain(ConditionNumber)}: Returns constraints describing the domain of the node.

\item \code{.grad(ConditionNumber)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{validate_args(ConditionNumber)}: Check that the matrix is square.

\item \code{dim_from_args(ConditionNumber)}: The dimensions of the atom determined from its arguments.

\item \code{sign_from_args(ConditionNumber)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(ConditionNumber)}: Is the atom convex?

\item \code{is_atom_concave(ConditionNumber)}: Is the atom concave?

\item \code{is_incr(ConditionNumber)}: Is the atom weakly increasing in the index?

\item \code{is_decr(ConditionNumber)}: Is the atom weakly decreasing in the index?

}}
\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a positive definite matrix.}
}}

