% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{cone-methods}
\alias{cone-methods}
\alias{num_cones}
\alias{cone_sizes}
\title{Second-Order Cone Methods}
\usage{
num_cones(object)

cone_sizes(object)
}
\arguments{
\item{object}{An \linkS4class{SOCAxis} object.}
}
\value{
The number of cones, or the size of a cone.
}
\description{
The number of elementwise cones or a list of the sizes of the elementwise cones.
}
