% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{cone_sizes}
\alias{cone_sizes}
\title{Get the Sizes of Individual Cones}
\usage{
cone_sizes(x, ...)
}
\arguments{
\item{x}{A cone constraint object.}

\item{...}{Not used.}
}
\value{
Integer vector of cone sizes.
}
\description{
Get the Sizes of Individual Cones
}
\keyword{internal}
