% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\docType{class}
\name{ConeDims-class}
\alias{ConeDims-class}
\alias{.ConeDims}
\title{Summary of cone dimensions present in constraints.}
\description{
Constraints must be formatted as dictionary that maps from
   constraint type to a list of constraints of that type.
}
\details{
Attributes
   ----------
   zero : int
       The dimension of the zero cone.
   nonpos : int
       The dimension of the non-positive cone.
   exp : int
       The dimension of the exponential cone.
   soc : list of int
       A list of the second-order cone dimensions.
   psd : list of int
       A list of the positive semidefinite cone dimensions, where the
       dimension of the PSD cone of k by k matrices is k.
}
