% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{ConicSolver.get_coeff_offset}
\alias{ConicSolver.get_coeff_offset}
\title{Return the coefficient and offset in A %*% x + b.}
\usage{
ConicSolver.get_coeff_offset(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object.}
}
\value{
The coefficient and offset in A %*% x + b.
}
\description{
Return the coefficient and offset in A %*% x + b.
}
