% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{ConicSolver.get_spacing_matrix}
\alias{ConicSolver.get_spacing_matrix}
\title{Returns a sparse matrix that spaces out an expression.}
\usage{
ConicSolver.get_spacing_matrix(dim, spacing, offset)
}
\arguments{
\item{dim}{A vector outlining the dimensions of the matrix.}

\item{spacing}{An int of the number of rows between the start of each non-zero block.}

\item{offset}{An int of the number of zeros at the beginning of the matrix.}
}
\value{
A sparse matrix that spaces out an expression
}
\description{
Returns a sparse matrix that spaces out an expression.
}
