% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{constants}
\alias{constants}
\title{Get the Constants in an Expression}
\usage{
constants(x, ...)
}
\arguments{
\item{x}{An expression or problem object.}

\item{...}{Not used.}
}
\value{
List of \code{\link{Constant}} objects.
}
\description{
Get the Constants in an Expression
}
