% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{constr_size}
\alias{constr_size}
\title{Get the Total Size of a Constraint}
\usage{
constr_size(x, ...)
}
\arguments{
\item{x}{A constraint object.}

\item{...}{Not used.}
}
\value{
Integer.
}
\description{
Get the Total Size of a Constraint
}
\keyword{internal}
