% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{constr_value}
\alias{constr_value}
\title{Is Constraint Violated?}
\usage{
constr_value(object, tolerance = 1e-08)
}
\arguments{
\item{object}{A \linkS4class{Constraint} object.}

\item{tolerance}{A numeric scalar representing the absolute tolerance to impose on the violation.}
}
\value{
A logical value indicating whether the violation is less than the \code{tolerance}. Raises an error if the residual is \code{NA}.
}
\description{
Checks whether the constraint violation is less than a tolerance.
}
