% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\docType{methods}
\name{construct_intermediate_chain,Problem,list-method}
\alias{construct_intermediate_chain,Problem,list-method}
\title{Builds a chain that rewrites a problem into an intermediate representation suitable for numeric reductions.}
\usage{
\S4method{construct_intermediate_chain}{Problem,list}(problem, candidates,
  gp = FALSE)
}
\arguments{
\item{problem}{The problem for which to build a chain}

\item{candidates}{A list of candidate solvers divided in qp_solvers}
}
\value{
A \linkS4class{Chain} object that can be used to convert the problem to an intermediate form.
}
\description{
Builds a chain that rewrites a problem into an intermediate representation suitable for numeric reductions.
}
