% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{Counter}
\alias{Counter}
\title{Counter class for ids}
\usage{
Counter()
}
\arguments{
\item{type}{the type of LinOp, one of the types above}

\item{dim}{the dim of the LinOp, a tuple, so for us a vector of integers}

\item{args}{the arguments of the LinOp}

\item{data}{the data for the LinOp, which is later set to C++ LinOp objects' linOp_data_ field}
}
\value{
an object of class "LinOp"
}
\description{
Counter class for ids
}
