% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{create_eq}
\alias{create_eq}
\title{Create an internal equality constraint}
\usage{
create_eq(lh_op, rh_op = NULL, constr_id = get_id())
}
\arguments{
\item{lh_op}{the lhs operator}

\item{rh_op}{the rhs operator}

\item{constr_id}{the id of the constraint}
}
\value{
the operator in the constraint
}
\description{
Create an internal equality constraint
}
