% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_atoms_cummax.R
\name{cummax_expr}
\alias{cummax_expr}
\title{Cumulative maximum along an axis}
\usage{
cummax_expr(x, axis = 2L)
}
\arguments{
\item{x}{An Expression}

\item{axis}{1 (across rows) or 2 (down columns, default)}
}
\value{
A Cummax atom
}
\description{
Cumulative maximum along an axis
}
