% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\docType{class}
\name{CvxAttr2Constr-class}
\alias{CvxAttr2Constr-class}
\alias{CvxAttr2Constr}
\alias{perform,CvxAttr2Constr,Problem-method}
\alias{invert,CvxAttr2Constr,Solution,list-method}
\title{The CvxAttr2Constr class.}
\usage{
\S4method{perform}{CvxAttr2Constr,Problem}(object, problem)

\S4method{invert}{CvxAttr2Constr,Solution,list}(object, solution,
  inverse_data)
}
\arguments{
\item{object}{A \linkS4class{CvxAttr2Constr} object.}

\item{problem}{A \linkS4class{Problem} object.}

\item{solution}{A \linkS4class{Solution} to a problem that generated the inverse data.}

\item{inverse_data}{The inverse data returned by an invocation to apply.}

\item{object}{A \linkS4class{CvxAttr2Constr} object.}
}
\description{
This class represents a reduction that expands convex variable attributes into constraints.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perform}: Expand convex variable attributes to constraints.

\item \code{invert}: Performs the reduction on a problem and returns an equivalent problem.
}}

