% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_atoms_norm.R
\name{cvxr_norm}
\alias{cvxr_norm}
\title{Compute a norm of an expression}
\usage{
cvxr_norm(x, p = 2, axis = NULL, keepdims = FALSE, max_denom = 1024L)
}
\arguments{
\item{x}{An Expression}

\item{p}{Norm type: 1, 2, Inf, or "fro" (Frobenius)}

\item{axis}{NULL (all), 0 (columns), or 1 (rows)}

\item{keepdims}{Logical}

\item{max_denom}{Integer max denominator}
}
\value{
A norm atom
}
\description{
Compute a norm of an expression
}
