% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.entr_canon}
\alias{Dcp2Cone.entr_canon}
\title{Dcp2Cone canonicalizer for the entropy atom}
\usage{
Dcp2Cone.entr_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from an entropy atom where 
the objective function is just the variable t with an ExpCone constraint.
}
\description{
Dcp2Cone canonicalizer for the entropy atom
}
