% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.exp_canon}
\alias{Dcp2Cone.exp_canon}
\title{Dcp2Cone canonicalizer for the exponential atom}
\usage{
Dcp2Cone.exp_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from an exponential atom 
where the objective function is the variable t with an ExpCone constraint.
}
\description{
Dcp2Cone canonicalizer for the exponential atom
}
