% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/202_geo_mean_canon.R
\name{Dcp2Cone.geo_mean_canon}
\alias{Dcp2Cone.geo_mean_canon}
\title{Dcp2Cone canonicalizer for the geometric mean atom}
\usage{
Dcp2Cone.geo_mean_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a geometric mean atom
where the objective function is the variable t with geometric mean constraints
}
\description{
Dcp2Cone canonicalizer for the geometric mean atom
}
