% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.huber_canon}
\alias{Dcp2Cone.huber_canon}
\title{Dcp2Cone canonicalizer for the huber atom}
\usage{
Dcp2Cone.huber_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a huber atom where the objective 
function is the variable t with square and absolute constraints
}
\description{
Dcp2Cone canonicalizer for the huber atom
}
