% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpcanons.R
\name{Dcp2Cone.indicator_canon}
\alias{Dcp2Cone.indicator_canon}
\title{Dcp2Cone canonicalizer for the indicator atom}
\usage{
Dcp2Cone.indicator_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from an indicator atom and
where 0 is the objective function with the given constraints
in the function.
}
\description{
Dcp2Cone canonicalizer for the indicator atom
}
