% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/205_kl_div_canon.R
\name{Dcp2Cone.kl_div_canon}
\alias{Dcp2Cone.kl_div_canon}
\title{Dcp2Cone canonicalizer for the KL Divergence atom}
\usage{
Dcp2Cone.kl_div_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a KL divergence atom
where t is the objective function with the ExpCone constraints.
}
\description{
Dcp2Cone canonicalizer for the KL Divergence atom
}
