% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.lambda_max_canon}
\alias{Dcp2Cone.lambda_max_canon}
\title{Dcp2Cone canonicalizer for the lambda maximization atom}
\usage{
Dcp2Cone.lambda_max_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a lambda maximization atom
where t is the objective function and a PSD constraint and a
constraint requiring I*t to be symmetric.
}
\description{
Dcp2Cone canonicalizer for the lambda maximization atom
}
