% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.lambda_sum_largest_canon}
\alias{Dcp2Cone.lambda_sum_largest_canon}
\title{Dcp2Cone canonicalizer for the largest lambda sum atom}
\usage{
Dcp2Cone.lambda_sum_largest_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a lambda sum of the k
largest elements atom where k*t + trace(Z) is the objective function.
t denotes the variable subject to constraints and Z is a PSD matrix variable
whose dimensions consist of the length of the vector at hand. The constraints
require the the diagonal matrix of the vector to be symmetric and PSD.
}
\description{
Dcp2Cone canonicalizer for the largest lambda sum atom
}
