% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.log1p_canon}
\alias{Dcp2Cone.log1p_canon}
\title{Dcp2Cone canonicalizer for the log 1p atom}
\usage{
Dcp2Cone.log1p_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a log 1p atom where
t is the objective function and the constraints consist of
ExpCone constraints + 1.
}
\description{
Dcp2Cone canonicalizer for the log 1p atom
}
