% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.log_det_canon}
\alias{Dcp2Cone.log_det_canon}
\title{Dcp2Cone canonicalizer for the log determinant atom}
\usage{
Dcp2Cone.log_det_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a log determinant atom where
the objective function is the sum of the log of the vector D
and the constraints consist of requiring the matrix Z to be
diagonal and the diagonal Z to equal D, Z to be upper triangular
and DZ; t(Z)A to be positive semidefinite, where A is a n by n
matrix.
}
\description{
Dcp2Cone canonicalizer for the log determinant atom
}
