% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.log_sum_exp_canon}
\alias{Dcp2Cone.log_sum_exp_canon}
\title{Dcp2Cone canonicalizer for the log sum of the exp atom}
\usage{
Dcp2Cone.log_sum_exp_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from the log sum
of the exp atom where the objective is the t variable
and the constraints consist of the ExpCone constraints and
requiring t to be less than a matrix of ones of the same size.
}
\description{
Dcp2Cone canonicalizer for the log sum of the exp atom
}
