% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpcanons.R
\name{Dcp2Cone.logistic_canon}
\alias{Dcp2Cone.logistic_canon}
\title{Dcp2Cone canonicalizer for the logistic function atom}
\usage{
Dcp2Cone.logistic_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from the logistic atom
where the objective function is given by t0 and the
constraints consist of the ExpCone constraints.
}
\description{
Dcp2Cone canonicalizer for the logistic function atom
}
