% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpcanons.R
\name{Dcp2Cone.matrix_frac_canon}
\alias{Dcp2Cone.matrix_frac_canon}
\title{Dcp2Cone canonicalizer for the matrix fraction atom}
\usage{
Dcp2Cone.matrix_frac_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from the matrix fraction
atom, where the objective function is the trace of Tvar, a
m by m matrix where the constraints consist of the matrix of
the Schur complement of Tvar to consist of P, an n by n, given
matrix, X, an n by m given matrix, and Tvar.
}
\description{
Dcp2Cone canonicalizer for the matrix fraction atom
}
