% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/214_mul_canon.R
\name{Dcp2Cone.mul_canon}
\alias{Dcp2Cone.mul_canon}
\title{Dcp2Cone canonicalizer for the multiplication of expressions atom}
\usage{
Dcp2Cone.mul_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from the multiplication of expressions atom.
}
\description{
Dcp2Cone canonicalizer for the multiplication of expressions atom
}
