% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpcanons.R
\name{Dcp2Cone.normNuc_canon}
\alias{Dcp2Cone.normNuc_canon}
\title{Dcp2Cone canonicalizer for the nuclear norm atom}
\usage{
Dcp2Cone.normNuc_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a nuclear norm atom,
where the objective function consists of .5 times the trace of
a matrix X of size m+n by m+n where the constraint consist of
the top right corner of the matrix being the original matrix.
}
\description{
Dcp2Cone canonicalizer for the nuclear norm atom
}
