% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.pnorm_canon}
\alias{Dcp2Cone.pnorm_canon}
\title{Dcp2Cone canonicalizer for the p norm atom}
\usage{
Dcp2Cone.pnorm_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a pnorm atom, where
the objective is a variable t of dimension of the original
vector in the problem and the constraints consist of geometric
mean constraints.
}
\description{
Dcp2Cone canonicalizer for the p norm atom
}
