% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.power_canon}
\alias{Dcp2Cone.power_canon}
\title{Dcp2Cone canonicalizer for the power atom}
\usage{
Dcp2Cone.power_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a power atom, where
the objective function consists of the variable t which is
of the dimension of the original vector from the power atom
and the constraints consists of geometric mean constraints.
}
\description{
Dcp2Cone canonicalizer for the power atom
}
