% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/219_quad_form_canon.R
\name{Dcp2Cone.quad_form_canon}
\alias{Dcp2Cone.quad_form_canon}
\title{Dcp2Cone canonicalizer for the quadratic form atom}
\usage{
Dcp2Cone.quad_form_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a quadratic form atom,
where the objective function consists of the scaled objective function
from the quadratic over linear canonicalization and same with the
constraints.
}
\description{
Dcp2Cone canonicalizer for the quadratic form atom
}
