% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpcanons.R
\name{Dcp2Cone.quad_over_lin_canon}
\alias{Dcp2Cone.quad_over_lin_canon}
\title{Dcp2Cone canonicalizer for the quadratic over linear term atom}
\usage{
Dcp2Cone.quad_over_lin_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a quadratic over linear
term atom where the objective function consists of a one
dimensional variable t with SOC constraints.
}
\description{
Dcp2Cone canonicalizer for the quadratic over linear term atom
}
