% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcp2cone.R
\name{Dcp2Cone.sigma_max_canon}
\alias{Dcp2Cone.sigma_max_canon}
\title{Dcp2Cone canonicalizer for the sigma max atom}
\usage{
Dcp2Cone.sigma_max_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a sigma max atom
where the objective function consists of the variable t
that is of the same dimension as the original expression
with specified constraints in the function.
}
\description{
Dcp2Cone canonicalizer for the sigma max atom
}
