% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/225_von_neumann_entr_canon.R
\name{Dcp2Cone.von_neumann_entr_canon}
\alias{Dcp2Cone.von_neumann_entr_canon}
\title{Dcp2Cone canonicalizer for the von Neumann entropy atom}
\usage{
Dcp2Cone.von_neumann_entr_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A cone program constructed from a von Neumann entropy atom
where the objective function consists of the variable t
that is of the same dimension as the original expression
with specified constraints in the function.
}
\description{
Dcp2Cone canonicalizer for the von Neumann entropy atom
}
