% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp2dcp.R
\name{Dgp2Dcp.add_canon}
\alias{Dgp2Dcp.add_canon}
\title{Dgp2Dcp canonicalizer for the addition atom}
\usage{
Dgp2Dcp.add_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the addition atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the addition atom
}
