% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/232_constant_canon.R
\name{Dgp2Dcp.constant_canon}
\alias{Dgp2Dcp.constant_canon}
\title{Dgp2Dcp canonicalizer for the constant atom}
\usage{
Dgp2Dcp.constant_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the constant atom of a DGP expression,
where the returned expression is the DCP equivalent resulting
from the log of the expression.
}
\description{
Dgp2Dcp canonicalizer for the constant atom
}
