% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/235_eye_minus_inv_canon.R
\name{Dgp2Dcp.eye_minus_inv_canon}
\alias{Dgp2Dcp.eye_minus_inv_canon}
\title{Dgp2Dcp canonicalizer for the (I - X)^{-1} atom}
\usage{
Dgp2Dcp.eye_minus_inv_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the (I - X)^{-1} atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the (I - X)^{-1} atom
}
