% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/238_log_canon.R
\name{Dgp2Dcp.log_canon}
\alias{Dgp2Dcp.log_canon}
\title{Dgp2Dcp canonicalizer for the log atom}
\usage{
Dgp2Dcp.log_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the log atom of a DGP expression,
where the returned expression is the log of the original expression..
}
\description{
Dgp2Dcp canonicalizer for the log atom
}
