% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/239_mul_canon.R
\name{Dgp2Dcp.mul_canon}
\alias{Dgp2Dcp.mul_canon}
\title{Dgp2Dcp canonicalizer for the multiplication atom}
\usage{
Dgp2Dcp.mul_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the multiplication atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the multiplication atom
}
