% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp2dcp.R
\name{Dgp2Dcp.one_minus_pos_canon}
\alias{Dgp2Dcp.one_minus_pos_canon}
\title{Dgp2Dcp canonicalizer for the 1-x atom}
\usage{
Dgp2Dcp.one_minus_pos_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the 1-x with 0 < x < 1 atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the 1-x atom
}
