% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/249_prod_canon.R
\name{Dgp2Dcp.prod_canon}
\alias{Dgp2Dcp.prod_canon}
\title{Dgp2Dcp canonicalizer for the product atom}
\usage{
Dgp2Dcp.prod_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the product atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the product atom
}
