% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/255_zero_constr_canon.R
\name{Dgp2Dcp.zero_constr_canon}
\alias{Dgp2Dcp.zero_constr_canon}
\title{Dgp2Dcp canonicalizer for the zero constraint atom}
\usage{
Dgp2Dcp.zero_constr_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of values for the expr variable}
}
\value{
A canonicalization of the zero constraint atom of a DGP expression,
where the returned expression is the transformed DCP equivalent.
}
\description{
Dgp2Dcp canonicalizer for the zero constraint atom
}
