% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp2dcp.R
\docType{class}
\name{DgpCanonMethods-class}
\alias{DgpCanonMethods-class}
\alias{.DgpCanonMethods}
\alias{names,DgpCanonMethods-method}
\alias{$,DgpCanonMethods-method}
\title{DGP canonical methods class.}
\usage{
\S4method{names}{DgpCanonMethods}(x)

\S4method{$}{DgpCanonMethods}(x, name)
}
\arguments{
\item{x}{A \linkS4class{DgpCanonMethods} object.}

\item{name}{The name of the atom or expression to canonicalize.}
}
\description{
Canonicalization of DGPs is a stateful procedure, hence the need for a class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{names}: Returns the name of all the canonicalization methods

\item \code{$}: Returns either a canonicalized variable or 
a corresponding Dgp2Dcp canonicalization method
}}

