% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/032_diag.R
\name{Diag}
\alias{Diag}
\title{The Diag atom.}
\usage{
Diag(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} that represents a vector or square matrix.}
}
\value{
An \linkS4class{Expression} representing the diagonal vector/matrix.
}
\description{
Turns an expression into a DiagVec object
}
