% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/033_diff.R
\name{Diff}
\alias{Diff}
\title{The Diff atom.}
\usage{
Diff(x, lag = 1, k = 1, axis = 2)
}
\arguments{
\item{x}{An \linkS4class{Expression} that represents a vector}

\item{lag}{The degree of lag between differences}

\item{k}{The integer value of the order of differences}

\item{axis}{The axis along which to apply the function. For a 2D matrix, \code{1} indicates rows and \code{2} indicates columns.}
}
\value{
Takes in a vector of length n and returns a vector of length n-k of the kth order differences
}
\description{
Takes the k-th order differences
}
