% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{DiffPos}
\alias{DiffPos}
\title{The DiffPos atom.}
\usage{
DiffPos(x, y)
}
\arguments{
\item{x}{An \linkS4class{Expression}}

\item{y}{An \linkS4class{Expression}}
}
\value{
The difference \eqn{x - y} with domain \eqn{x,y: x > y > 0}.
}
\description{
The difference between expressions, \eqn{x - y}, where \eqn{x > y > 0}.
}
