% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R, R/019_atom.R
\name{dim_from_args}
\alias{dim_from_args}
\alias{dim_from_args,Atom-method}
\title{Atom Dimensions}
\usage{
dim_from_args(object)

\S4method{dim_from_args}{Atom}(object)
}
\arguments{
\item{object}{An \linkS4class{Atom} object.}
}
\value{
A numeric vector \code{c(row, col)} indicating the dimensions of the atom.

A numeric vector \code{c(row, col)} indicating the dimensions of the atom.
}
\description{
Determine the dimensions of an atom based on its arguments.

Determine the dimensions of an atom based on its arguments.
}
