% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/023_dist_ratio.R
\docType{class}
\name{DistRatio-class}
\alias{DistRatio-class}
\alias{.DistRatio}
\alias{DistRatio}
\alias{to_numeric,DistRatio-method}
\alias{dim_from_args,DistRatio-method}
\alias{sign_from_args,DistRatio-method}
\alias{is_atom_convex,DistRatio-method}
\alias{is_atom_concave,DistRatio-method}
\alias{is_atom_quasiconvex,DistRatio-method}
\alias{is_atom_quasiconcave,DistRatio-method}
\alias{is_incr,DistRatio-method}
\alias{is_decr,DistRatio-method}
\alias{.grad,DistRatio-method}
\title{The DistRatio class.}
\usage{
DistRatio(x, a, b)

\S4method{to_numeric}{DistRatio}(object, values)

\S4method{dim_from_args}{DistRatio}(object)

\S4method{sign_from_args}{DistRatio}(object)

\S4method{is_atom_convex}{DistRatio}(object)

\S4method{is_atom_concave}{DistRatio}(object)

\S4method{is_atom_quasiconvex}{DistRatio}(object)

\S4method{is_atom_quasiconcave}{DistRatio}(object)

\S4method{is_incr}{DistRatio}(object, idx)

\S4method{is_decr}{DistRatio}(object, idx)

\S4method{.grad}{DistRatio}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{a}{A numeric value.}

\item{b}{A numeric value.}

\item{object}{A \linkS4class{DistRatio} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents the ratio of the \eqn{l_2}-norm distance from \eqn{x} to two points \eqn{a} and \eqn{b}, i.e.,
}
\details{
\deqn{||x - a||_2 / ||x - b||_2},

where \eqn{a} and \eqn{b} are constants with \eqn{norm(x - a)_2 \leq norm(x - b)}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(DistRatio)}: The distance ratio of the expression.

\item \code{dim_from_args(DistRatio)}: The dimensions of the atom determined from its arguments.

\item \code{sign_from_args(DistRatio)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(DistRatio)}: Is the atom convex?

\item \code{is_atom_concave(DistRatio)}: Is the atom concave?

\item \code{is_atom_quasiconvex(DistRatio)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(DistRatio)}: Is the atom quasiconcave?

\item \code{is_incr(DistRatio)}: Is the atom weakly increasing in the index?

\item \code{is_decr(DistRatio)}: Is the atom weakly decreasing in the index?

\item \code{.grad(DistRatio)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}
}}

