% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{domain}
\alias{domain}
\title{Get the Domain Constraints of an Expression}
\usage{
domain(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
List of constraints defining the domain.
}
\description{
Get the Domain Constraints of an Expression
}
\keyword{internal}
