% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.build_matrix_0}
\alias{.build_matrix_0}
\title{Get the \code{sparse} flag field for the LinOp object}
\usage{
.build_matrix_0(xp, v)
}
\arguments{
\item{xp}{the LinOpVector Object XPtr}

\item{v}{the \code{id_to_col} named int vector in R with integer names}
}
\value{
a XPtr to ProblemData Object
}
\description{
Get the \code{sparse} flag field for the LinOp object
}
