% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.build_matrix_1}
\alias{.build_matrix_1}
\title{Get the \code{sparse} flag field for the LinOp object}
\usage{
.build_matrix_1(xp, v1, v2)
}
\arguments{
\item{xp}{the LinOpVector Object XPtr}

\item{v1}{the \code{id_to_col} named int vector in R with integer names}

\item{v2}{the \code{constr_offsets} vector of offsets (an int vector in R)}
}
\value{
a XPtr to ProblemData Object
}
\description{
Get the \code{sparse} flag field for the LinOp object
}
