% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__args_push_back}
\alias{.LinOp__args_push_back}
\title{Perform a push back operation on the \code{args} field of LinOp}
\usage{
.LinOp__args_push_back(xp, yp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{yp}{the LinOp Object XPtr to push}
}
\description{
Perform a push back operation on the \code{args} field of LinOp
}
