% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_dense_data}
\alias{.LinOp__get_dense_data}
\title{Get the field \code{dense_data} for the LinOp object}
\usage{
.LinOp__get_dense_data(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
a MatrixXd object
}
\description{
Get the field \code{dense_data} for the LinOp object
}
