% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_id}
\alias{.LinOp__get_id}
\title{Get the id field of the LinOp Object}
\usage{
.LinOp__get_id(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
the value of the id field of the LinOp Object
}
\description{
Get the id field of the LinOp Object
}
