% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_size}
\alias{.LinOp__get_size}
\title{Get the field \code{size} for the LinOp object}
\usage{
.LinOp__get_size(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
an integer vector
}
\description{
Get the field \code{size} for the LinOp object
}
