% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_slice}
\alias{.LinOp__get_slice}
\title{Get the slice field of the LinOp Object}
\usage{
.LinOp__get_slice(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
the value of the slice field of the LinOp Object
}
\description{
Get the slice field of the LinOp Object
}
