% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_sparse}
\alias{.LinOp__get_sparse}
\title{Get the \code{sparse} flag field for the LinOp object}
\usage{
.LinOp__get_sparse(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
TRUE or FALSE
}
\description{
Get the \code{sparse} flag field for the LinOp object
}
