% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_sparse_data}
\alias{.LinOp__get_sparse_data}
\title{Get the field named \code{sparse_data} from the LinOp object}
\usage{
.LinOp__get_sparse_data(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
a \link[Matrix]{dgCMatrix-class} object
}
\description{
Get the field named \code{sparse_data} from the LinOp object
}
