% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__get_type}
\alias{.LinOp__get_type}
\title{Get the field named \code{type} for the LinOp object}
\usage{
.LinOp__get_type(xp)
}
\arguments{
\item{xp}{the LinOp Object XPtr}
}
\value{
an integer value for type
}
\description{
Get the field named \code{type} for the LinOp object
}
