% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__new}
\alias{.LinOp__new}
\title{Create a new LinOp object.}
\usage{
.LinOp__new()
}
\value{
an external ptr (Rcpp::XPtr) to a LinOp object instance.
}
\description{
Create a new LinOp object.
}
