% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_dense_data}
\alias{.LinOp__set_dense_data}
\title{Set the field \code{dense_data} of the LinOp object}
\usage{
.LinOp__set_dense_data(xp, denseMat)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{denseMat}{a standard matrix object in R}
}
\description{
Set the field \code{dense_data} of the LinOp object
}
