% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_size}
\alias{.LinOp__set_size}
\title{Set the field \code{size} of the LinOp object}
\usage{
.LinOp__set_size(xp, value)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{value}{an integer vector object in R}
}
\description{
Set the field \code{size} of the LinOp object
}
