% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_slice}
\alias{.LinOp__set_slice}
\title{Set the slice field of the LinOp Object}
\usage{
.LinOp__set_slice(xp, value)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{value}{a list of integer vectors, e.g. \code{list(1:10, 2L, 11:15)}}
}
\value{
the value of the slice field of the LinOp Object
}
\description{
Set the slice field of the LinOp Object
}
