% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_sparse}
\alias{.LinOp__set_sparse}
\title{Set the flag \code{sparse} of the LinOp object}
\usage{
.LinOp__set_sparse(xp, sparseSEXP)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{sparseSEXP}{an R boolean}
}
\description{
Set the flag \code{sparse} of the LinOp object
}
