% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_sparse_data}
\alias{.LinOp__set_sparse_data}
\title{Set the field named \code{sparse_data} of the LinOp object}
\usage{
.LinOp__set_sparse_data(xp, sparseMat)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{sparseMat}{a \link[Matrix]{dgCMatrix-class} object}
}
\description{
Set the field named \code{sparse_data} of the LinOp object
}
