% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__set_type}
\alias{.LinOp__set_type}
\title{Set the field named \code{type} for the LinOp object}
\usage{
.LinOp__set_type(xp, typeValue)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{typeValue}{an integer value}
}
\description{
Set the field named \code{type} for the LinOp object
}
