% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__size_push_back}
\alias{.LinOp__size_push_back}
\title{Perform a push back operation on the \code{size} field of LinOp}
\usage{
.LinOp__size_push_back(xp, intVal)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{intVal}{the integer value to push back}
}
\description{
Perform a push back operation on the \code{size} field of LinOp
}
