% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp__slice_push_back}
\alias{.LinOp__slice_push_back}
\title{Perform a push back operation on the \code{slice} field of LinOp}
\usage{
.LinOp__slice_push_back(xp, intVec)
}
\arguments{
\item{xp}{the LinOp Object XPtr}

\item{intVec}{an integer vector to push back}
}
\description{
Perform a push back operation on the \code{slice} field of LinOp
}
