% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOp_at_index}
\alias{.LinOp_at_index}
\title{Return the LinOp element at index i (0-based)}
\usage{
.LinOp_at_index(lvec, i)
}
\arguments{
\item{lvec}{the LinOpVector Object XPtr}

\item{i}{the index}
}
\description{
Return the LinOp element at index i (0-based)
}
