% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOpVector__new}
\alias{.LinOpVector__new}
\title{Create a new LinOpVector object.}
\usage{
.LinOpVector__new()
}
\value{
an external ptr (Rcpp::XPtr) to a LinOp object instance.
}
\description{
Create a new LinOpVector object.
}
