% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LinOpVector__push_back}
\alias{.LinOpVector__push_back}
\title{Perform a push back operation on the \code{args} field of LinOp}
\usage{
.LinOpVector__push_back(xp, yp)
}
\arguments{
\item{xp}{the LinOpVector Object XPtr}

\item{yp}{the LinOp Object XPtr to push}
}
\description{
Perform a push back operation on the \code{args} field of LinOp
}
