% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/099_pnorm.R
\name{.p_norm}
\alias{.p_norm}
\title{Internal method for calculating the p-norm}
\usage{
.p_norm(x, p)
}
\arguments{
\item{x}{A matrix}

\item{p}{A number grater than or equal to 1, or equal to positive infinity}
}
\value{
Returns the specified norm of matrix x
}
\description{
Internal method for calculating the p-norm
}
