% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_const_to_row}
\alias{.ProblemData__get_const_to_row}
\title{Get the const_to_row field of the ProblemData Object}
\usage{
.ProblemData__get_const_to_row(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
the const_to_row field as a named integer vector where the names are integers converted to characters
}
\description{
Get the const_to_row field of the ProblemData Object
}
