% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_const_vec}
\alias{.ProblemData__get_const_vec}
\title{Get the const_vec field from the ProblemData Object}
\usage{
.ProblemData__get_const_vec(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
a numeric vector of the field const_vec from the ProblemData Object
}
\description{
Get the const_vec field from the ProblemData Object
}
