% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_I}
\alias{.ProblemData__get_I}
\title{Get the I field of the ProblemData Object}
\usage{
.ProblemData__get_I(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
an integer vector of the field I from the ProblemData Object
}
\description{
Get the I field of the ProblemData Object
}
