% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_id_to_col}
\alias{.ProblemData__get_id_to_col}
\title{Get the id_to_col field of the ProblemData Object}
\usage{
.ProblemData__get_id_to_col(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
the id_to_col field as a named integer vector where the names are integers converted to characters
}
\description{
Get the id_to_col field of the ProblemData Object
}
