% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_J}
\alias{.ProblemData__get_J}
\title{Get the J field of the ProblemData Object}
\usage{
.ProblemData__get_J(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
an integer vector of the field J from the ProblemData Object
}
\description{
Get the J field of the ProblemData Object
}
