% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__get_V}
\alias{.ProblemData__get_V}
\title{Get the V field of the ProblemData Object}
\usage{
.ProblemData__get_V(xp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}
}
\value{
a numeric vector of doubles (the field V) from the ProblemData Object
}
\description{
Get the V field of the ProblemData Object
}
