% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__new}
\alias{.ProblemData__new}
\title{Create a new ProblemData object.}
\usage{
.ProblemData__new()
}
\value{
an external ptr (Rcpp::XPtr) to a ProblemData object instance.
}
\description{
Create a new ProblemData object.
}
