% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_const_to_row}
\alias{.ProblemData__set_const_to_row}
\title{Set the const_to_row map of the ProblemData Object}
\usage{
.ProblemData__set_const_to_row(xp, iv)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{iv}{a named integer vector with names being integers converted to characters}
}
\description{
Set the const_to_row map of the ProblemData Object
}
