% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_const_vec}
\alias{.ProblemData__set_const_vec}
\title{Set the const_vec field in the ProblemData Object}
\usage{
.ProblemData__set_const_vec(xp, cvp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{cvp}{a numeric vector of values for const_vec field of the ProblemData object}
}
\description{
Set the const_vec field in the ProblemData Object
}
