% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_I}
\alias{.ProblemData__set_I}
\title{Set the I field in the ProblemData Object}
\usage{
.ProblemData__set_I(xp, ip)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{ip}{an integer vector of values for field I of the ProblemData object}
}
\description{
Set the I field in the ProblemData Object
}
