% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_id_to_col}
\alias{.ProblemData__set_id_to_col}
\title{Set the id_to_col field of the ProblemData Object}
\usage{
.ProblemData__set_id_to_col(xp, iv)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{iv}{a named integer vector with names being integers converted to characters}
}
\description{
Set the id_to_col field of the ProblemData Object
}
