% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_J}
\alias{.ProblemData__set_J}
\title{Set the J field in the ProblemData Object}
\usage{
.ProblemData__set_J(xp, jp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{jp}{an integer vector of the values for field J of the ProblemData object}
}
\description{
Set the J field in the ProblemData Object
}
