% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ProblemData__set_V}
\alias{.ProblemData__set_V}
\title{Set the V field in the ProblemData Object}
\usage{
.ProblemData__set_V(xp, vp)
}
\arguments{
\item{xp}{the ProblemData Object XPtr}

\item{vp}{a numeric vector of values for field V}
}
\description{
Set the V field in the ProblemData Object
}
