% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/301_zzz_R_specific_data.R
\docType{data}
\name{dspop}
\alias{dspop}
\title{Direct Standardization: Population}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
\item{y}{Response variable}
\item{sex}{Sex of individual, coded male (0) and female (1)}
\item{age}{Age of individual}
}
}
\usage{
dspop
}
\description{
Randomly generated data for direct standardization example.
Sex was drawn from a Bernoulli distribution, and age was drawn from a uniform distribution on \eqn{10,\ldots,60}.
The response was drawn from a normal distribution with a mean that depends on sex and age, and a variance of 1.
}
\seealso{
\link[CVXR]{dssamp}
}
\keyword{data}
