% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{dual_cone}
\alias{dual_cone}
\title{Get the Dual Cone Constraint}
\usage{
dual_cone(x, ...)
}
\arguments{
\item{x}{A cone constraint object.}

\item{...}{Optional arguments.}
}
\value{
A cone constraint representing the dual cone.
}
\description{
Get the Dual Cone Constraint
}
\keyword{internal}
