% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{dual_residual}
\alias{dual_residual}
\title{Get the Dual Residual}
\usage{
dual_residual(x, ...)
}
\arguments{
\item{x}{A cone constraint object.}

\item{...}{Not used.}
}
\value{
Numeric residual.
}
\description{
Get the Dual Residual
}
\keyword{internal}
