% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\name{dual_value-methods}
\alias{dual_value-methods}
\alias{dual_value}
\alias{dual_value<-}
\title{Get and Set Dual Value}
\usage{
dual_value(object)

dual_value(object) <- value
}
\arguments{
\item{object}{A \linkS4class{Constraint} object.}

\item{value}{A numeric scalar, vector, or matrix to assign to the object.}
}
\description{
Get and set the value of the dual variable in a constraint.
}
