% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{Elementwise-class}
\alias{Elementwise-class}
\alias{Elementwise}
\alias{validate_args,Elementwise-method}
\alias{size_from_args,Elementwise-method}
\title{The Elementwise class.}
\usage{
\S4method{validate_args}{Elementwise}(object)

\S4method{size_from_args}{Elementwise}(object)
}
\arguments{
\item{object}{An \linkS4class{Elementwise} object.}
}
\description{
This virtual class represents an elementwise atom.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args}: Check all the shapes are the same or can be promoted.

\item \code{size_from_args}: Size is the same as the sum of the arguments' sizes.
}}

