% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_pwl.R
\docType{class}
\name{EliminatePwl-class}
\alias{EliminatePwl-class}
\alias{.EliminatePwl}
\alias{accepts,EliminatePwl,Problem-method}
\title{The EliminatePwl class.}
\usage{
\S4method{accepts}{EliminatePwl,Problem}(object, problem)
}
\arguments{
\item{object}{An \linkS4class{EliminatePwl} object.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
This class eliminates piecewise linear atoms.
}
\section{Methods (by generic)}{
\itemize{
\item \code{accepts(object = EliminatePwl, problem = Problem)}: Does this problem contain piecewise linear atoms?

}}
