% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_pwl.R
\name{EliminatePwl.abs_canon}
\alias{EliminatePwl.abs_canon}
\title{EliminatePwl canonicalizer for the absolute atom}
\usage{
EliminatePwl.abs_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the picewise-lienar atom
constructed from an absolute atom where the objective function 
consists of the variable that is of the same dimension as the 
original expression and the constraints consist of splitting 
the absolute value into two inequalities.
}
\description{
EliminatePwl canonicalizer for the absolute atom
}
