% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_pwl.R
\name{EliminatePwl.cummax_canon}
\alias{EliminatePwl.cummax_canon}
\title{EliminatePwl canonicalizer for the cumulative max atom}
\usage{
EliminatePwl.cummax_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-lienar atom
constructed from a cumulative max atom where the objective
function consists of the variable Y which is of the same
dimension as the original expression and the constraints
consist of row/column constraints depending on the axis
}
\description{
EliminatePwl canonicalizer for the cumulative max atom
}
