% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_pwl.R
\name{EliminatePwl.dotsort_canon}
\alias{EliminatePwl.dotsort_canon}
\title{EliminatePwl canonicalizer for the dot sort atom}
\usage{
EliminatePwl.dotsort_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-lienar atom
constructed from the dot sort atom where the objective
function consists of the variable t of the same size as
the original expression and the constraints consist of
a vector multiplied by a vector of 1's.
}
\description{
EliminatePwl canonicalizer for the dot sort atom
}
