% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/190_maximum_canon.R
\name{EliminatePwl.max_elemwise_canon}
\alias{EliminatePwl.max_elemwise_canon}
\title{EliminatePwl canonicalizer for the elementwise maximum atom}
\usage{
EliminatePwl.max_elemwise_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-linear atom
constructed by a elementwise maximum atom where the
objective function is the variable t of the same dimension
as the expression and the constraints consist of a simple
inequality.
}
\description{
EliminatePwl canonicalizer for the elementwise maximum atom
}
