% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/189_max_canon.R
\name{EliminatePwl.max_entries_canon}
\alias{EliminatePwl.max_entries_canon}
\title{EliminatePwl canonicalizer for the max entries atom}
\usage{
EliminatePwl.max_entries_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-linear atom
constructed from the max entries atom where the objective
function consists of the variable t of the same size as
the original expression and the constraints consist of
a vector multiplied by a vector of 1's.
}
\description{
EliminatePwl canonicalizer for the max entries atom
}
