% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/194_norm_inf_canon.R
\name{EliminatePwl.norm_inf_canon}
\alias{EliminatePwl.norm_inf_canon}
\title{EliminatePwl canonicalizer for the infinite norm atom}
\usage{
EliminatePwl.norm_inf_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-lienar atom
constructed by the infinite norm atom where the objective
function consists variable t of the same dimension as the
expression and the constraints consist of a vector
constructed by multiplying t to a vector of 1's
}
\description{
EliminatePwl canonicalizer for the infinite norm atom
}
