% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminate_pwl.R
\name{EliminatePwl.sum_largest_canon}
\alias{EliminatePwl.sum_largest_canon}
\title{EliminatePwl canonicalizer for the largest sum atom}
\usage{
EliminatePwl.sum_largest_canon(expr, args)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{args}{A list of \linkS4class{Constraint} objects}
}
\value{
A canonicalization of the piecewise-lienar atom
constructed by the k largest sums atom where the objective
function consists of the sum of variables t that is of
the same dimension as the expression plus k
}
\description{
EliminatePwl canonicalizer for the largest sum atom
}
