% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\docType{class}
\name{EvalParams-class}
\alias{EvalParams-class}
\alias{EvalParams}
\alias{perform,EvalParams,Problem-method}
\alias{invert,EvalParams,Solution,list-method}
\title{The EvalParams class.}
\usage{
\S4method{perform}{EvalParams,Problem}(object, problem)

\S4method{invert}{EvalParams,Solution,list}(object, solution, inverse_data)
}
\arguments{
\item{object}{A \linkS4class{EvalParams} object.}

\item{problem}{A \linkS4class{Problem} object.}

\item{solution}{A \linkS4class{Solution} to a problem that generated the inverse data.}

\item{inverse_data}{The inverse data returned by an invocation to apply.}

\item{object}{A \linkS4class{EvalParams} object.}
}
\description{
This class represents a reduction that replaces symbolic parameters with
their constaint values.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perform}: Replace parameters with constant values.

\item \code{invert}: Returns a solution to the original problem given the inverse_data.
}}

