% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/181_exotic2common.R
\docType{class}
\name{Exotic2Common-class}
\alias{Exotic2Common-class}
\alias{.Exotic2Common}
\title{The Exotic2Common class.}
\description{
This class represents a reduction of an exotic cone to a common cone constraint.
}
