% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{exp,Expression-method}
\alias{exp,Expression-method}
\alias{exp}
\title{Natural Exponential}
\usage{
\S4method{exp}{Expression}(x)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}
}
\value{
An \linkS4class{Expression} representing the natural exponential of the input.
}
\description{
The elementwise natural exponential.
}
\examples{
x <- Variable(5)
obj <- Minimize(sum(exp(x)))
prob <- Problem(obj, list(sum(x) == 1))
result <- solve(prob)
result$getValue(x)
}
