% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{expr_name}
\alias{expr_name}
\title{Get the Name of an Expression}
\usage{
expr_name(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Character string.
}
\description{
Get the Name of an Expression
}
\keyword{internal}
