% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{extract_dual_value}
\alias{extract_dual_value}
\title{Gets a specified value of a dual variable.}
\usage{
extract_dual_value(result_vec, offset, constraint)
}
\arguments{
\item{result_vec}{A vector containing the dual variable values.}

\item{offset}{An offset to get correct index of dual values.}

\item{constraint}{A list of the constraints in the problem.}
}
\value{
A list of a dual variable value and its offset.
}
\description{
Gets a specified value of a dual variable.
}
