% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{extract_mip_idx}
\alias{extract_mip_idx}
\title{Coalesces bool, int indices for variables.}
\usage{
extract_mip_idx(variables)
}
\arguments{
\item{variables}{A list of \linkS4class{Variable} objects.}
}
\value{
Coalesces bool, int indices for variables. The indexing scheme assumes that the variables will be coalesced into
a single one-dimensional variable, with each variable being reshaped in Fortran order.
}
\description{
Coalesces bool, int indices for variables.
}
