% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\docType{class}
\name{FlipObjective-class}
\alias{FlipObjective-class}
\alias{FlipObjective}
\alias{perform,FlipObjective,Problem-method}
\alias{invert,FlipObjective,Solution,list-method}
\title{The FlipObjective class.}
\usage{
\S4method{perform}{FlipObjective,Problem}(object, problem)

\S4method{invert}{FlipObjective,Solution,list}(object, solution,
  inverse_data)
}
\arguments{
\item{object}{A \linkS4class{FlipObjective} object.}

\item{problem}{A \linkS4class{Problem} object.}

\item{solution}{A \linkS4class{Solution} to a problem that generated the inverse data.}

\item{inverse_data}{The inverse data returned by an invocation to apply.}

\item{object}{A \linkS4class{FlipObjective} object.}
}
\description{
This class represents a reduction that flips a minimization objective to a
maximization and vice versa.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perform}: Flip a minimization objective to a maximization and vice versa.

\item \code{invert}: Map the solution of the flipped problem to that of the original.
}}

