% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{Floor-class}
\alias{Floor-class}
\alias{.Floor}
\alias{Floor}
\alias{to_numeric,Floor-method}
\alias{sign_from_args,Floor-method}
\alias{is_atom_convex,Floor-method}
\alias{is_atom_concave,Floor-method}
\alias{is_atom_log_log_convex,Floor-method}
\alias{is_atom_log_log_concave,Floor-method}
\alias{is_atom_quasiconvex,Floor-method}
\alias{is_atom_quasiconcave,Floor-method}
\alias{is_incr,Floor-method}
\alias{is_decr,Floor-method}
\alias{.grad,Floor-method}
\title{The Floor class.}
\usage{
Floor(x)

\S4method{to_numeric}{Floor}(object, values)

\S4method{sign_from_args}{Floor}(object)

\S4method{is_atom_convex}{Floor}(object)

\S4method{is_atom_concave}{Floor}(object)

\S4method{is_atom_log_log_convex}{Floor}(object)

\S4method{is_atom_log_log_concave}{Floor}(object)

\S4method{is_atom_quasiconvex}{Floor}(object)

\S4method{is_atom_quasiconcave}{Floor}(object)

\S4method{is_incr}{Floor}(object, idx)

\S4method{is_decr}{Floor}(object, idx)

\S4method{.grad}{Floor}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{object}{A \linkS4class{Floor} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents elementwise ceiling.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(Floor)}: The numeric value of the expression.

\item \code{sign_from_args(Floor)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(Floor)}: Is the atom convex?

\item \code{is_atom_concave(Floor)}: Is the atom concave?

\item \code{is_atom_log_log_convex(Floor)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(Floor)}: Is the atom log-log concave?

\item \code{is_atom_quasiconvex(Floor)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(Floor)}: Is the atom quasiconcave?

\item \code{is_incr(Floor)}: Is the atom weakly increasing in the index?

\item \code{is_decr(Floor)}: Is the atom weakly decreasing in the index?

\item \code{.grad(Floor)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}
}}

