% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/084_atoms_elementwise_ceil.R
\name{floor_expr}
\alias{floor_expr}
\title{Elementwise Floor}
\usage{
floor_expr(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
A \code{Floor} expression.
}
\description{
Returns the floor (largest integer <= x) of each element.
This atom is quasiconvex and quasiconcave but NOT convex or concave,
so it can only be used in DQCP problems (solved with \code{qcp = TRUE}).
}
\seealso{
\code{\link{ceil_expr}}
}
