% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{format_constr}
\alias{format_constr}
\title{Format Constraints}
\usage{
format_constr(object, eq_constr, leq_constr, dims, solver)
}
\arguments{
\item{object}{A \linkS4class{Constraint} object.}

\item{eq_constr}{A list of the equality constraints in the canonical problem.}

\item{leq_constr}{A list of the inequality constraints in the canonical problem.}

\item{dims}{A list with the dimensions of the conic constraints.}

\item{solver}{A string representing the solver to be called.}
}
\value{
A list containing equality constraints, inequality constraints, and dimensions.
}
\description{
Format constraints for the solver.
}
