% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{format_results}
\alias{format_results}
\title{Format Solver Results}
\usage{
format_results(solver, results_dict, data, cached_data)
}
\arguments{
\item{solver}{A \linkS4class{Solver} object.}

\item{results_dict}{A list containing the solver output.}

\item{data}{A list containing information about the problem.}

\item{cached_data}{A list mapping solver name to cached problem data.}
}
\value{
A list containing the solver output in standard form.
}
\description{
Converts the solver output into standard form.
}
