% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/126_atoms_gen_lambda_max.R
\name{gen_lambda_max}
\alias{gen_lambda_max}
\title{Maximum generalized eigenvalue}
\usage{
gen_lambda_max(A, B)
}
\arguments{
\item{A}{A square symmetric matrix expression}

\item{B}{A square PSD matrix expression of the same dimension as A}
}
\value{
An expression representing the maximum generalized eigenvalue
}
\description{
Computes the maximum generalized eigenvalue lambda_max(A, B).
Requires A symmetric and B positive semidefinite. This is a quasiconvex atom.
}
