% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{get_constr_expr}
\alias{get_constr_expr}
\title{Return the operator in the constraint left op right, i.e. rewrite as: left - right op 0}
\usage{
get_constr_expr(lh_op, rh_op = NULL)
}
\arguments{
\item{lh_op}{the operator}

\item{rh_op}{the right operator}
}
\value{
the operator in the constraint
}
\description{
Return the operator in the constraint left op right, i.e. rewrite as: left - right op 0
}
