% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{get_data}
\alias{get_data}
\title{Get Atom-Specific Data}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
List of data.
}
\description{
Get Atom-Specific Data
}
\keyword{internal}
