% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304_canonInterface.R
\name{get_default_canon_backend}
\alias{get_default_canon_backend}
\title{Returns the default canonicalization backend, which can be set globally using an environment variable.}
\usage{
get_default_canon_backend()
}
\description{
Returns the default canonicalization backend, which can be set globally using an environment variable.
}
