% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{get_expr_vars}
\alias{get_expr_vars}
\title{Get a list of the variables in the operator and their shapes}
\usage{
get_expr_vars(operator)
}
\arguments{
\item{operator}{the lhs operator}
}
\value{
a list of var id, var dim pairs
}
\description{
Get a list of the variables in the operator and their shapes
}
