% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_gurobiglue}
\alias{get_gurobiglue}
\title{Get our gurobiglue handle}
\usage{
get_gurobiglue()
}
\value{
the gurobiglue handle
}
\description{
Get the gurobiglue handle or fail if not available
}
\examples{
\dontrun{
   get_gurobiglue
}
}
