% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R, R/lin_utils.R
\name{get_id}
\alias{get_id}
\title{Get ID}
\usage{
get_id()

get_id()
}
\value{
A new unique integer identifier.

a new id
}
\description{
Get the next identifier value.
}
\examples{
\dontrun{
   get_id()
}
}
