% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonInterface.R
\name{get_matrix_and_offset_from_unparameterized_tensor}
\alias{get_matrix_and_offset_from_unparameterized_tensor}
\title{Get Matrix and Offset from Unparameterized Tensor}
\usage{
get_matrix_and_offset_from_unparameterized_tensor(
  problem_data_tensor,
  var_length
)
}
\arguments{
\item{problem_data_tensor}{tensor returned from get_problem_matrix, representing an affine map}

\item{var_length}{the number of variables}
}
\value{
A list containing A (a matrix with \code{var_length} columns) and b (a flattened array representing the constant offset).
}
\description{
Converts unparameterized tensor to matrix offset representation.
}
\details{
The \code{problem_data_tensor} must have been obtained from calling
\code{get_problem_matrix} on a problem with 0 parameters.
}
