% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304_canonInterface.R
\name{get_matrix_from_tensor}
\alias{get_matrix_from_tensor}
\title{Get Matrix from Tensor}
\usage{
get_matrix_from_tensor(
  problem_data_tensor,
  param_vec,
  var_length,
  nonzero_rows = NULL,
  with_offset = TRUE,
  problem_data_index = NULL
)
}
\arguments{
\item{problem_data_tensor}{tensor returned from get_problem_matrix, representing a parameterized affine map}

\item{param_vec}{flattened parameter vector}

\item{var_length}{the number of variables}

\item{nonzero_rows}{(optional) rows in the part of problem_data_tensor corresponding to A that have nonzeros in them (i.e., rows that are affected by parameters); if not NULL, then the corresponding entries in A will have explicit zeros.}

\item{with_offset}{(optional) return offset. Defaults to TRUE.}

\item{problem_data_index}{(optional) a tuple (indices, indptr, shape) for construction of the CSC matrix holding the problem data and offset}
}
\value{
A list containing A (a matrix with var_length columns) and b (a flattened array representing the constant offset). If with_offset=FALSE, returned b is NULL.
}
\description{
Applies problem_data_tensor to param_vec to obtain a a matrix representation of the corresponding affine map and (optionally) the offset.
}
