% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_mosekglue}
\alias{get_mosekglue}
\title{Get our mosekglue handle}
\usage{
get_mosekglue()
}
\value{
the mosekglue handle
}
\description{
Get the mosekglue handle or fail if not available
}
\examples{
\dontrun{
   get_mosekglue
}
}
