% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304_canonInterface.R
\name{get_parameter_vector}
\alias{get_parameter_vector}
\title{Get Parameter Vector}
\usage{
get_parameter_vector(
  param_size,
  param_id_to_col,
  param_id_to_size,
  param_id_to_value_fn,
  zero_offset = FALSE
)
}
\arguments{
\item{param_size}{The number of parameters}

\item{param_id_to_col}{A named list from parameter id to column offset}

\item{param_id_to_size}{A named list from parameter id to parameter size}

\item{param_id_to_value_fn}{A function that returns a value for a parameter id}

\item{zero_offset}{Logical; if TRUE, zero out the constant offset in the parameter vector (default: FALSE)}
}
\value{
A flattened numeric vector of parameter values, of length param_size + 1. Returns NULL if param_size is 0.
}
\description{
Returns a flattened parameter vector. The flattened vector includes a constant offset (i.e., a 1).
}
