% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glpk_solver.R
\docType{methods}
\name{status_map,GLPK-method}
\alias{status_map,GLPK-method}
\title{GLPK Status Map}
\usage{
\S4method{status_map}{GLPK}(solver, status)
}
\arguments{
\item{solver}{A \linkS4class{GLPK} object.}

\item{status}{An exit code returned by GLPK.}
}
\value{
A string indicating the status, either "optimal", "infeasible", "unbounded", "optimal_inaccurate", "infeasible_inaccurate", "unbounded_inaccurate", or "solver_error".
}
\description{
Map of GLPK status to CVXR status.
}
