% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\docType{class}
\name{GLPK_MI-class}
\alias{GLPK_MI-class}
\alias{GLPK_MI}
\alias{status_map,GLPK_MI-method}
\alias{name,GLPK_MI-method}
\alias{solve_via_data,GLPK_MI-method}
\title{An interface for the GLPK MI solver.}
\usage{
\S4method{status_map}{GLPK_MI}(solver, status)

\S4method{name}{GLPK_MI}(x)

\S4method{solve_via_data}{GLPK_MI}(object, data, warm_start, verbose,
  solver_opts, solver_cache = list())
}
\arguments{
\item{data}{Data generated via an apply call.}

\item{warm_start}{A boolean of whether to warm start the solver.}

\item{verbose}{A boolean of whether to enable solver verbosity.}

\item{solver_opts}{A list of Solver specific options}

\item{solver_cache}{Cache for the solver.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
An interface for the GLPK MI solver.
}
\section{Methods (by generic)}{
\itemize{
\item \code{status_map}: Converts status returned by the GLPK_MI solver to its respective CVXPY status.

\item \code{name}: Returns the name of the solver.

\item \code{solve_via_data}: Solve a problem represented by data returned from apply.
}}

