% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/078_gmatmul.R
\docType{class}
\name{GmatMul-class}
\alias{GmatMul-class}
\alias{.GmatMul}
\alias{GmatMul}
\alias{to_numeric,GmatMul-method}
\alias{name,GmatMul-method}
\alias{validate_args,GmatMul-method}
\alias{dim_from_args,GmatMul-method}
\alias{get_data,GmatMul-method}
\alias{sign_from_args,GmatMul-method}
\alias{is_atom_convex,GmatMul-method}
\alias{is_atom_concave,GmatMul-method}
\alias{parameters,GmatMul-method}
\alias{is_atom_log_log_convex,GmatMul-method}
\alias{is_atom_log_log_concave,GmatMul-method}
\alias{is_incr,GmatMul-method}
\alias{is_decr,GmatMul-method}
\alias{.grad,GmatMul-method}
\title{The GmatMul atom.}
\usage{
GmatMul(A, X)

\S4method{to_numeric}{GmatMul}(object, values)

\S4method{name}{GmatMul}(x)

\S4method{validate_args}{GmatMul}(object)

\S4method{dim_from_args}{GmatMul}(object)

\S4method{get_data}{GmatMul}(object)

\S4method{sign_from_args}{GmatMul}(object)

\S4method{is_atom_convex}{GmatMul}(object)

\S4method{is_atom_concave}{GmatMul}(object)

\S4method{parameters}{GmatMul}(object)

\S4method{is_atom_log_log_convex}{GmatMul}(object)

\S4method{is_atom_log_log_concave}{GmatMul}(object)

\S4method{is_incr}{GmatMul}(object, idx)

\S4method{is_decr}{GmatMul}(object, idx)

\S4method{.grad}{GmatMul}(object, values)
}
\arguments{
\item{A}{An \linkS4class{Expression} or numeric matrix.}

\item{X}{An \linkS4class{Expression} or numeric matrix.}

\item{object}{A \linkS4class{GmatMul} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
Geometric matrix multiplication: \eqn{A \mathbin{\diamond} X}.
}
\details{
For \eqn{A \in \mathbf{R}^{m \times n}} and \eqn{X \in \mathbf{R}^{n \times p}_{++}}, this atom represents

\deqn{\left[\begin{array}{ccc} \prod_{j=1}^n X_{j1}^{A_{1j}} & \cdots & \prod_{j=1}^n X_{pj}^{A_{1j}} \\ \vdots &  & \vdots \\ \prod_{j=1}^n X_{j1}^{A_{mj}} & \cdots & \prod_{j=1}^n X_{pj}^{A_{mj}} \end{array}\right]}

This atom is log-log affine in \eqn{X}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(GmatMul)}: Geometric matrix multiplication.

\item \code{name(GmatMul)}: The name and arguments of the atom.

\item \code{validate_args(GmatMul)}: Check if the arguments are valid.

\item \code{dim_from_args(GmatMul)}: The dimensions of the atom determined from its arguments.

\item \code{get_data(GmatMul)}: Returns the parameter \code{A}.

\item \code{sign_from_args(GmatMul)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(GmatMul)}: Is the atom convex?

\item \code{is_atom_concave(GmatMul)}: Is the atom concave?

\item \code{parameters(GmatMul)}: List of \linkS4class{Parameter} objects in the atom.

\item \code{is_atom_log_log_convex(GmatMul)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(GmatMul)}: Is the atom log-log concave?

\item \code{is_incr(GmatMul)}: Is the atom weakly increasing in the index?

\item \code{is_decr(GmatMul)}: Is the atom weakly decreasing in the index?

\item \code{.grad(GmatMul)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{A}}{An \linkS4class{Expression} representing a constant matrix.}

\item{\code{X}}{An \linkS4class{Expression} representing a positive matrix.}
}}

