% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/117_atoms_gmatmul.R
\name{gmatmul}
\alias{gmatmul}
\title{Geometric matrix multiplication A diamond X}
\usage{
gmatmul(A, X)
}
\arguments{
\item{A}{A constant matrix}

\item{X}{An Expression (positive matrix)}
}
\value{
A Gmatmul atom
}
\description{
Computes the geometric matrix product where (A diamond X)_ij = prod_k X_kj^A_ik.
Log-log affine atom for DGP. Solve with
\code{psolve(problem, gp = TRUE)}.
}
\examples{
x <- Variable(2, pos = TRUE)
A <- matrix(c(1, 0, 0, 1), 2, 2)
prob <- Problem(Minimize(sum(gmatmul(A, x))), list(x >= 0.5))
\dontrun{psolve(prob, gp = TRUE)}
}
