% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{grad}
\alias{grad}
\title{Get the Gradient of an Expression}
\usage{
grad(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Gradient information.
}
\description{
Get the Gradient of an Expression
}
\keyword{internal}
