% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{graph_implementation}
\alias{graph_implementation}
\title{Get the Graph Implementation of an Atom}
\usage{
graph_implementation(x, arg_objs, shape, data = NULL, ...)
}
\arguments{
\item{x}{An atom object.}

\item{arg_objs}{List of canonicalized argument LinOps.}

\item{shape}{Integer vector: target shape.}

\item{data}{Optional atom-specific data.}

\item{...}{Additional arguments.}
}
\value{
List with (expression, constraints).
}
\description{
Get the Graph Implementation of an Atom
}
\keyword{internal}
