% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{group_constraints}
\alias{group_constraints}
\title{Organize the constraints into a dictionary keyed by constraint names.}
\usage{
group_constraints(constraints)
}
\arguments{
\item{constraints}{a list of constraints.}
}
\value{
A list of constraint types where constr_map[[cone_type]] maps to a list.
}
\description{
Organize the constraints into a dictionary keyed by constraint names.
}
