% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gurobi-solver.R
\docType{methods}
\name{status_map,GUROBI-method}
\alias{status_map,GUROBI-method}
\title{GUROBI Status Map}
\usage{
\S4method{status_map}{GUROBI}(solver, status)
}
\arguments{
\item{solver}{A \linkS4class{GUROBI} object.}

\item{status}{An exit code returned by GUROBI. See the \href{http://www.gurobi.com/documentation/7.5/refman/optimization_status_codes.html}{GUROBI documentation} for details.}
}
\value{
A string indicating the status, either "optimal", "infeasible", "unbounded", "optimal_inaccurate", "infeasible_inaccurate", "unbounded_inaccurate", or "solver_error".
}
\description{
Map of GUROBI status to CVXR status.
}
