% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/079_harmonic_mean.R
\name{HarmonicMean}
\alias{HarmonicMean}
\title{The HarmonicMean atom.}
\usage{
HarmonicMean(x)
}
\arguments{
\item{x}{An expression or number whose harmonic mean is to be computed. Must have positive entries.}
}
\value{
The harmonic mean of \code{x}.
}
\description{
The harmonic mean of \eqn{x}, \eqn{\frac{1}{n} \sum_{i=1}^n x_i^{-1}}, where \eqn{n} is the length of \eqn{x}.
}
