% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{HermitianWrap-class}
\alias{HermitianWrap-class}
\alias{.HermitianWrap}
\alias{HermitianWrap}
\alias{validate_args,HermitianWrap-method}
\alias{is_hermitian,HermitianWrap-method}
\title{The HermitianWrap class.}
\usage{
HermitianWrap(arg)

\S4method{validate_args}{HermitianWrap}(object)

\S4method{is_hermitian}{HermitianWrap}(object)
}
\arguments{
\item{arg}{A \linkS4class{Expression} object or matrix.}

\item{object}{A \linkS4class{HermitianWrap} object.}
}
\description{
A no-op wrapper to assert the input argument is Hermitian.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args(HermitianWrap)}: Check the input is a real square matrix.

\item \code{is_hermitian(HermitianWrap)}: Is the atom Hermitian?

}}
