% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{HStack-class}
\alias{HStack-class}
\alias{.HStack}
\alias{HStack}
\alias{validate_args,HStack-method}
\alias{to_numeric,HStack-method}
\alias{size_from_args,HStack-method}
\alias{graph_implementation,HStack-method}
\title{The HStack class.}
\usage{
HStack(...)

\S4method{validate_args}{HStack}(object)

\S4method{to_numeric}{HStack}(object, values)

\S4method{size_from_args}{HStack}(object)

\S4method{graph_implementation}{HStack}(object, arg_objs, size,
  data = NA_real_)
}
\arguments{
\item{...}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of rows.}

\item{object}{A \linkS4class{HStack} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
Horizontal concatenation of values.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args}: Check all arguments have the same height.

\item \code{to_numeric}: Horizontally concatenate the values using \code{cbind}.

\item \code{size_from_args}: The size of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{...}}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of rows.}
}}

