% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/035_imag.R
\docType{class}
\name{Imag-class}
\alias{Imag-class}
\alias{.Imag}
\alias{Imag}
\alias{to_numeric,Imag-method}
\alias{dim_from_args,Imag-method}
\alias{is_imag,Imag-method}
\alias{is_complex,Imag-method}
\alias{is_symmetric,Imag-method}
\title{The Imag class.}
\usage{
Imag(expr)

\S4method{to_numeric}{Imag}(object, values)

\S4method{dim_from_args}{Imag}(object)

\S4method{is_imag}{Imag}(object)

\S4method{is_complex}{Imag}(object)

\S4method{is_symmetric}{Imag}(object)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a vector or matrix.}

\item{object}{An \linkS4class{Imag} object.}

\item{values}{A list of arguments to the atom.}
}
\description{
This class represents the imaginary part of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(Imag)}: The imaginary part of the given value.

\item \code{dim_from_args(Imag)}: The dimensions of the atom.

\item \code{is_imag(Imag)}: Is the atom imaginary?

\item \code{is_complex(Imag)}: Is the atom complex valued?

\item \code{is_symmetric(Imag)}: Is the atom symmetric?

}}
\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a vector or matrix.}
}}

