% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/085_atoms_elementwise_logic.R
\name{implies}
\alias{implies}
\title{Logical Implication}
\usage{
implies(x, y)
}
\arguments{
\item{x, y}{Boolean \link{Variable}s or logic expressions.}
}
\value{
An \link{Or} expression representing \code{!x | y}.
}
\description{
Logical implication: x => y.
Returns 1 unless x = 1 and y = 0. Equivalent to \code{Or(Not(x), y)}.
}
\examples{
\dontrun{
x <- Variable(boolean = TRUE)
y <- Variable(boolean = TRUE)
expr <- implies(x, y)
}
}
\seealso{
\code{\link[=iff]{iff()}}, \code{\link[=Not]{Not()}}, \code{\link[=And]{And()}}, \code{\link[=Or]{Or()}}, \code{\link[=Xor]{Xor()}}
}
