% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/139_transforms_indicator.R
\name{indicator}
\alias{indicator}
\title{Indicator function for constraints}
\usage{
indicator(constraints, err_tol = 0.001)
}
\arguments{
\item{constraints}{A list of constraint objects}

\item{err_tol}{Numeric tolerance for checking constraint satisfaction (default 1e-3)}
}
\value{
An Indicator expression
}
\description{
Creates an expression that equals 0 if all constraints are satisfied
and +Inf otherwise. Use this to embed constraints into the objective.
}
