% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/132_constraints_nonpos.R
\name{Inequality}
\alias{Inequality}
\title{Create an Inequality Constraint}
\usage{
Inequality(lhs, rhs, constr_id = NULL)
}
\arguments{
\item{lhs}{A CVXR expression (left-hand side).}

\item{rhs}{A CVXR expression (right-hand side).}

\item{constr_id}{Optional integer constraint ID.}
}
\value{
An \code{Inequality} constraint object.
}
\description{
Constrains the left-hand side to be less than or equal to the
right-hand side elementwise: \eqn{lhs \le rhs}. Typically created
via the \code{<=} operator on CVXR expressions.
}
\seealso{
\code{\link{Equality}}, \code{\link{NonPos}}, \code{\link{NonNeg}}
}
