% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{IntConstr-class}
\alias{IntConstr-class}
\alias{.IntConstr}
\alias{IntConstr}
\title{The IntConstr class.}
\usage{
IntConstr(lin_op)
}
\arguments{
\item{lin_op}{A list representing the linear operator equal to the \code{.noncvx_var}.}
}
\description{
This class represents an integer constraint, \eqn{X_{ij} \in \mathbf{Z}} for all \eqn{i,j}.
}
\section{Slots}{

\describe{
\item{\code{constr_id}}{(Internal) A unique integer identification number used internally.}

\item{\code{lin_op}}{A list representing the linear operator equal to the \code{.noncvx_var}.}

\item{\code{.noncvx_var}}{(Internal) A list representing the variable constrained to be elementwise integer.}
}}

