% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/023_interface_matrix_utilities.R
\name{intf_is_hermitian}
\alias{intf_is_hermitian}
\title{Check if a matrix is symmetric (and Hermitian for real case)}
\usage{
intf_is_hermitian(val)
}
\arguments{
\item{val}{A numeric matrix or sparse matrix}
}
\value{
List with \code{is_symmetric} (logical) and \code{is_hermitian} (logical)
}
\description{
Check if a matrix is symmetric (and Hermitian for real case)
}
\keyword{internal}
