% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/024_interface_matrix_utilities.R
\name{intf_is_skew_symmetric}
\alias{intf_is_skew_symmetric}
\title{Check if a matrix is skew-symmetric (A + A^T == 0)}
\usage{
intf_is_skew_symmetric(val)
}
\arguments{
\item{val}{A numeric matrix or sparse matrix}
}
\value{
Logical scalar
}
\description{
Check if a matrix is skew-symmetric (A + A^T == 0)
}
\keyword{internal}
