% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/024_interface_matrix_utilities.R
\name{intf_sign}
\alias{intf_sign}
\title{Determine the sign of a numeric value}
\usage{
intf_sign(val)
}
\arguments{
\item{val}{A numeric matrix or sparse matrix}
}
\value{
List with \code{is_nonneg} (logical) and \code{is_nonpos} (logical)
}
\description{
Determine the sign of a numeric value
}
\keyword{internal}
