% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/075_atoms_elementwise_inv_pos.R
\name{inv_pos}
\alias{inv_pos}
\title{Inverse position: \eqn{x^{-1}} (for x > 0)}
\usage{
inv_pos(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A Power atom with p=-1
}
\description{
Inverse position: \eqn{x^{-1}} (for x > 0)
}
