% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{invert}
\alias{invert}
\title{Return Original Solution}
\usage{
invert(object, solution, inverse_data)
}
\arguments{
\item{object}{A \linkS4class{Reduction} object.}

\item{solution}{A \linkS4class{Solution} to a problem that generated \code{inverse_data}.}

\item{inverse_data}{A \linkS4class{InverseData} object encoding the original problem.}
}
\value{
A \linkS4class{Solution} to the original problem.
}
\description{
Returns a solution to the original problem given the inverse data.
}
