% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080_inv_prod.R
\name{InvProd}
\alias{InvProd}
\title{The InvProd atom.}
\usage{
InvProd(value)
}
\arguments{
\item{x}{An expression or vector whose reciprocal product is to be computed. Must have positive entries.}
}
\value{
The reciprocal product of \code{x}.
}
\description{
The reciprocal of a product of the entries of a vector \eqn{x}, \eqn{(\prod_{i=1}^n x_i)^{-1}}, where \eqn{n} is the length of \eqn{x}.
}
