% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_atom_concave}
\alias{is_atom_concave}
\title{Check if Atom is Concave}
\usage{
is_atom_concave(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Atom is Concave
}
\keyword{internal}
