% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_atom_log_log_convex}
\alias{is_atom_log_log_convex}
\title{Check if Atom is Log-Log Convex}
\usage{
is_atom_log_log_convex(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Atom is Log-Log Convex
}
\keyword{internal}
