% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_complex}
\alias{is_complex}
\title{Check if Expression is Complex}
\usage{
is_complex(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Complex
}
\keyword{internal}
