% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_dcp}
\alias{is_dcp}
\title{Check if an Expression is DCP-Compliant}
\usage{
is_dcp(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Tests whether an expression follows the Disciplined Convex Programming
(DCP) rules.
}
