% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_dnlp}
\alias{is_dnlp}
\title{Check if an Expression or Problem is DNLP-Compliant}
\usage{
is_dnlp(x, ...)
}
\arguments{
\item{x}{An expression, objective, constraint, or \code{\link{Problem}}.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Tests whether the object follows the Disciplined Nonlinear Programming
(DNLP) rules: smooth representable, i.e. linearizable-convex or
linearizable-concave.
}
