% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_dqcp}
\alias{is_dqcp}
\title{Check if Expression is DQCP-Compliant}
\usage{
is_dqcp(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Tests whether an expression follows the Disciplined Quasiconvex
Programming (DQCP) rules.
}
