% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/019_atom.R
\name{is_imag,Atom-method}
\alias{is_imag,Atom-method}
\title{Is an atom imaginary?}
\usage{
\S4method{is_imag}{Atom}(object)
}
\arguments{
\item{object}{an atom}
}
\value{
a logical value indicating whether the atom is imaginary
}
\description{
Is an atom imaginary?
}
