% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_imag}
\alias{is_imag}
\title{Check if Expression is Imaginary}
\usage{
is_imag(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Imaginary
}
\keyword{internal}
