% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_log_log_convex}
\alias{is_log_log_convex}
\title{Check if Expression is Log-Log Convex}
\usage{
is_log_log_convex(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Log-Log Convex
}
