% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_problems_problem.R
\name{is_mixed_integer}
\alias{is_mixed_integer}
\title{Check if a Problem is Mixed-Integer}
\usage{
is_mixed_integer(problem)
}
\arguments{
\item{problem}{A \code{\link{Problem}} object.}
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if any variable in the problem has a
\code{boolean} or \code{integer} attribute.
}
