% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/019_atom.R
\name{is_nonneg,Atom-method}
\alias{is_nonneg,Atom-method}
\title{Is an atom non_negative?}
\usage{
\S4method{is_nonneg}{Atom}(object)
}
\arguments{
\item{object}{an atom}
}
\value{
a logical value indicating whether the atom is nonnegative
}
\description{
Is an atom non_negative?
}
