% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/031_expressions_constants_parameter.R
\name{is_param_free}
\alias{is_param_free}
\title{Check if Expression is Parameter-Free}
\usage{
is_param_free(expr)
}
\arguments{
\item{expr}{A CVXR expression.}
}
\value{
Logical scalar.
}
\description{
Returns \code{TRUE} if the expression contains no parameters.
}
\keyword{internal}
