% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_psd}
\alias{is_psd}
\title{Check if Expression is Positive Semidefinite}
\usage{
is_psd(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Positive Semidefinite
}
