% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R,
%   R/297_zzz_R_specific_exports.R
\name{is_pwl}
\alias{is_pwl}
\title{Check if Expression is Piecewise Linear}
\usage{
is_pwl(x, ...)
}
\arguments{
\item{x}{A CVXR expression.}

\item{...}{Not used.}
}
\value{
Logical scalar.

Logical.
}
\description{
Check if Expression is Piecewise Linear

Is the Expression Piecewise Linear?
}
