% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_qp}
\alias{is_qp}
\title{Check if a Problem is a Quadratic Program}
\usage{
is_qp(x, ...)
}
\arguments{
\item{x}{A \code{\link{Problem}} object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if a Problem is a Quadratic Program
}
