% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_qpwa}
\alias{is_qpwa}
\title{Check if Expression is Quadratic or Piecewise Affine}
\usage{
is_qpwa(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Quadratic or Piecewise Affine
}
\keyword{internal}
