% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_quadratic}
\alias{is_quadratic}
\title{Check if an Expression is Quadratic}
\usage{
is_quadratic(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if an Expression is Quadratic
}
