% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{is_skew_symmetric}
\alias{is_skew_symmetric}
\title{Check if Expression is Skew-Symmetric}
\usage{
is_skew_symmetric(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Tests whether \code{X + t(X) == 0} (real matrices only).
CVXPY SOURCE: expression.py line 470-473
}
\keyword{internal}
