% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/295_zzz_R_extra.R
\name{is_stuffed_cone_constraint}
\alias{is_stuffed_cone_constraint}
\title{Is the constraint a stuffed cone constraint?}
\usage{
is_stuffed_cone_constraint(constraint)
}
\arguments{
\item{constraint}{A \linkS4class{Constraint} object.}
}
\value{
Is the constraint a stuffed-cone constraint?
}
\description{
Is the constraint a stuffed cone constraint?
}
