% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/295_zzz_R_extra.R
\name{is_stuffed_cone_objective}
\alias{is_stuffed_cone_objective}
\title{Is the objective a stuffed cone objective?}
\usage{
is_stuffed_cone_objective(objective)
}
\arguments{
\item{objective}{An \linkS4class{Objective} object.}
}
\value{
Is the objective a stuffed-cone objective?
}
\description{
Is the objective a stuffed cone objective?
}
