% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_solver.R
\name{is_stuffed_qp_objective}
\alias{is_stuffed_qp_objective}
\title{Is the QP objective stuffed?}
\usage{
is_stuffed_qp_objective(objective)
}
\arguments{
\item{objective}{A \linkS4class{Minimize} or \linkS4class{Maximize} object representing the optimization objective.}
}
\value{
Is the objective a stuffed QP?
}
\description{
Is the QP objective stuffed?
}
