% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069_atoms_elementwise_kl_div.R
\name{kl_div}
\alias{kl_div}
\title{KL Divergence: x*log(x/y) - x + y}
\usage{
kl_div(x, y)
}
\arguments{
\item{x}{An Expression}

\item{y}{An Expression}
}
\value{
A KlDiv atom
}
\description{
KL Divergence: x*log(x/y) - x + y
}
