% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053_atoms_affine_kron.R
\name{kron}
\alias{kron}
\title{Kronecker product of two expressions}
\usage{
kron(a, b)
}
\arguments{
\item{a}{An Expression (one must be constant)}

\item{b}{An Expression}
}
\value{
A Kron atom
}
\description{
Kronecker product of two expressions
}
