% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/108_atoms_lambda_max.R
\name{lambda_min}
\alias{lambda_min}
\title{Minimum eigenvalue}
\usage{
lambda_min(A)
}
\arguments{
\item{A}{A square matrix expression}
}
\value{
An expression representing the minimum eigenvalue of A
}
\description{
Minimum eigenvalue
}
