% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{LambdaMin}
\alias{LambdaMin}
\title{The LambdaMin atom.}
\usage{
LambdaMin(A)
}
\arguments{
\item{A}{An \linkS4class{Expression} or numeric matrix.}
}
\value{
Returns the minimum eigenvalue of a matrix.
}
\description{
The minimum eigenvalue of a matrix, \eqn{\lambda_{\min}(A)}.
}
