% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/084_lambda_sum_smallest.R
\name{LambdaSumSmallest}
\alias{LambdaSumSmallest}
\title{The LambdaSumSmallest atom.}
\usage{
LambdaSumSmallest(A, k)
}
\arguments{
\item{A}{An \linkS4class{Expression} or numeric matrix.}

\item{k}{A positive integer.}
}
\value{
Returns the sum of the k smallest eigenvalues of a matrix.
}
\description{
This class represents the sum of the \code{k} smallest eigenvalues of a matrix.
}
