% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{leaf-attr}
\alias{leaf-attr}
\alias{is_pos}
\alias{is_neg}
\title{Attributes of an Expression Leaf}
\usage{
is_pos(object)

is_neg(object)
}
\arguments{
\item{object}{A \linkS4class{Leaf} object.}
}
\value{
A logical value.
}
\description{
Determine if an expression is positive or negative.
}
