% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011_leaf.R
\docType{class}
\name{Leaf-class}
\alias{Leaf-class}
\alias{Leaf}
\alias{get_data,Leaf-method}
\alias{dim,Leaf-method}
\alias{variables,Leaf-method}
\alias{parameters,Leaf-method}
\alias{constants,Leaf-method}
\alias{atoms,Leaf-method}
\alias{is_convex,Leaf-method}
\alias{is_concave,Leaf-method}
\alias{is_log_log_convex,Leaf-method}
\alias{is_log_log_concave,Leaf-method}
\alias{is_nonneg,Leaf-method}
\alias{is_nonpos,Leaf-method}
\alias{is_pos,Leaf-method}
\alias{is_neg,Leaf-method}
\alias{is_hermitian,Leaf-method}
\alias{is_symmetric,Leaf-method}
\alias{is_imag,Leaf-method}
\alias{is_complex,Leaf-method}
\alias{domain,Leaf-method}
\alias{project,Leaf-method}
\alias{value,Leaf-method}
\alias{value<-,Leaf-method}
\alias{project_and_assign,Leaf-method}
\alias{validate_val,Leaf-method}
\alias{is_psd,Leaf-method}
\alias{is_nsd,Leaf-method}
\alias{is_quadratic,Leaf-method}
\alias{has_quadratic_term,Leaf-method}
\alias{is_pwl,Leaf-method}
\alias{is_dpp,Leaf-method}
\title{The Leaf class.}
\usage{
\S4method{get_data}{Leaf}(object)

\S4method{dim}{Leaf}(x)

\S4method{variables}{Leaf}(object)

\S4method{parameters}{Leaf}(object)

\S4method{constants}{Leaf}(object)

\S4method{atoms}{Leaf}(object)

\S4method{is_convex}{Leaf}(object)

\S4method{is_concave}{Leaf}(object)

\S4method{is_log_log_convex}{Leaf}(object)

\S4method{is_log_log_concave}{Leaf}(object)

\S4method{is_nonneg}{Leaf}(object)

\S4method{is_nonpos}{Leaf}(object)

\S4method{is_pos}{Leaf}(object)

\S4method{is_neg}{Leaf}(object)

\S4method{is_hermitian}{Leaf}(object)

\S4method{is_symmetric}{Leaf}(object)

\S4method{is_imag}{Leaf}(object)

\S4method{is_complex}{Leaf}(object)

\S4method{domain}{Leaf}(object)

\S4method{project}{Leaf}(object, value)

\S4method{value}{Leaf}(object)

\S4method{value}{Leaf}(object) <- value

\S4method{project_and_assign}{Leaf}(object, value)

\S4method{validate_val}{Leaf}(object, val)

\S4method{is_psd}{Leaf}(object)

\S4method{is_nsd}{Leaf}(object)

\S4method{is_quadratic}{Leaf}(object)

\S4method{has_quadratic_term}{Leaf}(object)

\S4method{is_pwl}{Leaf}(object)

\S4method{is_dpp}{Leaf}(object, context = c("dcp", "dgp"))

\S4method{atoms}{Leaf}(object)
}
\arguments{
\item{object, x}{A \linkS4class{Leaf} object.}

\item{value}{A numeric scalar, vector, or matrix.}

\item{val}{The assigned value.}
}
\description{
This class represents a leaf node, i.e. a Variable, Constant, or Parameter.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_data(Leaf)}: Leaves are not copied.

\item \code{dim(Leaf)}: The dimensions of the leaf node.

\item \code{variables(Leaf)}: List of \linkS4class{Variable} objects in the leaf node.

\item \code{parameters(Leaf)}: List of \linkS4class{Parameter} objects in the leaf node.

\item \code{constants(Leaf)}: List of \linkS4class{Constant} objects in the leaf node.

\item \code{atoms(Leaf)}: List of \linkS4class{Atom} objects in the leaf node.

\item \code{is_convex(Leaf)}: A logical value indicating whether the leaf node is convex.

\item \code{is_concave(Leaf)}: A logical value indicating whether the leaf node is concave.

\item \code{is_log_log_convex(Leaf)}: Is the expression log-log convex?

\item \code{is_log_log_concave(Leaf)}: Is the expression log-log concave?

\item \code{is_nonneg(Leaf)}: A logical value indicating whether the leaf node is nonnegative.

\item \code{is_nonpos(Leaf)}: A logical value indicating whether the leaf node is nonpositive.

\item \code{is_pos(Leaf)}: Is the expression positive?

\item \code{is_neg(Leaf)}: Is the expression negative?

\item \code{is_hermitian(Leaf)}: A logical value indicating whether the leaf node is hermitian.

\item \code{is_symmetric(Leaf)}: A logical value indicating whether the leaf node is symmetric.

\item \code{is_imag(Leaf)}: A logical value indicating whether the leaf node is imaginary.

\item \code{is_complex(Leaf)}: A logical value indicating whether the leaf node is complex.

\item \code{domain(Leaf)}: A list of constraints describing the closure of the region where the leaf node is finite. Default is the full domain.

\item \code{project(Leaf)}: Project value onto the attribute set of the leaf.

\item \code{value(Leaf)}: Get the value of the leaf.

\item \code{value(Leaf) <- value}: Set the value of the leaf.

\item \code{project_and_assign(Leaf)}: Project and assign a value to the leaf.

\item \code{validate_val(Leaf)}: Check that \code{val} satisfies symbolic attributes of leaf.

\item \code{is_psd(Leaf)}: A logical value indicating whether the leaf node is a positive semidefinite matrix.

\item \code{is_nsd(Leaf)}: A logical value indicating whether the leaf node is a negative semidefinite matrix.

\item \code{is_quadratic(Leaf)}: Leaf nodes are always quadratic.

\item \code{has_quadratic_term(Leaf)}: Leaf nodes are not quadratic terms.

\item \code{is_pwl(Leaf)}: Leaf nodes are always piecewise linear.

\item \code{is_dpp(Leaf)}: Is the expression a disciplined parametrized expression?

\item \code{atoms(Leaf)}: A list of the atoms in the expression.

}}
\section{Slots}{

\describe{
\item{\code{id}}{(Internal) A unique integer identification number used internally.}

\item{\code{dim}}{The dimensions of the leaf.}

\item{\code{value}}{The numeric value of the leaf.}

\item{\code{nonneg}}{Is the leaf constrained to be nonnegative?}

\item{\code{nonpos}}{Is the leaf constrained to be nonpositive?}

\item{\code{complex}}{Is the leaf complex valued?}

\item{\code{imag}}{Is the leaf imaginary?}

\item{\code{symmetric}}{Is the leaf symmetric?}

\item{\code{diag}}{Is the leaf diagonal?}

\item{\code{PSD}}{Is the leaf constrained to be positive semidefinite?}

\item{\code{NSD}}{Is the leaf constrained to be negative semidefinite?}

\item{\code{hermitian}}{Is the leaf Hermitian?}

\item{\code{boolean}}{Is the leaf boolean? Can be \code{TRUE} = entire leaf is boolean, \code{FALSE} = entire leaf is not boolean, or a vector of
indices which should be constrained as boolean, where each index is a vector of length exactly equal to the length of \code{dim}.}

\item{\code{integer}}{Is the leaf integral? The semantics are the same as the \code{boolean} argument.}

\item{\code{sparsity}}{A matrix representing the fixed sparsity pattern of the leaf.}

\item{\code{pos}}{Is the leaf positive?}

\item{\code{neg}}{Is the leaf negative?}
}}

