% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.conv}
\alias{lo.conv}
\title{Discrete convolution of two vectors}
\usage{
lo.conv(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{the left-hand operator in the convolution}

\item{rh_op}{the right-hand operator in the convolution}

\item{dim}{the shape of the LinOp after convolution}
}
\value{
a LinOp representing the convolution
}
\description{
Discrete convolution of two vectors
}
