% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.create_const}
\alias{lo.create_const}
\title{Wrap a constant}
\usage{
lo.create_const(value, dim, sparse = FALSE)
}
\arguments{
\item{value}{a scalar, matrix or sparse matrix to wrap}

\item{dim}{a tuple, the (rows, cols) dimensions of the variable}

\item{sparse}{a boolean flag indicating if sparse matrix or not}
}
\value{
a LinOp wrapping the constant
}
\description{
Wrap a constant
}
