% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.create_param}
\alias{lo.create_param}
\title{Wrap a parameter.  (This is the v1.3.0 version and the 1.5.2 version is different!)}
\usage{
lo.create_param(dim, param_id = get_id())
}
\arguments{
\item{dim}{a tuple, the (rows, cols) dimensions of the variable}

\item{param_id}{the id of the variable}
}
\value{
a LinOp wrapping the parameter
}
\description{
Wrap a parameter.  (This is the v1.3.0 version and the 1.5.2 version is different!)
}
