% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.create_var}
\alias{lo.create_var}
\title{Create a new internal variable}
\usage{
lo.create_var(dim, var_id = get_id())
}
\arguments{
\item{dim}{a tuple, the (rows, cols) dimensions of the variable}

\item{var_id}{the id of the variable}
}
\value{
a LinOp representing the new variable
}
\description{
Create a new internal variable
}
