% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.diag_mat}
\alias{lo.diag_mat}
\title{Convert the diagonal of a LinOp matrix to a vector}
\usage{
lo.diag_mat(operator)
}
\arguments{
\item{operator}{the operator to convert}
}
\value{
a LinOp representing the matrix diagonal
}
\description{
Convert the diagonal of a LinOp matrix to a vector
}
