% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.diag_vec}
\alias{lo.diag_vec}
\title{Convert a LinOp vector into a diagonal matrix. (Note version different for cvxpy v1.5.2 and this is v1.3.0 version)}
\usage{
lo.diag_vec(operator)
}
\arguments{
\item{operator}{the operator to convert}
}
\value{
a LinOp representing the matrix diagonal
}
\description{
Convert a LinOp vector into a diagonal matrix. (Note version different for cvxpy v1.5.2 and this is v1.3.0 version)
}
