% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.div_expr}
\alias{lo.div_expr}
\title{Divide one LinOp by another}
\usage{
lo.div_expr(lh_op, rh_op)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}
}
\value{
a LinOp representing the product
}
\description{
Divide one LinOp by another
}
