% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.hstack}
\alias{lo.hstack}
\title{Concatenate operators horizontally, (like `cbind`)}
\usage{
lo.hstack(operators, dim)
}
\arguments{
\item{operators}{the operators to horizontally stack}

\item{dim}{the (rows, cols) of the stacked operators}
}
\value{
a LinOp representing the stacked expression
}
\description{
Concatenate operators horizontally, (like `cbind`)
}
