% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.index}
\alias{lo.index}
\title{Indexes/slices an operator}
\usage{
lo.index(operator, dim, keys)
}
\arguments{
\item{operator}{a LinOp}

\item{dim}{the shape of the LinOp after indexing/slicing}

\item{keys}{the (row slice, column slice)}
}
\value{
a LinOp representing the index/slice
}
\description{
Indexes/slices an operator
}
