% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.kron_r}
\alias{lo.kron_r}
\title{Construct Kronecker product of two matrices, where the right operand is a Variable}
\usage{
lo.kron_r(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}

\item{dim}{the dimension of the result}
}
\value{
a LinOp representing the kronecker product
}
\description{
Construct Kronecker product of two matrices, where the right operand is a Variable
}
