% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.multiply}
\alias{lo.multiply}
\title{Multiply two LinOps elementwise}
\usage{
lo.multiply(lh_op, rh_op)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}
}
\value{
a LinOp representing the product
}
\description{
Multiply two LinOps elementwise
}
