% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.promote}
\alias{lo.promote}
\title{Promote a scalar operator to a specified shape}
\usage{
lo.promote(operator, dim)
}
\arguments{
\item{operator}{a LinOp}

\item{dim}{the desired shape}
}
\value{
a LinOp representing the promotion
}
\description{
Promote a scalar operator to a specified shape
}
