% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.promote_lin_ops_for_mul}
\alias{lo.promote_lin_ops_for_mul}
\title{Promote LinOp arguments for multiplication}
\usage{
lo.promote_lin_ops_for_mul(lh_op, rh_op)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}
}
\value{
a list of promoted LinOps respectively and the new dimension
}
\description{
Promote LinOp arguments for multiplication
}
