% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.redim}
\alias{lo.redim}
\title{Reshape an operator}
\usage{
lo.redim(operator, dim)
}
\arguments{
\item{operator}{the operator to reshape}

\item{dim}{the (rows, cols) of the reshaped operator}
}
\value{
a LinOp representing the reshaped operator
}
\description{
Reshape an operator
}
