% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.rmul_expr}
\alias{lo.rmul_expr}
\title{Multiply two LinOps, with constant on right}
\usage{
lo.rmul_expr(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}

\item{dim}{the dim of the result}
}
\value{
a LinOp representing the product
}
\description{
Multiply two LinOps, with constant on right
}
