% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.sub_expr}
\alias{lo.sub_expr}
\title{Subtract one LinOp from another}
\usage{
lo.sub_expr(lh_op, rh_op)
}
\arguments{
\item{lh_op}{a LinOp}

\item{rh_op}{a LinOp}
}
\value{
a LinOp representing the subtracted result
}
\description{
Subtract one LinOp from another
}
