% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.sum_entries}
\alias{lo.sum_entries}
\title{Sum the entries of an operator}
\usage{
lo.sum_entries(operator, dim)
}
\arguments{
\item{operator}{a LinOp}

\item{dim}{the dimensions of the sum}
}
\value{
a LinOp representing the sum
}
\description{
Sum the entries of an operator
}
