% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.sum_expr}
\alias{lo.sum_expr}
\title{Sum a list of LinOps}
\usage{
lo.sum_expr(operators)
}
\arguments{
\item{operators}{a list of LinOps}
}
\value{
a LinOp representing the sum
}
\description{
Sum a list of LinOps
}
