% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.trace}
\alias{lo.trace}
\title{Sum the diagonal entries of an operator}
\usage{
lo.trace(operator)
}
\arguments{
\item{operator}{a LinOp}
}
\value{
a LinOp representing the sum
}
\description{
Sum the diagonal entries of an operator
}
