% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.transpose}
\alias{lo.transpose}
\title{Trnaspose an operator}
\usage{
lo.transpose(operator)
}
\arguments{
\item{operator}{a LinOp}
}
\value{
a LinOp representing the transpose
}
\description{
Trnaspose an operator
}
