% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.upper_tri}
\alias{lo.upper_tri}
\title{Create vectorized upper triangular portion of a square matrix (excluding diagonal)}
\usage{
lo.upper_tri(operator)
}
\arguments{
\item{operator}{the operator to convert}
}
\value{
a LinOp representing the vectorized upper triangle
}
\description{
Create vectorized upper triangular portion of a square matrix (excluding diagonal)
}
