% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lo.vstack}
\alias{lo.vstack}
\title{Concatenate operators vertically, (like `rbind`)}
\usage{
lo.vstack(operators, dim)
}
\arguments{
\item{operators}{the operators to vertically stack}

\item{dim}{the (rows, cols) of the stacked operators}
}
\value{
a LinOp representing the stacked expression
}
\description{
Concatenate operators vertically, (like `rbind`)
}
