% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{Log1p-class}
\alias{Log1p-class}
\alias{.Log1p}
\alias{Log1p}
\alias{to_numeric,Log1p-method}
\alias{sign_from_args,Log1p-method}
\alias{.grad,Log1p-method}
\alias{.domain,Log1p-method}
\title{The Log1p class.}
\usage{
Log1p(x)

\S4method{to_numeric}{Log1p}(object, values)

\S4method{sign_from_args}{Log1p}(object)

\S4method{.grad}{Log1p}(object, values)

\S4method{.domain}{Log1p}(object)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric constant.}

\item{object}{A \linkS4class{Log1p} object.}

\item{values}{A list of arguments to the atom.}

\item{values}{A list of numeric values for the arguments}
}
\description{
This class represents the elementwise operation \eqn{\log(1 + x)}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: The elementwise natural logarithm of one plus the input value.

\item \code{sign_from_args}: The sign of the atom.

\item \code{.grad}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.domain}: Returns constraints describng the domain of the node
}}

\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} or numeric constant.}
}}

