% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\name{log_log_curvature-atom}
\alias{log_log_curvature-atom}
\alias{is_atom_log_log_convex}
\alias{is_atom_log_log_concave}
\alias{is_atom_log_log_affine}
\title{Log-Log Curvature of an Atom}
\usage{
is_atom_log_log_convex(object)

is_atom_log_log_concave(object)

is_atom_log_log_affine(object)
}
\arguments{
\item{object}{A \linkS4class{Atom} object.}
}
\value{
A logical value.
}
\description{
Determine if an atom is log-log convex, concave, or affine.
}
