% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/expressions.R
\docType{methods}
\name{log_log_curvature}
\alias{log_log_curvature}
\alias{log_log_curvature,Expression-method}
\title{Log-Log Curvature of Expression}
\usage{
log_log_curvature(object)

\S4method{log_log_curvature}{Expression}(object)
}
\arguments{
\item{object}{An \linkS4class{Expression} object.}
}
\value{
A string indicating the log-log curvature of the expression, either "LOG_LOG_CONSTANT", "LOG_LOG_AFFINE", "LOG_LOG_CONVEX, "LOG_LOG_CONCAVE", or "UNKNOWN".

A string indicating the log-log curvature of the expression, either "LOG_LOG_CONSTANT", "LOG_LOG_AFFINE", "LOG_LOG_CONVEX", "LOG_LOG_CONCAVE", or "UNKNOWN".
}
\description{
The log-log curvature of an expression.

The log-log curvature of an expression.
}
