% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/074_atoms_elementwise_logistic.R
\name{logistic}
\alias{logistic}
\title{Logistic function: log(1 + exp(x)) -- elementwise}
\usage{
logistic(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A Logistic atom
}
\description{
Logistic function: log(1 + exp(x)) -- elementwise
}
