% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\name{LogNormCdf}
\alias{LogNormCdf}
\title{The LogNormCdf atom.}
\usage{
LogNormCdf(x)
}
\arguments{
\item{x}{An \linkS4class{Expression} object.}
}
\value{
An expression representing the log-norm of \code{x}.
}
\description{
The elementwise log of the cumulative distribution function of a standard normal random variable.
}
\details{
This implementation is a quadratic approximation with modest accuracy over \link{-4, 4}.
For details on the nature of the approximation, refer to CVXPY GitHub PR #1224
https://github.com/cvxpy/cvxpy/pull/1224#issue-793221374.

Note: Python SciPy's analog of LogNormCdf is called log_ndtr:
https://docs.scipy.org/doc/scipy/reference/generated/scipy.special.log_ndtr.html
}
