% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.conv}
\alias{lu.conv}
\title{Create 1D discrete convolution of two vectors}
\usage{
lu.conv(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{the left-hand operator in the convolution}

\item{rh_op}{the right-hand operator in the convolution}

\item{dim}{the dimension of the convolution}
}
\value{
a linear operator representing the convolution
}
\description{
Create 1D discrete convolution of two vectors
}
