% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.copy_constr}
\alias{lu.copy_constr}
\title{Creates a copy of the constraint modified according to func}
\usage{
lu.copy_constr(constr, func)
}
\arguments{
\item{constr}{the `LinConstraint` constraint to modify}

\item{func}{a function to modify the constraint expression}
}
\value{
a copy of the constraint with the specified changes
}
\description{
Creates a copy of the constraint modified according to func
}
