% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.create_const}
\alias{lu.create_const}
\title{Wrap a constant}
\usage{
lu.create_const(value, dim, sparse = FALSE)
}
\arguments{
\item{value}{a scalar, matrix or sparse matrix}

\item{dim}{the rows, cols dimensions of the constant as a vector, in order}

\item{sparse}{a flag indicating if object is sparse matrix or not}
}
\value{
a LinOp wrapping the constant
}
\description{
Wrap a constant
}
