% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.create_geq}
\alias{lu.create_geq}
\title{Create an internal greater than or equal to constraint as a LinOp}
\usage{
lu.create_geq(lh_op, rh_op, constr_id = lu.get_id())
}
\arguments{
\item{lh_op}{the left-hand operator in the constraint}

\item{rh_op}{the right-hand operator in the constraint}

\item{constr_id}{a id for the constraint, automatically generated if not specified}
}
\value{
the LinOp representing the greater than or equal to constraint
}
\description{
Create an internal greater than or equal to constraint as a LinOp
}
