% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.create_param}
\alias{lu.create_param}
\title{Wrap a parameter}
\usage{
lu.create_param(dim, param_id = lu.get_id())
}
\arguments{
\item{dim}{the rows, cols dimensions of the parameter as a vector in order}

\item{id}{the id of the parameter, automatically generated if not specified}
}
\value{
a LinOp wrapping the parameter
}
\description{
Wrap a parameter
}
