% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.create_var}
\alias{lu.create_var}
\title{Create a new internal variable}
\usage{
lu.create_var(dim, var_id = lu.get_id())
}
\arguments{
\item{dim}{the rows, cols dimensions of the variable as a vector, in order}

\item{id}{the id of the variable, automatically generated if not specified}
}
\value{
a LinOp representing the new variable
}
\description{
Create a new internal variable
}
