% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.diag_mat}
\alias{lu.diag_mat}
\title{Convert the diagonal of a matrix to a vector}
\usage{
lu.diag_mat(operator)
}
\arguments{
\item{operator}{the operator to convert to a vector}
}
\value{
LinOp representing the matrix diagonal
}
\description{
Convert the diagonal of a matrix to a vector
}
