% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.diag_vec}
\alias{lu.diag_vec}
\title{Convert a vector to a diagonal matrix}
\usage{
lu.diag_vec(operator)
}
\arguments{
\item{operator}{he operator to convert to a diagonal matrix}
}
\value{
LinOp representing the diagonal matrix
}
\description{
Convert a vector to a diagonal matrix
}
