% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.div_expr}
\alias{lu.div_expr}
\title{Divide one linear operator by another where the right operand is a scalar constant}
\usage{
lu.div_expr(lh_op, rh_op)
}
\arguments{
\item{lh_op}{the left-hand operator in the quotient}

\item{rh_op}{the right-hand operator in the quotient}
}
\value{
a linear operator representing the quotient
}
\description{
Divide one linear operator by another where the right operand is a scalar constant
}
