% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.get_constr_expr}
\alias{lu.get_constr_expr}
\title{Return a constraint as a LinOp}
\usage{
lu.get_constr_expr(lh_op, rh_op)
}
\arguments{
\item{lh_op}{the left-hand operator in the constraint}

\item{rh_op}{the right-hand operator in the constraint}
}
\value{
the LinOp representing the constraint
}
\description{
Return a constraint as a LinOp
}
