% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.get_expr_params}
\alias{lu.get_expr_params}
\title{Get a list of parameters in the operator}
\usage{
lu.get_expr_params(operator)
}
\arguments{
\item{operator}{the operator whose parameters are needed}
}
\value{
a list of parameters in the operator
}
\description{
Get a list of parameters in the operator
}
