% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.get_expr_vars}
\alias{lu.get_expr_vars}
\title{Get a list of variables in the operator and their dimensions}
\usage{
lu.get_expr_vars(operator)
}
\arguments{
\item{operator}{the operator whose variables and dimensions are needed}
}
\value{
a list of var_id, var dim pairs
}
\description{
Get a list of variables in the operator and their dimensions
}
