% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.hstack}
\alias{lu.hstack}
\title{Concatenate operators horizontally}
\usage{
lu.hstack(operators, dim)
}
\arguments{
\item{operators}{the operators to concatenate}

\item{dim}{the dimensions of the result}
}
\value{
LinOp representing the horizontally stacked operators
}
\description{
Concatenate operators horizontally
}
