% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.index}
\alias{lu.index}
\title{Index/slice an operator}
\usage{
lu.index(operator, dim, keys)
}
\arguments{
\item{operator}{the expression to index}

\item{dim}{the dimensions of the expression after indexing}

\item{keys}{a list of row slice, column slice}
}
\value{
an operator representing the indexing
}
\description{
Index/slice an operator
}
