% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.is_const}
\alias{lu.is_const}
\title{Is the LinOp a constant?}
\usage{
lu.is_const(operator)
}
\arguments{
\item{operator}{the LinOp to test}
}
\value{
`TRUE` if a constant, else `FALSE`
}
\description{
Is the LinOp a constant?
}
