% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.is_scalar}
\alias{lu.is_scalar}
\title{Is the LinOp a scalar?}
\usage{
lu.is_scalar(operator)
}
\arguments{
\item{operator}{the LinOp to test}
}
\value{
\code{TRUE} if a scalar, else \code{FALSE}
}
\description{
Is the LinOp a scalar?
}
