% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.kron_l}
\alias{lu.kron_l}
\title{Compute the Kronecker product of two matrices, where the left operand is a variable}
\usage{
lu.kron_l(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{the left-hand operator in the product}

\item{rh_op}{the right-hand operator in the product}

\item{dim}{the dimension of the product}
}
\value{
a linear operator representing the Kronecker product
}
\description{
Compute the Kronecker product of two matrices, where the left operand is a variable
}
