% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.mul_expr}
\alias{lu.mul_expr}
\title{Multiply two linear operators, with the constant on the left}
\usage{
lu.mul_expr(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{the left-hand operator in the product}

\item{rh_op}{the right-hand operator in the product}

\item{dim}{the shape of the product}
}
\value{
a linear operator representing the product
}
\description{
Multiply two linear operators, with the constant on the left
}
