% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.multiply}
\alias{lu.multiply}
\title{Multiply two linear operators, elementwise}
\usage{
lu.multiply(lh_op, rh_op)
}
\arguments{
\item{lh_op}{the left-hand operator in the product}

\item{rh_op}{the right-hand operator in the product}
}
\value{
a linear operator representing the product
}
\description{
Multiply two linear operators, elementwise
}
