% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.neg_expr}
\alias{lu.neg_expr}
\title{Negate an operator}
\usage{
lu.neg_expr(operator)
}
\arguments{
\item{operator}{the LinOp to negate}
}
\value{
the negated operator
}
\description{
Negate an operator
}
