% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.promote}
\alias{lu.promote}
\title{Promote a scalar operator to the specified dimensions}
\usage{
lu.promote(operator, dim)
}
\arguments{
\item{operator}{the operator to promote}

\item{dim}{the dimensions to promote to}
}
\value{
a linear operator representing the promotion
}
\description{
Promote a scalar operator to the specified dimensions
}
