% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.promote_lin_ops_for_mul}
\alias{lu.promote_lin_ops_for_mul}
\title{Promote arguments for multiplication}
\usage{
lu.promote_lin_ops_for_mul(lh_op, rh_op)
}
\arguments{
\item{lh_op}{LinOp
The left-hand operator in the multiplication}

\item{rh_op}{LinOp
The right-hand operator in the multiplication}
}
\value{
A list containing the following elements:
  \item{lh_op}{Promoted left-hand operator}
  \item{rh_op}{Promoted right-hand operator}
  \item{dim}{Shape of the product}
}
\description{
Promote arguments for multiplication
}
