% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.replace_new_vars}
\alias{lu.replace_new_vars}
\title{Replace the given variables in the expression}
\usage{
lu.replace_new_vars(expr, id_to_new_var)
}
\arguments{
\item{expr}{the `LinOp` expression to replace variables in}

\item{id_to_new_var}{a map of id to new variable}
}
\value{
an LinOp identical to expr, but with the given variables replaced
}
\description{
Replace the given variables in the expression
}
