% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.replace_params_with_consts}
\alias{lu.replace_params_with_consts}
\title{Replace parameters with constant nodes}
\usage{
lu.replace_params_with_consts(expr)
}
\arguments{
\item{expr}{the \code{LinOp} expression to replace parameters in}
}
\value{
an LinOp identical to expr, but with the parameters replaced
}
\description{
Replace parameters with constant nodes
}
