% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.reshape}
\alias{lu.reshape}
\title{Reshape an operator}
\usage{
lu.reshape(operator, dim)
}
\arguments{
\item{operator}{the operator to reshape}

\item{dim}{the dimensions of the reshaped operator}
}
\value{
an operator representing the reshaping
}
\description{
Reshape an operator
}
