% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.rmul_expr}
\alias{lu.rmul_expr}
\title{Multiply two linear operators, with the constant on the right}
\usage{
lu.rmul_expr(lh_op, rh_op, dim)
}
\arguments{
\item{lh_op}{the left-hand operator in the product}

\item{rh_op}{the right-hand operator in the product}
}
\value{
a linear operator representing the product
}
\description{
Multiply two linear operators, with the constant on the right
}
