% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.sub_expr}
\alias{lu.sub_expr}
\title{Subtract two operators}
\usage{
lu.sub_expr(lh_op, rh_op)
}
\arguments{
\item{lh_op}{the LinOp on LHS}

\item{rh_op}{the LinOp on RHS}
}
\value{
the resulting subtracted operator
}
\description{
Subtract two operators
}
