% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.sum_entries}
\alias{lu.sum_entries}
\title{Sum the entries of the given operator.}
\usage{
lu.sum_entries(operator, dim)
}
\arguments{
\item{operator}{the operator to sum the entries of}

\item{dim}{the dimensions of the sum}
}
\value{
an operator representing the sum
}
\description{
Sum the entries of the given operator.
}
