% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.sum_expr}
\alias{lu.sum_expr}
\title{Add linear operators}
\usage{
lu.sum_expr(operators)
}
\arguments{
\item{a}{list of linear operators to sum}
}
\value{
a LinOp representing the sum of the operators
}
\description{
Add linear operators
}
