% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.trace}
\alias{lu.trace}
\title{Sum the diagonal entries of an operator}
\usage{
lu.trace(operator)
}
\arguments{
\item{expr}{the operator to sum the diagonal entries of}
}
\value{
an operator representing the sum of the diagonal entries
}
\description{
Sum the diagonal entries of an operator
}
