% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/133_lin_utils.R
\name{lu.transpose}
\alias{lu.transpose}
\title{Transpose an operator}
\usage{
lu.transpose(operator)
}
\arguments{
\item{operator}{the operator to sum the entries of}
}
\value{
an operator representing the transpose
}
\description{
Transpose an operator
}
