% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.upper_tri}
\alias{lu.upper_tri}
\title{Create vectorized upper triangular portion of a square matrix}
\usage{
lu.upper_tri(operator)
}
\arguments{
\item{operator}{the matrix operator}
}
\value{
LinOp representing the vectorized upper triangle
}
\description{
Create vectorized upper triangular portion of a square matrix
}
