% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_utils.R
\name{lu.vstack}
\alias{lu.vstack}
\title{Concatenate operators vertically}
\usage{
lu.vstack(operators, dim)
}
\arguments{
\item{operators}{the operators to concatenate}

\item{dim}{the dimensions of the result}
}
\value{
LinOp representing the vertically stacked operators
}
\description{
Concatenate operators vertically
}
