% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304_canonInterface.R
\name{make_linC_from_linR}
\alias{make_linC_from_linR}
\title{Construct a C++ LinOp corresponding to LinPy.}
\usage{
make_linC_from_linR(linR, linR_to_linC)
}
\arguments{
\item{linR}{A LinPy object.}

\item{linR_to_linC}{A mapping from LinPy objects to LinC objects.}
}
\description{
Children of linR are retrieved from linR_to_linC.
}
