% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{matrix_prop-methods}
\alias{matrix_prop-methods}
\alias{is_psd}
\alias{is_nsd}
\alias{is_hermitian}
\alias{is_symmetric}
\title{Matrix Properties}
\usage{
is_psd(object)

is_nsd(object)

is_hermitian(object)

is_symmetric(object)
}
\arguments{
\item{object}{An \linkS4class{Expression} object.}
}
\value{
A logical value.
}
\description{
Determine if an expression is positive semidefinite, negative semidefinite, hermitian, and/or symmetric.
}
