% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\docType{class}
\name{MatrixStuffing-class}
\alias{MatrixStuffing-class}
\alias{MatrixStuffing}
\alias{perform,MatrixStuffing,Problem-method}
\alias{invert,MatrixStuffing,Solution,InverseData-method}
\title{The MatrixStuffing class.}
\usage{
\S4method{perform}{MatrixStuffing,Problem}(object, problem)

\S4method{invert}{MatrixStuffing,Solution,InverseData}(object, solution, inverse_data)
}
\arguments{
\item{object}{A \linkS4class{MatrixStuffing} object.}

\item{problem}{A \linkS4class{Problem} object to stuff; the arguments of every constraint must be affine.}

\item{solution}{A \linkS4class{Solution} to a problem that generated the inverse data.}

\item{inverse_data}{The data encoding the original problem.}
}
\description{
This class stuffs a problem into a standard form for a family of solvers.
}
\section{Methods (by generic)}{
\itemize{
\item \code{perform(object = MatrixStuffing, problem = Problem)}: Returns a stuffed problem. The returned problem is a minimization problem in which every
constraint in the problem has affine arguments that are expressed in the form A %*% x + b.

\item \code{invert(
  object = MatrixStuffing,
  solution = Solution,
  inverse_data = InverseData
)}: Returns the solution to the original problem given the inverse_data.

}}
