% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{mean.Expression}
\alias{mean.Expression}
\alias{mean}
\title{Arithmetic Mean}
\usage{
\method{mean}{Expression}(x, trim = 0, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An \linkS4class{Expression} object.}

\item{trim}{(Unimplemented) The fraction (0 to 0.5) of observations to be trimmed from each end of \eqn{x} before the mean is computed.}

\item{na.rm}{(Unimplemented) A logical value indicating whether missing values should be removed.}

\item{...}{(Unimplemented) Optional arguments.}
}
\value{
An \linkS4class{Expression} representing the mean of the input.
}
\description{
The arithmetic mean of an expression.
}
\examples{
A <- Variable(2,2)
val <- cbind(c(-5,2), c(-3,1))
prob <- Problem(Minimize(mean(A)), list(A == val))
result <- solve(prob)
result$value
}
