% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\docType{class}
\name{Minimize-class}
\alias{Minimize-class}
\alias{.Minimize}
\alias{Minimize}
\alias{canonicalize,Minimize-method}
\alias{is_dcp,Minimize-method}
\alias{is_dgp,Minimize-method}
\title{The Minimize class.}
\usage{
Minimize(expr)

\S4method{canonicalize}{Minimize}(object)

\S4method{is_dcp}{Minimize}(object)

\S4method{is_dgp}{Minimize}(object)
}
\arguments{
\item{expr}{A scalar \linkS4class{Expression} to minimize.}

\item{object}{A \linkS4class{Minimize} object.}
}
\description{
This class represents an optimization objective for minimization.
}
\section{Methods (by generic)}{
\itemize{
\item \code{canonicalize}: Pass on the target expression's objective and constraints.

\item \code{is_dcp}: A logical value indicating whether the objective is convex.

\item \code{is_dgp}: A logical value indicating whether the objective is log-log convex.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{A scalar \linkS4class{Expression} to minimize.}
}}

