% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{mip_capable}
\alias{mip_capable}
\title{Solver Capabilities}
\usage{
mip_capable(solver)
}
\arguments{
\item{solver}{A \linkS4class{ReductionSolver} object.}
}
\value{
A logical value.
}
\description{
Determine if a solver is capable of solving a mixed-integer program (MIP).
}
\examples{
mip_capable(ECOS())
}
