% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/091_mixed_norm.R
\name{MixedNorm}
\alias{MixedNorm}
\title{The MixedNorm atom.}
\usage{
MixedNorm(X, p = 2, q = 1)
}
\arguments{
\item{X}{The matrix to take the \eqn{l_{p,q}} norm of}

\item{p}{The type of inner norm}

\item{q}{The type of outer norm}
}
\value{
Returns the mixed norm of X with specified parameters p and q
}
\description{
The \eqn{l_{p,q}} norm of X, \eqn{(\sum_k (\sum_l ||X_{k,l}||^p)^{q/p})^{1/q}}.
}
