% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosek-solver.R
\docType{methods}
\name{status_map,MOSEK-method}
\alias{status_map,MOSEK-method}
\title{MOSEK Status Map}
\usage{
\S4method{status_map}{MOSEK}(solver, status)
}
\arguments{
\item{solver}{A \linkS4class{MOSEK} object.}

\item{status}{An exit code returned by MOSEK. See the \href{http://docs.mosek.com/8.0/dotnetfusion/solution_status.html}{MOSEK documentation} for details.}
}
\value{
A string indicating the status, either "optimal", "infeasible", "unbounded", "optimal_inaccurate", "infeasible_inaccurate", "unbounded_inaccurate", or "solver_error".
}
\description{
Map of MOSEK status to CVXR status.
}
