% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/294_mosek_conif.R
\name{MOSEK.parse_dual_vars}
\alias{MOSEK.parse_dual_vars}
\title{Parses MOSEK dual variables into corresponding CVXR constraints and dual values}
\usage{
MOSEK.parse_dual_vars(dual_var, constr_id_to_constr_dim)
}
\arguments{
\item{dual_var}{List of the dual variables returned by the MOSEK solution.}

\item{constr_id_to_constr_dim}{A list that contains the mapping of entry "id"
that is the index of the CVXR \linkS4class{Constraint} object to which the
next "dim" entries of the dual variable belong.}
}
\value{
A list with the mapping of the CVXR \linkS4class{Constraint} object
indices with the corresponding dual values.
}
\description{
Parses MOSEK dual variables into corresponding CVXR constraints and dual values
}
