% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{Multiply-class}
\alias{Multiply-class}
\alias{.Multiply}
\alias{Multiply}
\alias{to_numeric,Multiply-method}
\alias{dim_from_args,Multiply-method}
\alias{is_atom_log_log_convex,Multiply-method}
\alias{is_atom_log_log_concave,Multiply-method}
\alias{is_psd,Multiply-method}
\alias{is_nsd,Multiply-method}
\alias{graph_implementation,Multiply-method}
\title{The Multiply class.}
\usage{
Multiply(lh_exp, rh_exp)

\S4method{to_numeric}{Multiply}(object, values)

\S4method{dim_from_args}{Multiply}(object)

\S4method{is_atom_log_log_convex}{Multiply}(object)

\S4method{is_atom_log_log_concave}{Multiply}(object)

\S4method{is_psd}{Multiply}(object)

\S4method{is_nsd}{Multiply}(object)

\S4method{graph_implementation}{Multiply}(object, arg_objs, dim, data = NA_real_)
}
\arguments{
\item{lh_exp}{An \linkS4class{Expression} or R numeric data.}

\item{rh_exp}{An \linkS4class{Expression} or R numeric data.}

\item{object}{A \linkS4class{Multiply} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the elementwise product of two expressions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: Multiplies the values elementwise.

\item \code{dim_from_args}: The sum of the argument dimensions - 1.

\item \code{is_atom_log_log_convex}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave}: Is the atom log-log concave?

\item \code{is_psd}: Is the expression a positive semidefinite matrix?

\item \code{is_nsd}: Is the expression a negative semidefinite matrix?

\item \code{graph_implementation}: The graph implementation of the expression.
}}

