% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\name{Neg}
\alias{Neg}
\title{An alias for -MinElemwise(x, 0)}
\usage{
Neg(x)
}
\arguments{
\item{x}{An R numeric value or \linkS4class{Expression}.}
}
\value{
An alias for -MinElemwise(x, 0)
}
\description{
An alias for -MinElemwise(x, 0)
}
