% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/affine.R
\docType{class}
\name{-,Expression,missing-method}
\alias{-,Expression,missing-method}
\alias{-,Expression,Expression-method}
\alias{-,Expression,ConstVal-method}
\alias{-,ConstVal,Expression-method}
\alias{NegExpression-class}
\alias{.NegExpression}
\alias{NegExpression}
\alias{dim_from_args,NegExpression-method}
\alias{sign_from_args,NegExpression-method}
\alias{is_incr,NegExpression-method}
\alias{is_decr,NegExpression-method}
\alias{is_symmetric,NegExpression-method}
\alias{is_hermitian,NegExpression-method}
\alias{graph_implementation,NegExpression-method}
\title{The NegExpression class.}
\usage{
\S4method{-}{Expression,missing}(e1, e2)

\S4method{-}{Expression,Expression}(e1, e2)

\S4method{-}{Expression,ConstVal}(e1, e2)

\S4method{-}{ConstVal,Expression}(e1, e2)

\S4method{dim_from_args}{NegExpression}(object)

\S4method{sign_from_args}{NegExpression}(object)

\S4method{is_incr}{NegExpression}(object, idx)

\S4method{is_decr}{NegExpression}(object, idx)

\S4method{is_symmetric}{NegExpression}(object)

\S4method{is_hermitian}{NegExpression}(object)

\S4method{graph_implementation}{NegExpression}(object, arg_objs, dim, data = NA_real_)
}
\arguments{
\item{e1, e2}{The \linkS4class{Expression} objects or numeric constants to subtract.}

\item{object}{A \linkS4class{NegExpression} object.}

\item{idx}{An index into the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the negation of an affine expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{dim_from_args}: The (row, col) dimensions of the expression.

\item \code{sign_from_args}: The (is positive, is negative) sign of the expression.

\item \code{is_incr}: The expression is not weakly increasing in any argument.

\item \code{is_decr}: The expression is weakly decreasing in every argument.

\item \code{is_symmetric}: Is the expression symmetric?

\item \code{is_hermitian}: Is the expression Hermitian?

\item \code{graph_implementation}: The graph implementation of the expression.
}}

