% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{NonNegConstraint-class}
\alias{NonNegConstraint-class}
\alias{.NonNegConstraint}
\alias{name,NonNegConstraint-method}
\alias{is_dcp,NonNegConstraint-method}
\alias{is_dgp,NonNegConstraint-method}
\alias{is_dqcp,NonNegConstraint-method}
\alias{residual,NonNegConstraint-method}
\alias{violation,NonNegConstraint-method}
\title{The NonNegConstraint class}
\usage{
\S4method{name}{NonNegConstraint}(x)

\S4method{is_dcp}{NonNegConstraint}(object, dpp = FALSE)

\S4method{is_dgp}{NonNegConstraint}(object, dpp = FALSE)

\S4method{is_dqcp}{NonNegConstraint}(object)

\S4method{residual}{NonNegConstraint}(object)

\S4method{violation}{NonNegConstraint}(object)
}
\arguments{
\item{x, object}{A \linkS4class{NonNegConstraint} object.}
}
\description{
A constraint of the form \eqn{x \geq 0}.
}
\details{
This class was created to account for the fact that the ConicSolver interface
returns matrix data stated with respect to the nonnegative orthant, rather than
the nonpositive orthant. This class can be removed if the behavior of ConicSolver is
changed. However the current behavior of ConicSolver means CVXR's dual variable
and Lagrangian convention follows the most common convention in the literature.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name(NonNegConstraint)}: The string representation of the constraint.

\item \code{is_dcp(NonNegConstraint)}: A non-positive constraint is DCP if its argument is convex.

\item \code{is_dgp(NonNegConstraint)}: Is the constraint DGP?

\item \code{is_dqcp(NonNegConstraint)}: Is the constraint DQCP?

\item \code{residual(NonNegConstraint)}: The residual of the constraint.

\item \code{violation(NonNegConstraint)}: The violation of the constraint.

}}
