% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/127_constraints_nonpos.R
\name{NonPos}
\alias{NonPos}
\title{Create a Non-Positive Constraint}
\usage{
NonPos(expr, constr_id = NULL)
}
\arguments{
\item{expr}{A CVXR expression.}

\item{constr_id}{Optional integer constraint ID.}
}
\value{
A \code{NonPos} constraint object.
}
\description{
Constrains an expression to be non-positive elementwise: \eqn{x \le 0}.
}
\seealso{
\code{\link{NonNeg}}, \code{\link{Inequality}}
}
