% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/092_norm.R
\name{Norm}
\alias{Norm}
\title{The Norm atom.}
\usage{
Norm(x, p = 2, axis = NA_real_, keepdims = FALSE)
}
\arguments{
\item{x}{The matrix to take the norm of}

\item{p}{The type of norm. Valid options include any positive integer, 'fro' (for frobenius),
'nuc' (sum of singular values), np.inf or 'inf' (infinity norm).}

\item{axis}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}

\item{keepdims}{(Optional) Should dimensions be maintained when applying the atom along an axis? If \code{FALSE}, result will be collapsed into an \eqn{n x 1} column vector. The default is \code{FALSE}.}
}
\value{
Returns the specified norm of x.
}
\description{
Wrapper around the different norm atoms.
}
