% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{Norm1-class}
\alias{Norm1-class}
\alias{.Norm1}
\alias{Norm1}
\alias{name,Norm1-method}
\alias{allow_complex,Norm1-method}
\alias{.domain,Norm1-method}
\alias{.grad,Norm1-method}
\alias{.column_grad,Norm1-method}
\title{The Norm1 class.}
\usage{
\S4method{name}{Norm1}(x)

\S4method{allow_complex}{Norm1}(object)

\S4method{.domain}{Norm1}(object)

\S4method{.grad}{Norm1}(object, values)

\S4method{.column_grad}{Norm1}(object, value)
}
\arguments{
\item{x, object}{A \linkS4class{Norm1} object.}

\item{values}{A list of numeric values for the arguments}
}
\description{
This class represents the 1-norm of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: The name and arguments of the atom.

\item \code{allow_complex}: Does the atom handle complex numbers?

\item \code{.domain}: Returns constraints describing the domain of the node

\item \code{.grad}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.column_grad}: Gives the (sub/super)gradient of the atom w.r.t. each column variable
}}

\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} object.}
}}

