% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_zzz_R_specific_exports.R
\name{norm2}
\alias{norm2}
\title{Euclidean norm (deprecated alias)}
\usage{
norm2(x)
}
\arguments{
\item{x}{An Expression.}
}
\value{
An Expression representing the L2 norm.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{p_norm(x, 2)} instead.
}
\seealso{
\code{\link[=p_norm]{p_norm()}}
}
