% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/086_atoms_elementwise_logic.R
\name{Not}
\alias{Not}
\title{Logical NOT}
\usage{
Not(x, id = NULL)
}
\arguments{
\item{x}{A boolean \link{Variable} or logic expression.}

\item{id}{Optional integer ID (internal use).}
}
\value{
A \code{Not} expression.
}
\description{
Returns \code{1 - x}, flipping 0 to 1 and 1 to 0.
Can also be written with the \code{!} operator: \code{!x}.
}
\examples{
\dontrun{
x <- Variable(boolean = TRUE)
not_x <- !x          # operator syntax
not_x <- Not(x)      # functional syntax
}
}
\seealso{
\code{\link[=And]{And()}}, \code{\link[=Or]{Or()}}, \code{\link[=Xor]{Xor()}}, \code{\link[=implies]{implies()}}, \code{\link[=iff]{iff()}}
}
