% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{num_cones}
\alias{num_cones}
\title{Get the Number of Cones in a Constraint}
\usage{
num_cones(x, ...)
}
\arguments{
\item{x}{A cone constraint object.}

\item{...}{Not used.}
}
\value{
Integer.
}
\description{
Get the Number of Cones in a Constraint
}
\keyword{internal}
