% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{Objective-arith}
\alias{Objective-arith}
\alias{-,Minimize,missing-method}
\alias{+,Minimize,Minimize-method}
\alias{+,Minimize,Maximize-method}
\alias{+,Minimize,numeric-method}
\alias{+,numeric,Minimize-method}
\alias{-,Minimize,Minimize-method}
\alias{-,Minimize,Maximize-method}
\alias{-,Minimize,numeric-method}
\alias{-,numeric,Minimize-method}
\alias{*,Minimize,numeric-method}
\alias{*,Maximize,numeric-method}
\alias{*,numeric,Minimize-method}
\alias{/,Minimize,numeric-method}
\alias{-,Maximize,missing-method}
\alias{+,Maximize,Maximize-method}
\alias{+,Maximize,Minimize-method}
\title{Arithmetic Operations on Objectives}
\usage{
\S4method{-}{Minimize,missing}(e1, e2)

\S4method{+}{Minimize,Minimize}(e1, e2)

\S4method{+}{Minimize,Maximize}(e1, e2)

\S4method{+}{Minimize,numeric}(e1, e2)

\S4method{+}{numeric,Minimize}(e1, e2)

\S4method{-}{Minimize,Minimize}(e1, e2)

\S4method{-}{Minimize,Maximize}(e1, e2)

\S4method{-}{Minimize,numeric}(e1, e2)

\S4method{-}{numeric,Minimize}(e1, e2)

\S4method{*}{Minimize,numeric}(e1, e2)

\S4method{*}{Maximize,numeric}(e1, e2)

\S4method{*}{numeric,Minimize}(e1, e2)

\S4method{/}{Minimize,numeric}(e1, e2)

\S4method{-}{Maximize,missing}(e1, e2)

\S4method{+}{Maximize,Maximize}(e1, e2)

\S4method{+}{Maximize,Minimize}(e1, e2)
}
\arguments{
\item{e1}{The left-hand \linkS4class{Minimize}, \linkS4class{Maximize}, or numeric value.}

\item{e2}{The right-hand \linkS4class{Minimize}, \linkS4class{Maximize}, or numeric value.}
}
\value{
A \linkS4class{Minimize} or \linkS4class{Maximize} object.
}
\description{
Add, subtract, multiply, or divide optimization objectives.
}
