% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\docType{class}
\name{Objective-class}
\alias{Objective-class}
\alias{.Objective}
\alias{Objective}
\alias{value,Objective-method}
\alias{is_quadratic,Objective-method}
\alias{is_qpwa,Objective-method}
\title{The Objective class.}
\usage{
Objective(expr)

\S4method{value}{Objective}(object)

\S4method{is_quadratic}{Objective}(object)

\S4method{is_qpwa}{Objective}(object)
}
\arguments{
\item{expr}{A scalar \linkS4class{Expression} to optimize.}

\item{object}{An \linkS4class{Objective} object.}
}
\description{
This class represents an optimization objective.
}
\section{Methods (by generic)}{
\itemize{
\item \code{value(Objective)}: The value of the objective expression.

\item \code{is_quadratic(Objective)}: Is the objective a quadratic function?

\item \code{is_qpwa(Objective)}: Is the objective a quadratic of piecewise affine function?

}}
\section{Slots}{

\describe{
\item{\code{expr}}{A scalar \linkS4class{Expression} to optimize.}
}}

