% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/015_parameter.R
\docType{class}
\name{Parameter-class}
\alias{Parameter-class}
\alias{Parameter}
\alias{get_data,Parameter-method}
\alias{name,Parameter-method}
\alias{is_constant,Parameter-method}
\alias{value,Parameter-method}
\alias{value<-,Parameter-method}
\alias{grad,Parameter-method}
\alias{parameters,Parameter-method}
\alias{canonicalize,Parameter-method}
\title{The Parameter class.}
\usage{
\S4method{get_data}{Parameter}(object)

\S4method{name}{Parameter}(x)

\S4method{is_constant}{Parameter}(object)

\S4method{value}{Parameter}(object)

\S4method{value}{Parameter}(object) <- value

\S4method{grad}{Parameter}(object)

\S4method{parameters}{Parameter}(object)

\S4method{canonicalize}{Parameter}(object)
}
\arguments{
\item{object, x}{A \linkS4class{Parameter} object.}
}
\description{
This class represents a parameter in an optimization problem.
}
\details{
Parameters are constant expressions whose value may be specified
after problem creation. The only way to modify a problem after its
creation is through parameters. For example, you might choose to declare
the hyper-parameters of a machine learning model to be Parameter objects;
more generally, Parameters are useful for computing trade-off curves.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_data(Parameter)}: Returns information needed to reconstruct the expression besides args: \code{list(dim, name, value, id, attributes)}.

\item \code{name(Parameter)}: The name of the parameter.

\item \code{is_constant(Parameter)}: Is the expression constant?

\item \code{value(Parameter)}: The value of the parameter.

\item \code{value(Parameter) <- value}: Set the value of the parameter.

\item \code{grad(Parameter)}: An empty list since the gradient of a parameter is zero.

\item \code{parameters(Parameter)}: Returns itself as a parameter.

\item \code{canonicalize(Parameter)}: The canonical form of the parameter.

}}
\section{Slots}{

\describe{
\item{\code{dim}}{the dimensions of the parameter.}

\item{\code{name}}{the name of this parameter}

\item{\code{PARAM_COUNT}}{count of parameters, (currently unused?)}

\item{\code{venv}}{an environment used for reference semantics in setting values for parameter}

\item{\code{delta}}{the delta parameter}

\item{\code{gradient}}{the gradient parameter}

\item{\code{is_constant}}{flag indicating the parameter is constant}

\item{\code{is_vector}}{flag indicating the parameter is a vector}
}}

\examples{
x <- Parameter(3, name = "x0", nonpos = TRUE) ## 3-vec negative
is_nonneg(x)
is_nonpos(x)
size(x)
}
