% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{parameters}
\alias{parameters}
\title{Get the Parameters in an Expression}
\usage{
parameters(x, ...)
}
\arguments{
\item{x}{An expression or problem object.}

\item{...}{Not used.}
}
\value{
List of \code{\link{Parameter}} objects.
}
\description{
Get the Parameters in an Expression
}
