% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/252_transforms_partial_optimize.R
\name{PartialProblem}
\alias{PartialProblem}
\title{Partial optimization of a Problem}
\usage{
PartialProblem(
  prob,
  opt_vars,
  dont_opt_vars,
  solver = NULL,
  solve_kwargs = list(),
  id = NULL
)
}
\description{
A \code{PartialProblem} is an Expression that represents the optimal
value of an inner Problem as a function of the variables you choose
NOT to optimise over.  Build one with \code{\link[=partial_optimize]{partial_optimize()}} rather
than constructing the class directly.
}
\keyword{internal}
