% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/038_partial_trace.R
\name{PartialTrace}
\alias{PartialTrace}
\title{The PartialTrace atom.}
\usage{
PartialTrace(expr, dims, axis)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a 2D expression of which to take the partial trace.}

\item{dims}{A vector of integers encoding the dimensions of each subsystem.}

\item{axis}{The index of the subsystem to be traced out from the tensor product that defines \code{expr}.}
}
\value{
The partial trace of \code{expr}.
}
\description{
Assume \eqn{expr = X_1 \otimes \cdots \otimes X_n} is a 2D Kronecker product
composed of \eqn{n = length(dims)} implicit subsystems.
Letting \eqn{k = axis}, this class represents the partial trace of \eqn{expr}
along its \eqn{k}-th implicit subsystem:
}
\details{
\deqn{tr(X_k) (X_1 \otimes \cdots \otimes X_{k-1} \otimes X_{k+1} \otimes \cdots \otimes X_n)}
}
