% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_qp_solver.R
\name{perform,QpSolver,Problem-method}
\alias{perform,QpSolver,Problem-method}
\title{Constructs a QP problem data stored in a list}
\usage{
\S4method{perform}{QpSolver,Problem}(object, problem)
}
\description{
Constructs a QP problem data stored in a list
}
