% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{Perspective-class}
\alias{Perspective-class}
\alias{.Perspective}
\alias{Perspective}
\alias{validate_args,Perspective-method}
\alias{to_numeric,Perspective-method}
\alias{sign_from_args,Perspective-method}
\alias{is_atom_convex,Perspective-method}
\alias{is_atom_concave,Perspective-method}
\alias{is_incr,Perspective-method}
\alias{is_decr,Perspective-method}
\alias{dim_from_args,Perspective-method}
\title{The Perspective class.}
\usage{
Perspective(f, s)

\S4method{validate_args}{Perspective}(object)

\S4method{to_numeric}{Perspective}(object, values)

\S4method{sign_from_args}{Perspective}(object)

\S4method{is_atom_convex}{Perspective}(object)

\S4method{is_atom_concave}{Perspective}(object)

\S4method{is_incr}{Perspective}(object, idx)

\S4method{is_decr}{Perspective}(object, idx)

\S4method{dim_from_args}{Perspective}(object)
}
\arguments{
\item{f}{An \linkS4class{Expression}.}

\item{s}{A \linkS4class{Variable}.}

\item{values}{A list of arguments to the atom.}

\item{idx}{An index into the atom.}
}
\description{
This class represents the perspective transform of a convex or concave scalar expression.
It uses the fact that, given a cone form for the epigraph of \eqn{f} via
}
\details{
\deqn{\{(t, x) \in \mathbb{R}^{n+1} : t \geq f(x) \} = \{(t, x) : Fx + gt + e \in K\}},

the epigraph of the perspective transform of \eqn{f} can be given by

\deqn{\{(t, x, s) \in \mathbb{R}^{n+2} : t \geq sf(x/s) \} = \{ (t, x, s) : Fx + gt + se \in K \}}.

(See https://web.stanford.edu/~boyd/papers/pdf/sw_aff_ctrl.pdf).

Note that this is the perspective transform of a scalar expression viewed as
a function of its underlying variables. The perspective atom does not return
a `Callable`, so you cannot create compositions such as \eqn{p(g(x),s)}, where
\eqn{p(z,s) = sf(z/s)} is the perspective transform of \eqn{f}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args(Perspective)}: Checks the dimensions of the arguments.

\item \code{to_numeric(Perspective)}: Returns the perspective sf(x/s).

\item \code{sign_from_args(Perspective)}: Returns the sign (is positive, is negative) of the atom.

\item \code{is_atom_convex(Perspective)}: Is the atom convex?

\item \code{is_atom_concave(Perspective)}: Is the atom concave?

\item \code{is_incr(Perspective)}: Is the atom weakly increasing in the argument \code{idx}?

\item \code{is_decr(Perspective)}: Is the atom weakly decreasing in the argument \code{idx}?

\item \code{dim_from_args(Perspective)}: The dimensions of the atom.

}}
\section{Slots}{

\describe{
\item{\code{f}}{An \linkS4class{Expression}.}

\item{\code{s}}{A \linkS4class{Variable}.}
}}

