% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_atoms_pf_eigenvalue.R
\name{pf_eigenvalue}
\alias{pf_eigenvalue}
\title{Perron-Frobenius eigenvalue of a positive matrix}
\usage{
pf_eigenvalue(X)
}
\arguments{
\item{X}{An Expression (positive square matrix)}
}
\value{
A PfEigenvalue atom (scalar)
}
\description{
Log-log convex atom for DGP. Solve with
\code{psolve(problem, gp = TRUE)}.
}
\examples{
X <- Variable(c(2, 2), pos = TRUE)
prob <- Problem(Minimize(pf_eigenvalue(X)),
                list(X[1,1] >= 0.1, X[2,2] >= 0.1))
\dontrun{psolve(prob, gp = TRUE)}
}
