% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{Problem-arith}
\alias{Problem-arith}
\alias{+,Problem,missing-method}
\alias{-,Problem,missing-method}
\alias{+,Problem,numeric-method}
\alias{+,numeric,Problem-method}
\alias{+,Problem,Problem-method}
\alias{-,Problem,numeric-method}
\alias{-,numeric,Problem-method}
\alias{-,Problem,Problem-method}
\alias{*,Problem,numeric-method}
\alias{*,numeric,Problem-method}
\alias{/,Problem,numeric-method}
\title{Arithmetic Operations on Problems}
\usage{
\S4method{+}{Problem,missing}(e1, e2)

\S4method{-}{Problem,missing}(e1, e2)

\S4method{+}{Problem,numeric}(e1, e2)

\S4method{+}{numeric,Problem}(e1, e2)

\S4method{+}{Problem,Problem}(e1, e2)

\S4method{-}{Problem,numeric}(e1, e2)

\S4method{-}{numeric,Problem}(e1, e2)

\S4method{-}{Problem,Problem}(e1, e2)

\S4method{*}{Problem,numeric}(e1, e2)

\S4method{*}{numeric,Problem}(e1, e2)

\S4method{/}{Problem,numeric}(e1, e2)
}
\arguments{
\item{e1}{The left-hand \linkS4class{Problem} object.}

\item{e2}{The right-hand \linkS4class{Problem} object.}
}
\value{
A \linkS4class{Problem} object.
}
\description{
Add, subtract, multiply, or divide DCP optimization problems.
}
