% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{project_and_assign}
\alias{project_and_assign}
\title{Projects and assigns a value onto the attribute set of a \linkS4class{Leaf}.}
\usage{
project_and_assign(object, value)
}
\description{
Projects and assigns a value onto the attribute set of a \linkS4class{Leaf}.
}
