% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/295_zzz_R_extra.R
\name{psd_coeff_offset}
\alias{psd_coeff_offset}
\title{Given a problem returns a PSD constrain}
\usage{
psd_coeff_offset(problem, c)
}
\arguments{
\item{problem}{A \linkS4class{Problem} object.}

\item{c}{A vector of coefficients.}
}
\value{
Returns an array G and vector h such that the given constraint is
equivalent to G*z <=_{PSD} h.
}
\description{
Given a problem returns a PSD constrain
}
