% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solver_mosek.R
\name{psd_format_mat,MOSEK-method}
\alias{psd_format_mat,MOSEK-method}
\title{Construct a linear operator to multiply by PSD constraint coefficients.}
\usage{
\S4method{psd_format_mat}{MOSEK}(solver, constr)
}
\arguments{
\item{constr}{a list of constraints}
}
\value{
a linear operator to multiply by PSD constraint coefficients.
}
\description{
Construct a linear operator to multiply by PSD constraint coefficients.
}
\details{
Special cases PSD constraints, as MOSEK expects constraints to be
imposed on solely the lower triangular part of the variable matrix.
This function differs from the SCS version only in the fact that it does not apply the \eqn{\sqrt(2)} scaling on off-diagonal entries. This difference is necessary based on how we implement the \code{MOSEK$bar_data}.
}
