% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/287_scs_conif.R
\name{psd_format_mat,SCS-method}
\alias{psd_format_mat,SCS-method}
\title{Construct a linear operator to multiply by PSD constraint coefficients.}
\usage{
\S4method{psd_format_mat}{SCS}(solver, constr)
}
\arguments{
\item{constr}{a list of constraints}
}
\value{
a linear operator to multiply by PSD constraint coefficients.
}
\description{
Construct a linear operator to multiply by PSD constraint coefficients.
}
\details{
Special cases PSD constraints, as SCS expects constraints to be
imposed on solely the lower triangular part of the variable matrix.
Moreover, it requires the off-diagonal coefficients to be scaled by
sqrt(2), and applies to the symmetric part of the constrained expression.
}
