% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{PSDWrap-class}
\alias{PSDWrap-class}
\alias{.PSDWrap}
\alias{PSDWrap}
\alias{is_psd,PSDWrap-method}
\title{The PSDWrap class.}
\usage{
PSDWrap(arg)

\S4method{is_psd}{PSDWrap}(object)
}
\arguments{
\item{arg}{A \linkS4class{Expression} object or matrix.}
}
\description{
A no-op wrapper to assert the input argument is positive semidefinite.
}
\section{Methods (by generic)}{
\itemize{
\item \code{is_psd}: Is the atom positive semidefinite?
}}

