% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/268_qp2symbolic_qp.R
\docType{class}
\name{Qp2SymbolicQp-class}
\alias{Qp2SymbolicQp-class}
\alias{.Qp2SymbolicQp}
\title{The Qp2SymbolicQp class.}
\description{
This class reduces a quadratic problem to a problem that consists of affine
expressions and symbolic quadratic forms.
}
