% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/269_qp_matrix_stuffing.R
\docType{class}
\name{QpMatrixStuffingConeDims-class}
\alias{QpMatrixStuffingConeDims-class}
\alias{.QpMatrixStuffingConeDims}
\title{The QpMatrixStuffingConeDims class.}
\description{
This class contains a summary of cone dimensions present in constraints.
}
\section{Slots}{

\describe{
\item{\code{zero}}{The dimension of the zero cone.}

\item{\code{nonpos}}{The dimension of the nonpositive cone.}

\item{\code{exp}}{The number of 3-dimensional exponential cones.}

\item{\code{soc}}{A vector of the second-order cone dimensions.}

\item{\code{psd}}{A vector the positive semidefinite cone dimensions, where the dimension of the PSD cone of k by k matrices is k.}
}}

