% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_utilities_scopes.R
\name{quad_form_dpp_scope_active}
\alias{quad_form_dpp_scope_active}
\title{Check if a quad_form DPP Scope is Active}
\usage{
quad_form_dpp_scope_active()
}
\value{
Logical; \code{TRUE} if a \code{with_quad_form_dpp_scope} block is active.
}
\description{
Check if a quad_form DPP Scope is Active
}
\keyword{internal}
