% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{Rdict-class}
\alias{Rdict-class}
\alias{Rdict}
\alias{Rdict}
\alias{$,Rdict-method}
\alias{length,Rdict-method}
\alias{is.element,ANY,Rdict-method}
\alias{[,Rdict,ANY,ANY,ANY-method}
\alias{[<-,Rdict,ANY,ANY,ANY-method}
\title{The Rdict class.}
\usage{
Rdict(keys = list(), values = list())

\S4method{$}{Rdict}(x, name)

\S4method{length}{Rdict}(x)

\S4method{is.element}{ANY,Rdict}(el, set)

\S4method{[}{Rdict,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{Rdict,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{keys}{A list of keys.}

\item{values}{A list of values corresponding to the keys.}

\item{x, set}{A \linkS4class{Rdict} object.}

\item{name}{Either "keys" for a list of keys, "values" for a list of values, or "items" for a list of lists where each nested list is a (key, value) pair.}

\item{el}{The element to search the dictionary of values for.}

\item{i}{A key into the dictionary.}

\item{j, drop, ...}{Unused arguments.}

\item{value}{The value to assign to key \code{i}.}
}
\description{
A simple, internal dictionary composed of a list of keys and a list of values. These keys/values can be any type, including nested lists, S4 objects, etc.
Incredibly inefficient hack, but necessary for the geometric mean atom, since it requires mixed numeric/gmp objects.
}
\section{Slots}{

\describe{
\item{\code{keys}}{A list of keys.}

\item{\code{values}}{A list of values corresponding to the keys.}
}}

