% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{Rdictdefault-class}
\alias{Rdictdefault-class}
\alias{Rdictdefault}
\alias{[,Rdictdefault,ANY,ANY,ANY-method}
\title{The Rdictdefault class.}
\usage{
Rdictdefault(keys = list(), values = list(), default)

\S4method{[}{Rdictdefault,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{keys}{A list of keys.}

\item{values}{A list of values corresponding to the keys.}

\item{default}{A function that takes as input a key and outputs a value to assign to that key.}

\item{x}{A \linkS4class{Rdictdefault} object.}

\item{i}{A key into the dictionary.}

\item{j, drop, ...}{Unused arguments.}
}
\description{
This is a subclass of \linkS4class{Rdict} that contains an additional slot for a default function, which assigns a value to an input key.
Only partially implemented, but working well enough for the geometric mean. Will be combined with \linkS4class{Rdict} later.
}
\section{Slots}{

\describe{
\item{\code{keys}}{A list of keys.}

\item{\code{values}}{A list of values corresponding to the keys.}

\item{\code{default}}{A function that takes as input a key and outputs a value to assign to that key.}
}}

\seealso{
\linkS4class{Rdict}
}
