% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{Real-class}
\alias{Real-class}
\alias{.Real}
\alias{Real}
\alias{to_numeric,Real-method}
\alias{dim_from_args,Real-method}
\alias{is_imag,Real-method}
\alias{is_complex,Real-method}
\alias{is_symmetric,Real-method}
\title{The Real class.}
\usage{
Real(expr)

\S4method{to_numeric}{Real}(object, values)

\S4method{dim_from_args}{Real}(object)

\S4method{is_imag}{Real}(object)

\S4method{is_complex}{Real}(object)

\S4method{is_symmetric}{Real}(object)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a vector or matrix.}

\item{object}{An \linkS4class{Real} object.}

\item{values}{A list of arguments to the atom.}
}
\description{
This class represents the real part of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: The imaginary part of the given value.

\item \code{dim_from_args}: The dimensions of the atom.

\item \code{is_imag}: Is the atom imaginary?

\item \code{is_complex}: Is the atom complex valued?

\item \code{is_symmetric}: Is the atom symmetric?
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a vector or matrix.}
}}

