% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/272_zzz_R_extra.R, R/281_zzz_R_extra.R
\name{reconcile_solver_options}
\alias{reconcile_solver_options}
\title{Update solver options using defaults and return list}
\usage{
reconcile_solver_options(solver_opts, defaults)

reconcile_solver_options(solver_opts, defaults)
}
\arguments{
\item{solver_opts}{the solver options named list}

\item{the}{defaults the named list to reconcile against}
}
\value{
list of solver options

list of solver options
}
\description{
Update solver options using defaults and return list

Update solver options using defaults and return list
}
