% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{reduce}
\alias{reduce}
\title{Reduce a Problem}
\usage{
reduce(object)
}
\arguments{
\item{object}{A \linkS4class{Reduction} object.}
}
\value{
An equivalent problem, encoded either as a \linkS4class{Problem} object or a list.
}
\description{
Reduces the owned problem to an equivalent problem.
}
