% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{reduction_apply}
\alias{reduction_apply}
\title{Apply a Reduction to a Problem}
\usage{
reduction_apply(x, problem, ...)
}
\arguments{
\item{x}{A Reduction object.}

\item{problem}{A Problem object.}

\item{...}{Additional arguments.}
}
\value{
List with (new_problem, inverse_data).
}
\description{
Apply a Reduction to a Problem
}
\keyword{internal}
