% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{reduction_invert}
\alias{reduction_invert}
\title{Invert a Solution through a Reduction}
\usage{
reduction_invert(x, solution, inverse_data, ...)
}
\arguments{
\item{x}{A Reduction object.}

\item{solution}{A solution object.}

\item{inverse_data}{Inverse data from apply.}

\item{...}{Additional arguments.}
}
\value{
A solution to the original problem.
}
\description{
Invert a Solution through a Reduction
}
\keyword{internal}
