% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_rel_entr.R
\docType{class}
\name{RelEntr-class}
\alias{RelEntr-class}
\alias{.RelEntr}
\alias{RelEntr}
\alias{to_numeric,RelEntr-method}
\alias{sign_from_args,RelEntr-method}
\alias{is_atom_convex,RelEntr-method}
\alias{is_atom_concave,RelEntr-method}
\alias{is_incr,RelEntr-method}
\alias{is_decr,RelEntr-method}
\alias{.grad,RelEntr-method}
\alias{.domain,RelEntr-method}
\title{The RelEntr class.}
\usage{
RelEntr(x, y)

\S4method{to_numeric}{RelEntr}(object, values)

\S4method{sign_from_args}{RelEntr}(object)

\S4method{is_atom_convex}{RelEntr}(object)

\S4method{is_atom_concave}{RelEntr}(object)

\S4method{is_incr}{RelEntr}(object, idx)

\S4method{is_decr}{RelEntr}(object, idx)

\S4method{.grad}{RelEntr}(object, values)

\S4method{.domain}{RelEntr}(object)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{y}{An \linkS4class{Expression}.}

\item{object}{A \linkS4class{RelEntr} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents elementwise \eqn{x\log(x/y)}.
}
\details{
For disambiguation between RelEntr and KLDiv, see https://github.com/cvxpy/cvxpy/issues/733
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(RelEntr)}: The numeric value of the expression.

\item \code{sign_from_args(RelEntr)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(RelEntr)}: Is the atom convex?

\item \code{is_atom_concave(RelEntr)}: Is the atom concave?

\item \code{is_incr(RelEntr)}: Is the atom weakly increasing in the index?

\item \code{is_decr(RelEntr)}: Is the atom weakly decreasing in the index?

\item \code{.grad(RelEntr)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.domain(RelEntr)}: Returns constraints describing the domain of the node

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}

\item{\code{y}}{An \linkS4class{Expression}.}
}}

