% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone2cone.R
\name{RelEntrConeQuad_canon}
\alias{RelEntrConeQuad_canon}
\title{Use linear and SOC constraints to approximately enforce
con@x * log(con@x / con@y) <= con@z.}
\usage{
RelEntrConeQuad_canon(con, args)
}
\description{
We rely on an SOC characterization of 2-by-2 PSD matrices.
Namely, a matrix
     [ a, b ]
     [ b, c ]
is PSD if and only if (a, c) >= 0 and a*c >= b^2.
That system of constraints can be expressed as
a >= quad_over_lin(b, c).
}
\details{
Note: constraint canonicalization in CVXR uses a return format
list(lead_con, con_list) where lead_con is a Constraint that might be
used in dual variable recovery and con_list consists of extra
Constraint objects as needed.
}
