% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_qp_solver.R
\name{requires_constr,QpSolver-method}
\alias{requires_constr,QpSolver-method}
\title{Can the solver solve problems that do not have constraints?}
\usage{
\S4method{requires_constr}{QpSolver}(solver)
}
\description{
Can the solver solve problems that do not have constraints?
}
