% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{residual-methods}
\alias{residual-methods}
\alias{residual}
\alias{violation}
\title{Constraint Residual}
\usage{
residual(object)

violation(object)
}
\arguments{
\item{object}{A \linkS4class{Constraint} object.}
}
\value{
A \linkS4class{Expression} representing the residual, or the value of this expression.
}
\description{
The residual expression of a constraint, i.e. the amount by which it is violated, and the value of that violation.
For instance, if our constraint is \eqn{g(x) \leq 0}, the residual is \eqn{max(g(x), 0)} applied elementwise.
}
