% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{residual}
\alias{residual}
\title{Get the Residual of a Constraint}
\usage{
residual(x, ...)
}
\arguments{
\item{x}{A constraint object.}

\item{...}{Not used.}
}
\value{
Numeric array, or \code{NULL} if expression has no value.
}
\description{
Returns the residual of a constraint, measuring how much the constraint
is violated or satisfied.
}
