% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{retrieve}
\alias{retrieve}
\title{Retrieve Solution}
\usage{
retrieve(object, solution)
}
\arguments{
\item{object}{A \linkS4class{Reduction} object.}

\item{solution}{A \linkS4class{Solution} object.}
}
\value{
A \linkS4class{Solution} to the problem emitted by \code{\link{reduce}}.
}
\description{
Retrieves a solution to the owned problem.
}
