% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{save_dual_value}
\alias{save_dual_value}
\title{Save Dual Variable Values from Solver Output}
\usage{
save_dual_value(x, val)
}
\arguments{
\item{x}{A constraint object.}

\item{val}{The dual value from the solver.}
}
\value{
Invisible constraint (side effect: sets dual variable values).
}
\description{
Save Dual Variable Values from Solver Output
}
\keyword{internal}
