% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/090_zzz_R_specific_convenience.R
\name{scalar_product}
\alias{scalar_product}
\title{Scalar product (alias for vdot)}
\usage{
scalar_product(x, y)
}
\arguments{
\item{x}{An Expression or numeric value.}

\item{y}{An Expression or numeric value.}
}
\value{
A scalar Expression representing sum(x * y).
}
\description{
Scalar product (alias for vdot)
}
