% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/028_binary_operators.R
\name{ScalarProduct}
\alias{ScalarProduct}
\title{The ScalarProduct atom.}
\usage{
ScalarProduct(x, y)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric value. The conjugate-linear argument to the inner product.}

\item{y}{An \linkS4class{Expression} or numeric value. The linear argument to the inner product.}
}
\value{
An \linkS4class{Expression} representing the standard inner product of (\code{x}, \code{y}), conjugate-linear in \code{x}.
}
\description{
Standard inner product or scalar product of (x,y)
}
