% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solver_mosek.R
\name{scaled_lower_tri}
\alias{scaled_lower_tri}
\title{Utility methods for special handling of semidefinite constraints.}
\usage{
scaled_lower_tri(matrix)
}
\arguments{
\item{matrix}{The matrix to get the lower triangular matrix for}
}
\value{
The lower triangular part of the matrix, stacked in column-major order
}
\description{
Utility methods for special handling of semidefinite constraints.
}
