% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/284_clarabel_conif.R
\name{scaled_upper_tri}
\alias{scaled_upper_tri}
\title{Utility methods for special handling of semidefinite constraints.}
\usage{
scaled_upper_tri(matrix)
}
\arguments{
\item{matrix}{The matrix to get the upper triangular matrix for}
}
\value{
The upper triangular part of the matrix, stacked in column-major order
}
\description{
Utility methods for special handling of semidefinite constraints.
}
