% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\docType{methods}
\name{status_map,SCS-method}
\alias{status_map,SCS-method}
\title{SCS Status Map}
\usage{
\S4method{status_map}{SCS}(solver, status)
}
\arguments{
\item{solver}{A \linkS4class{SCS} object.}

\item{status}{An exit code returned by SCS.}
}
\value{
A string indicating the status, either "optimal", "infeasible", "unbounded", "optimal_inaccurate", "infeasible_inaccurate", "unbounded_inaccurate", or "solver_error".
}
\description{
Map of SCS status to CVXR status.
}
