% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{SCS.dims_to_solver_dict}
\alias{SCS.dims_to_solver_dict}
\title{Utility method for formatting a ConeDims instance into a dictionary
that can be supplied to SCS.}
\usage{
SCS.dims_to_solver_dict(cone_dims)
}
\arguments{
\item{cone_dims}{A \linkS4class{ConeDims} instance.}
}
\value{
The dimensions of the cones.
}
\description{
Utility method for formatting a ConeDims instance into a dictionary
that can be supplied to SCS.
}
