% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Semidef}
\alias{Semidef}
\title{Positive Semidefinite Variable}
\usage{
Semidef(n, name = NA_character_)
}
\arguments{
\item{n}{The number of rows/columns in the matrix.}

\item{name}{(Optional) A character string representing the name of the variable.}
}
\value{
An \linkS4class{Expression} representing the positive semidefinite matrix.
}
\description{
An expression representing a positive semidefinite matrix.
}
\examples{
x <- Semidef(5) ## 5 by 5 semidefinite matrix expression
}
