% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_globals.R
\name{set_id_counter}
\alias{set_id_counter}
\title{Set ID Counter}
\usage{
set_id_counter(value = 0L)
}
\arguments{
\item{value}{The value to assign as ID.}
}
\value{
the changed value
}
\description{
Set the CVXR variable/constraint identification number counter.
}
\examples{
\dontrun{
  setIdCounter(value = 0L)
}
}
