% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonInterface.R
\name{set_linC_data}
\alias{set_linC_data}
\title{Set numerical data fields in linC}
\usage{
set_linC_data(linC, linR)
}
\arguments{
\item{linC}{An object in which the data fields will be set.}

\item{linR}{An object that contains the data to be set in linC.}
}
\value{
None
}
\description{
This function sets numerical data fields in the linC object based on the
linR object's data.
}
