% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{set_variable_of_provenance}
\alias{set_variable_of_provenance}
\title{Set a variable attributes using another}
\usage{
set_variable_of_provenance(object, variable)
}
\arguments{
\item{object}{A \linkS4class{Variable} object.}

\item{variable}{Another \linkS4class{Variable} object.}
}
\value{
A modified object with the attributes set
}
\description{
Set a variable attributes using another
}
