% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_globals.R
\name{setIdCounter}
\alias{setIdCounter}
\title{Set ID Counter}
\usage{
setIdCounter(value = 0L)
}
\arguments{
\item{value}{The value to assign as ID.}
}
\value{
the changed value of the package global \code{.CVXR_options}.
}
\description{
Set the CVXR variable/constraint identification number counter.
}
\examples{
\dontrun{
  setIdCounter(value = 0L)
}
}
