% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{shape_from_args}
\alias{shape_from_args}
\title{Infer Shape from Arguments}
\usage{
shape_from_args(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
Integer vector of shape dimensions.
}
\description{
Infer Shape from Arguments
}
\keyword{internal}
