% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/104_atoms_sigma_max.R
\name{sigma_max}
\alias{sigma_max}
\title{Maximum singular value}
\usage{
sigma_max(A)
}
\arguments{
\item{A}{A matrix expression}
}
\value{
An expression representing the maximum singular value of A
}
\description{
Maximum singular value
}
