% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\docType{methods}
\name{sign,Expression-method}
\alias{sign,Expression-method}
\title{Sign of Expression}
\usage{
\S4method{sign}{Expression}(x)
}
\arguments{
\item{x}{An \linkS4class{Expression} object.}
}
\value{
A string indicating the sign of the expression, either "ZERO", "NONNEGATIVE", "NONPOSITIVE", or "UNKNOWN".
}
\description{
The sign of an expression.
}
