% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{SignEntries-class}
\alias{SignEntries-class}
\alias{.SignEntries}
\alias{SignEntries}
\alias{to_numeric,SignEntries-method}
\alias{dim_from_args,SignEntries-method}
\alias{sign_from_args,SignEntries-method}
\alias{is_atom_convex,SignEntries-method}
\alias{is_atom_concave,SignEntries-method}
\alias{is_atom_quasiconvex,SignEntries-method}
\alias{is_atom_quasiconcave,SignEntries-method}
\alias{is_incr,SignEntries-method}
\alias{is_decr,SignEntries-method}
\alias{.grad,SignEntries-method}
\title{The SignEntries class.}
\usage{
SignEntries(x = x)

\S4method{to_numeric}{SignEntries}(object, values)

\S4method{dim_from_args}{SignEntries}(object)

\S4method{sign_from_args}{SignEntries}(object)

\S4method{is_atom_convex}{SignEntries}(object)

\S4method{is_atom_concave}{SignEntries}(object)

\S4method{is_atom_quasiconvex}{SignEntries}(object)

\S4method{is_atom_quasiconcave}{SignEntries}(object)

\S4method{is_incr}{SignEntries}(object, idx)

\S4method{is_decr}{SignEntries}(object, idx)

\S4method{.grad}{SignEntries}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{object}{A \linkS4class{SignEntries} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
The sign of an expression (-1 for x <= 0, +1 for x > 0).
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(SignEntries)}: The sign of \code{x}.

\item \code{dim_from_args(SignEntries)}: The (row, col) shape of the atom.

\item \code{sign_from_args(SignEntries)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(SignEntries)}: Is the atom convex?

\item \code{is_atom_concave(SignEntries)}: Is the atom concave?

\item \code{is_atom_quasiconvex(SignEntries)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(SignEntries)}: Is the atom quasiconcave?

\item \code{is_incr(SignEntries)}: Is the atom weakly increasing?

\item \code{is_decr(SignEntries)}: Is the atom weakly decreasing?

\item \code{.grad(SignEntries)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}
}}

