% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/299_zzz_R_specific_exports.R
\name{size}
\alias{size}
\title{Get Expression Size}
\usage{
size(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
An integer (product of shape dimensions).
}
\description{
Returns the total number of elements in the expression.
}
\seealso{
\code{\link[=is_scalar]{is_scalar()}}, \code{\link[=is_vector]{is_vector()}}, \code{\link[=is_matrix]{is_matrix()}}
}
