% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/atoms.R
\name{size_from_args}
\alias{size_from_args}
\alias{size_from_args,Atom-method}
\title{Atom Size}
\usage{
size_from_args(object)

\S4method{size_from_args}{Atom}(object)
}
\arguments{
\item{object}{A \linkS4class{Atom} object.}
}
\value{
A numeric vector \code{c(row, col)} indicating the size of the atom.
}
\description{
Determine the size of an atom based on its arguments.
}
