% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{SkewSymmetricWrap-class}
\alias{SkewSymmetricWrap-class}
\alias{.SkewSymmetricWrap}
\alias{SkewSymmetricWrap}
\alias{validate_args,SkewSymmetricWrap-method}
\alias{is_skew_symmetric,SkewSymmetricWrap-method}
\title{The SkewSymmetricWrap class.}
\usage{
SkewSymmetricWrap(arg)

\S4method{validate_args}{SkewSymmetricWrap}(object)

\S4method{is_skew_symmetric}{SkewSymmetricWrap}(object)
}
\arguments{
\item{arg}{A \linkS4class{Expression} object or matrix.}

\item{object}{A \linkS4class{SkewSymmetricWrap} object.}
}
\description{
A no-op wrapper to assert the input argument is skew symmetric.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args(SkewSymmetricWrap)}: Check the input is a real square matrix.

\item \code{is_skew_symmetric(SkewSymmetricWrap)}: Is the atom skew symmetric?

}}
