% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{Solver.choose_solver}
\alias{Solver.choose_solver}
\title{Choose a Solver}
\usage{
Solver.choose_solver(constraints)
}
\arguments{
\item{constraints}{A list of canonicalized constraints.}
}
\value{
A \linkS4class{Solver} object.
}
\description{
Determines the appropriate solver.
}
