% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/015_settings.R
\name{solver-constants}
\alias{solver-constants}
\alias{SCS_SOLVER}
\alias{OSQP_SOLVER}
\alias{CLARABEL_SOLVER}
\alias{DIFFCP_SOLVER}
\alias{HIGHS_SOLVER}
\alias{MOSEK_SOLVER}
\alias{GUROBI_SOLVER}
\alias{GLPK_SOLVER}
\alias{GLPK_MI_SOLVER}
\alias{ECOS_SOLVER}
\alias{ECOS_BB_SOLVER}
\alias{CPLEX_SOLVER}
\alias{CVXOPT_SOLVER}
\alias{PIQP_SOLVER}
\alias{SCIP_SOLVER}
\alias{XPRESS_SOLVER}
\alias{IPOPT_SOLVER}
\alias{KNITRO_SOLVER}
\alias{UNO_SOLVER}
\alias{COPT_SOLVER}
\title{Solver Name Constants}
\usage{
SCS_SOLVER

OSQP_SOLVER

CLARABEL_SOLVER

DIFFCP_SOLVER

HIGHS_SOLVER

MOSEK_SOLVER

GUROBI_SOLVER

GLPK_SOLVER

GLPK_MI_SOLVER

ECOS_SOLVER

ECOS_BB_SOLVER

CPLEX_SOLVER

CVXOPT_SOLVER

PIQP_SOLVER

SCIP_SOLVER

XPRESS_SOLVER

IPOPT_SOLVER

KNITRO_SOLVER

UNO_SOLVER

COPT_SOLVER
}
\value{
A character string.
}
\description{
Character string constants identifying the available solvers.
}
