% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{Solver.solve}
\alias{Solver.solve}
\title{Call to Solver}
\usage{
Solver.solve(
  solver,
  objective,
  constraints,
  cached_data,
  warm_start,
  verbose,
  ...
)
}
\arguments{
\item{solver}{A \linkS4class{Solver} object.}

\item{objective}{A list representing the canonicalized objective.}

\item{constraints}{A list of canonicalized constraints.}

\item{cached_data}{A list mapping solver name to cached problem data.}

\item{warm_start}{A logical value indicating whether the previous solver result should be used to warm start.}

\item{verbose}{A logical value indicating whether to print solver output.}

\item{...}{Additional arguments to the solver.}
}
\value{
A list containing the status, optimal value, primal variable, and dual variables for the equality and inequality constraints.
}
\description{
Returns the result of the call to the solver.
}
