% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\docType{class}
\name{SolverStats-class}
\alias{SolverStats-class}
\alias{.SolverStats}
\alias{SolverStats}
\title{The SolverStats class.}
\usage{
SolverStats(results_dict = list(), solver_name = NA_character_)
}
\arguments{
\item{results_dict}{A list containing the results returned by the solver.}

\item{solver_name}{The name of the solver.}
}
\value{
A list containing
\describe{
  \item{\code{solver_name}}{The name of the solver.}
  \item{\code{solve_time}}{The time (in seconds) it took for the solver to solve the problem.}
  \item{\code{setup_time}}{The time (in seconds) it took for the solver to set up the problem.}
  \item{\code{num_iters}}{The number of iterations the solver had to go through to find a solution.}
  \item{\code{extra_stats}}{Extra statistics specific to the solver; these statistics are typically returned directly from the solver, without modification by CVXR.}
}
}
\description{
This class contains the miscellaneous information that is returned by a solver after solving, but that is not captured directly by the \linkS4class{Problem} object.
}
\section{Slots}{

\describe{
\item{\code{solver_name}}{The name of the solver.}

\item{\code{solve_time}}{The time (in seconds) it took for the solver to solve the problem.}

\item{\code{setup_time}}{The time (in seconds) it took for the solver to set up the problem.}

\item{\code{num_iters}}{The number of iterations the solver had to go through to find a solution.}

\item{\code{extra_stats}}{Extra statistics specific to the solver; these statistics are typically returned directly from the solver, without modification by CVXR.}
}}

