% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/277_solving_chain.R
\name{SolvingChain.reductions_for_problem_class}
\alias{SolvingChain.reductions_for_problem_class}
\title{Builds a chain that rewrites a problem into an intermediate representation suitable for numeric reductions.}
\usage{
SolvingChain.reductions_for_problem_class(
  problem,
  candidates,
  gp = FALSE,
  solver_opts = NULL
)
}
\arguments{
\item{problem}{The problem for which to build a chain.}

\item{candidates}{List of candidate solvers divided into qp_solvers and conic_solvers.}

\item{gp}{If TRUE, the problem is parsed as a Disciplined Geometric Program instead of a Disciplined Convex Program}
}
\value{
A list of Reduction objects that can be used to convert the problem to an intermediate form.
}
\description{
Builds a chain that rewrites a problem into an intermediate representation suitable for numeric reductions.
}
