% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072_sqrt.R, R/305_exports.R
\docType{methods}
\name{sqrt,Expression-method}
\alias{sqrt,Expression-method}
\alias{sqrt}
\title{Square Root}
\usage{
\S4method{sqrt}{Expression}(x)

\S4method{sqrt}{Expression}(x)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}
}
\value{
An \linkS4class{Expression} representing the square root of the input.
A <- Variable(2,2)
val <- cbind(c(2,4), c(16,1))
prob <- Problem(Maximize(sqrt(A)\link{1,2}), list(A == val))
result <- solve(prob)
result$value

An \linkS4class{Expression} representing the square root of the input.
A <- Variable(2,2)
val <- cbind(c(2,4), c(16,1))
prob <- Problem(Maximize(sqrt(A)\link{1,2}), list(A == val))
result <- solve(prob)
result$value
}
\description{
The elementwise square root.

The elementwise square root.
}
