% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{stack_vals}
\alias{stack_vals}
\title{Stacks the values of the given variables.}
\usage{
stack_vals(variables, default, byrow = FALSE)
}
\arguments{
\item{variables}{a list of variables.}

\item{default}{value to use when variable value is NA}

\item{byrow}{logical value indicating whether to fill the matrix by rows. Defaults to FALSE.}
}
\description{
Stacks the values of the given variables.
}
