% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/107_atoms_sum_largest.R
\name{sum_largest}
\alias{sum_largest}
\title{Sum of k largest entries}
\usage{
sum_largest(x, k, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{k}{Number of largest entries to sum}

\item{axis}{NULL (all entries), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical; keep the reduced dimension as size 1}
}
\value{
A SumLargest atom
}
\description{
Sum of k largest entries
}
