% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/007_utilities_sign.R
\name{sum_signs}
\alias{sum_signs}
\title{Sign of a sum of expressions}
\usage{
sum_signs(exprs)
}
\arguments{
\item{exprs}{List of Expression objects}
}
\value{
Named logical vector c(is_nonneg, is_nonpos)
}
\description{
Determines whether the sum of a list of expressions is nonnegative,
nonpositive, or unknown.
}
\keyword{internal}
