% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{SumEntries-class}
\alias{SumEntries-class}
\alias{.SumEntries}
\alias{SumEntries}
\alias{to_numeric,SumEntries-method}
\alias{graph_implementation,SumEntries-method}
\title{The SumEntries class.}
\usage{
SumEntries(expr, axis = NA_real_)

\S4method{to_numeric}{SumEntries}(object, values)

\S4method{graph_implementation}{SumEntries}(object, arg_objs, size,
  data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a vector or matrix.}

\item{axis}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}

\item{object}{A \linkS4class{SumEntries} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the sum of all entries in a vector or matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: Sum the entries along the specified axis.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a vector or matrix.}

\item{\code{axis}}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}
}}

