% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{SumSmallest}
\alias{SumSmallest}
\title{The SumSmallest atom.}
\usage{
SumSmallest(x, k)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric matrix.}

\item{k}{The number of smallest values to sum over.}
}
\value{
Sum of the smlalest k values
}
\description{
The sum of the smallest k values of a matrix.
}
