% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{SumSquares}
\alias{SumSquares}
\title{The SumSquares atom.}
\usage{
SumSquares(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} or numeric matrix.}
}
\value{
Sum of the squares of the entries in the expression.
}
\description{
The sum of the squares of the entries.
}
