% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\docType{class}
\name{SuperSCS-class}
\alias{SuperSCS-class}
\alias{SuperSCS}
\alias{name,SuperSCS-method}
\alias{import_solver,SuperSCS-method}
\alias{solve_via_data,SuperSCS-method}
\title{The SuperSCS class.}
\usage{
\S4method{name}{SuperSCS}(x)

\S4method{import_solver}{SuperSCS}(solver)

\S4method{solve_via_data}{SuperSCS}(
  object,
  data,
  warm_start,
  verbose,
  solver_opts,
  solver_cache = list()
)
}
\arguments{
\item{solver, object, x}{A \linkS4class{SuperSCS} object.}

\item{data}{Data generated via an apply call.}

\item{warm_start}{An option for warm start.}

\item{verbose}{A boolean of whether to enable solver verbosity.}

\item{solver_opts}{A list of Solver specific options}

\item{solver_cache}{Cache for the solver.}
}
\description{
Not Implemented.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: Returns the name of the SuperSCS solver

\item \code{import_solver}: imports the SuperSCS solver.

\item \code{solve_via_data}: Solve a problem represented by data returned from apply.
}}

