% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{SuppFuncAtom-class}
\alias{SuppFuncAtom-class}
\alias{.SuppFuncAtom}
\alias{SuppFuncAtom}
\alias{variables,SuppFuncAtom-method}
\alias{parameters,SuppFuncAtom-method}
\alias{constants,SuppFuncAtom-method}
\alias{dim_from_args,SuppFuncAtom-method}
\alias{sign_from_args,SuppFuncAtom-method}
\alias{is_nonneg,SuppFuncAtom-method}
\alias{is_nonpos,SuppFuncAtom-method}
\alias{is_imag,SuppFuncAtom-method}
\alias{is_complex,SuppFuncAtom-method}
\alias{is_atom_convex,SuppFuncAtom-method}
\alias{is_atom_concave,SuppFuncAtom-method}
\alias{is_atom_log_log_convex,SuppFuncAtom-method}
\alias{is_atom_log_log_concave,SuppFuncAtom-method}
\alias{is_atom_quasiconvex,SuppFuncAtom-method}
\alias{is_atom_quasiconcave,SuppFuncAtom-method}
\alias{is_incr,SuppFuncAtom-method}
\alias{is_decr,SuppFuncAtom-method}
\alias{is_convex,SuppFuncAtom-method}
\alias{is_concave,SuppFuncAtom-method}
\alias{is_log_log_convex,SuppFuncAtom-method}
\alias{is_log_log_concave,SuppFuncAtom-method}
\alias{is_quasiconvex,SuppFuncAtom-method}
\alias{is_quasiconcave,SuppFuncAtom-method}
\alias{value_impl,SuppFuncAtom-method}
\alias{.grad,SuppFuncAtom-method}
\alias{<,SuppFuncAtom,Expression-method}
\alias{>,SuppFuncAtom,Expression-method}
\title{The SuppFuncAtom class.}
\usage{
SuppFuncAtom(y, parent)

\S4method{variables}{SuppFuncAtom}(object)

\S4method{parameters}{SuppFuncAtom}(object)

\S4method{constants}{SuppFuncAtom}(object)

\S4method{dim_from_args}{SuppFuncAtom}(object)

\S4method{sign_from_args}{SuppFuncAtom}(object)

\S4method{is_nonneg}{SuppFuncAtom}(object)

\S4method{is_nonpos}{SuppFuncAtom}(object)

\S4method{is_imag}{SuppFuncAtom}(object)

\S4method{is_complex}{SuppFuncAtom}(object)

\S4method{is_atom_convex}{SuppFuncAtom}(object)

\S4method{is_atom_concave}{SuppFuncAtom}(object)

\S4method{is_atom_log_log_convex}{SuppFuncAtom}(object)

\S4method{is_atom_log_log_concave}{SuppFuncAtom}(object)

\S4method{is_atom_quasiconvex}{SuppFuncAtom}(object)

\S4method{is_atom_quasiconcave}{SuppFuncAtom}(object)

\S4method{is_incr}{SuppFuncAtom}(object, idx)

\S4method{is_decr}{SuppFuncAtom}(object, idx)

\S4method{is_convex}{SuppFuncAtom}(object)

\S4method{is_concave}{SuppFuncAtom}(object)

\S4method{is_log_log_convex}{SuppFuncAtom}(object)

\S4method{is_log_log_concave}{SuppFuncAtom}(object)

\S4method{is_quasiconvex}{SuppFuncAtom}(object)

\S4method{is_quasiconcave}{SuppFuncAtom}(object)

\S4method{value_impl}{SuppFuncAtom}(object)

\S4method{.grad}{SuppFuncAtom}(object, values)

\S4method{<}{SuppFuncAtom,Expression}(e1, e2)

\S4method{>}{SuppFuncAtom,Expression}(e1, e2)
}
\arguments{
\item{y}{An \linkS4class{Expression} with \code{is_affine(y) == TRUE}.}

\item{parent}{A \linkS4class{SuppFunc} containing data for the convex set associated with this atom.}

\item{object}{A \linkS4class{SuppFuncAtom} object.}

\item{idx}{An index into the atom.}

\item{values}{A list of numeric values for the arguments}
}
\description{
This class represents the support function of an affine expression \eqn{y}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{variables(SuppFuncAtom)}: List of \linkS4class{Variable} objects in the atom.

\item \code{parameters(SuppFuncAtom)}: List of \linkS4class{Parameter} objects in the atom.

\item \code{constants(SuppFuncAtom)}: List of \linkS4class{Constant} objects in the atom.

\item \code{dim_from_args(SuppFuncAtom)}: The dimensions of the atom.

\item \code{sign_from_args(SuppFuncAtom)}: The sign (is positive, is negative) of the atom.

\item \code{is_nonneg(SuppFuncAtom)}: A logical value indicating whether the atom is nonnegative.

\item \code{is_nonpos(SuppFuncAtom)}: A logical value indicating whether the atom is nonpositive.

\item \code{is_imag(SuppFuncAtom)}: A logical value indicating whether the atom is imaginary.

\item \code{is_complex(SuppFuncAtom)}: A logical value indicating whether the atom is complex valued.

\item \code{is_atom_convex(SuppFuncAtom)}: Is the atom convex?

\item \code{is_atom_concave(SuppFuncAtom)}: Is the atom concave?

\item \code{is_atom_log_log_convex(SuppFuncAtom)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(SuppFuncAtom)}: Is the atom log-log concave?

\item \code{is_atom_quasiconvex(SuppFuncAtom)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(SuppFuncAtom)}: Is the atom quasiconcave?

\item \code{is_incr(SuppFuncAtom)}: Is the atom weakly increasing in argument \code{idx}?

\item \code{is_decr(SuppFuncAtom)}: Is the atom weakly decreasing in argument \code{idx}?

\item \code{is_convex(SuppFuncAtom)}: Is the atom convex?

\item \code{is_concave(SuppFuncAtom)}: Is the atom concave?

\item \code{is_log_log_convex(SuppFuncAtom)}: Is the atom log-log convex?

\item \code{is_log_log_concave(SuppFuncAtom)}: Is the atom log-log concave?

\item \code{is_quasiconvex(SuppFuncAtom)}: Is the atom quasiconvex?

\item \code{is_quasiconcave(SuppFuncAtom)}: Is the atom quasiconcave?

\item \code{value_impl(SuppFuncAtom)}: Returns the value of each of the components in the atom.

\item \code{.grad(SuppFuncAtom)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{e1 < e2}: Strict inequalities are not allowed.

\item \code{e1 > e2}: Strict inequalities are not allowed.

}}
\section{Slots}{

\describe{
\item{\code{y}}{An \linkS4class{Expression} with \code{is_affine(y) == TRUE}.}

\item{\code{parent}}{A \linkS4class{SuppFunc} containing data for the convex set associated with this atom.}
}}

