% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_qp_solver.R
\name{supported_constraints,QpSolver-method}
\alias{supported_constraints,QpSolver-method}
\title{What classes of constraints does the solver support?}
\usage{
\S4method{supported_constraints}{QpSolver}(solver)
}
\arguments{
\item{object}{A \linkS4class{QpSolver} object.}
}
\description{
What classes of constraints does the solver support?
}
