% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{SymbolicQuadForm-class}
\alias{SymbolicQuadForm-class}
\alias{.SymbolicQuadForm}
\alias{SymbolicQuadForm}
\alias{.grad,SymbolicQuadForm-method}
\title{The SymbolicQuadForm class.}
\usage{
SymbolicQuadForm(x, P, expr)

\S4method{.grad}{SymbolicQuadForm}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric vector.}

\item{P}{An \linkS4class{Expression}, numeric matrix, or vector.}

\item{values}{A list of numeric values for the arguments}

\item{original_expression}{The original \linkS4class{Expression}.}
}
\description{
The SymbolicQuadForm class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{.grad}: Gives the (sub/super)gradient of the atom w.r.t. each variable
}}

\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} or numeric vector.}

\item{\code{P}}{An \linkS4class{Expression}, numeric matrix, or vector.}

\item{\code{original_expression}}{The original \linkS4class{Expression}.}
}}

