% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.R
\name{Symmetric}
\alias{Symmetric}
\title{Symmetric Variable}
\usage{
Symmetric(n, name = NA_character_)
}
\arguments{
\item{n}{The number of rows/columns in the matrix.}

\item{name}{(Optional) A character string representing the name of the variable.}
}
\value{
An \linkS4class{Expression} representing the symmetric matrix.
}
\description{
An expression representing a symmetric matrix.
}
\examples{
x <- Symmetric(3, name="s3")
}
