% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050_wraps.R
\docType{class}
\name{SymmetricWrap-class}
\alias{SymmetricWrap-class}
\alias{.SymmetricWrap}
\alias{SymmetricWrap}
\alias{validate_args,SymmetricWrap-method}
\alias{is_symmetric,SymmetricWrap-method}
\alias{is_hermitian,SymmetricWrap-method}
\title{The SymmetricWrap class.}
\usage{
SymmetricWrap(arg)

\S4method{validate_args}{SymmetricWrap}(object)

\S4method{is_symmetric}{SymmetricWrap}(object)

\S4method{is_hermitian}{SymmetricWrap}(object)
}
\arguments{
\item{arg}{A \linkS4class{Expression} object or matrix.}

\item{object}{A \linkS4class{SymmetricWrap} object.}
}
\description{
A no-op wrapper to assert the input argument is symmetric.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args(SymmetricWrap)}: Check the input is a real square matrix.

\item \code{is_symmetric(SymmetricWrap)}: Is the atom symmetric?

\item \code{is_hermitian(SymmetricWrap)}: Is the atom Hermitian?

}}
