% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{to_numeric}
\alias{to_numeric}
\title{Numeric Value of Atom}
\usage{
to_numeric(object, values)
}
\arguments{
\item{object}{An \linkS4class{Atom} object.}

\item{values}{A list of arguments to the atom.}
}
\value{
A numeric scalar, vector, or matrix.
}
\description{
Returns the numeric value of the atom evaluated on the specified arguments.
}
