% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/110_total_variation.R
\name{TotalVariation}
\alias{TotalVariation}
\title{The TotalVariation atom.}
\usage{
TotalVariation(value, ...)
}
\arguments{
\item{value}{An \linkS4class{Expression} representing the value to take the total variation of.}

\item{...}{Additional matrices extending the third dimension of value.}
}
\value{
An expression representing the total variation.
}
\description{
The total variation of a vector, matrix, or list of matrices.
Uses L1 norm of discrete gradients for vectors and L2 norm of discrete gradients for matrices.
}
