% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{Trace-class}
\alias{Trace-class}
\alias{.Trace}
\alias{Trace}
\alias{Trace}
\alias{validate_args,Trace-method}
\alias{to_numeric,Trace-method}
\alias{size_from_args,Trace-method}
\alias{graph_implementation,Trace-method}
\title{The Trace class.}
\usage{
Trace(expr)

\S4method{validate_args}{Trace}(object)

\S4method{to_numeric}{Trace}(object, values)

\S4method{size_from_args}{Trace}(object)

\S4method{graph_implementation}{Trace}(object, arg_objs, size,
  data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a matrix.}

\item{object}{A \linkS4class{Trace} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the sum of the diagonal entries in a matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args}: Check the argument is a square matrix.

\item \code{to_numeric}: Sum the diagonal entries.

\item \code{size_from_args}: The atom is a scalar.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a matrix.}
}}

