% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{Transpose-class}
\alias{Transpose-class}
\alias{Transpose}
\alias{to_numeric,Transpose-method}
\alias{size_from_args,Transpose-method}
\alias{graph_implementation,Transpose-method}
\title{The Transpose class.}
\usage{
\S4method{to_numeric}{Transpose}(object, values)

\S4method{size_from_args}{Transpose}(object)

\S4method{graph_implementation}{Transpose}(object, arg_objs, size, data = NA_real_)
}
\arguments{
\item{object}{A \linkS4class{Transpose} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the matrix transpose.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: The transpose of the given value.

\item \code{size_from_args}: The size of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

