% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\docType{methods}
\name{t.Expression}
\alias{t.Expression}
\alias{t}
\alias{t,Expression-method}
\title{Matrix Transpose}
\usage{
\method{t}{Expression}(x)

\S4method{t}{Expression}(x)
}
\arguments{
\item{x}{An \linkS4class{Expression} representing a matrix.}
}
\value{
An \linkS4class{Expression} representing the transposed matrix.
}
\description{
The transpose of a matrix.
}
\examples{
x <- Variable(3, 4)
t(x)
}
