% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{tri_to_full}
\alias{tri_to_full}
\title{Expands lower triangular to full matrix.}
\usage{
tri_to_full(lower_tri, n)
}
\arguments{
\item{lower_tri}{A matrix representing the lower triangular part of the matrix,
stacked in column-major order}

\item{n}{The number of rows (columns) in the full square matrix.}
}
\value{
A matrix that is the scaled expansion of the lower triangular matrix.
}
\description{
Expands lower triangular to full matrix.
}
