% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{TrInv-class}
\alias{TrInv-class}
\alias{.TrInv}
\alias{TrInv}
\alias{to_numeric,TrInv-method}
\alias{validate_args,TrInv-method}
\alias{dim_from_args,TrInv-method}
\alias{sign_from_args,TrInv-method}
\alias{is_atom_convex,TrInv-method}
\alias{is_atom_concave,TrInv-method}
\alias{is_incr,TrInv-method}
\alias{is_decr,TrInv-method}
\alias{.grad,TrInv-method}
\alias{.domain,TrInv-method}
\alias{value,TrInv-method}
\title{The TrInv atom.}
\usage{
TrInv(X)

\S4method{to_numeric}{TrInv}(object, values)

\S4method{validate_args}{TrInv}(object)

\S4method{dim_from_args}{TrInv}(object)

\S4method{sign_from_args}{TrInv}(object)

\S4method{is_atom_convex}{TrInv}(object)

\S4method{is_atom_concave}{TrInv}(object)

\S4method{is_incr}{TrInv}(object, idx)

\S4method{is_decr}{TrInv}(object, idx)

\S4method{.grad}{TrInv}(object, values)

\S4method{.domain}{TrInv}(object)

\S4method{value}{TrInv}(object)
}
\arguments{
\item{X}{An \linkS4class{Expression}}

\item{object}{A \linkS4class{TrInv} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
The trace of the inverse of a positive definite matrix \eqn{X}, \eqn{tr(X^{-1})}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(TrInv)}: The trace of the inverse of a positive definite matrix.

\item \code{validate_args(TrInv)}: Check that the matrix is square.

\item \code{dim_from_args(TrInv)}: The dimensions of the atom determined from its arguments.

\item \code{sign_from_args(TrInv)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(TrInv)}: Is the atom convex?

\item \code{is_atom_concave(TrInv)}: Is the atom concave?

\item \code{is_incr(TrInv)}: Is the atom weakly increasing in the index?

\item \code{is_decr(TrInv)}: Is the atom weakly decreasing in the index?

\item \code{.grad(TrInv)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.domain(TrInv)}: Returns constraints describing the domain of the node.

\item \code{value(TrInv)}: Returns the value of the atom.

}}
\section{Slots}{

\describe{
\item{\code{X}}{An \linkS4class{Expression} representing a positive definite matrix.}
}}

