% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarabel.R
\name{triu_to_full}
\alias{triu_to_full}
\title{Expands upper triangular to full matrix.}
\usage{
triu_to_full(upper_tri, n)
}
\arguments{
\item{upper_tri}{A matrix representing the uppertriangular part of the matrix,
stacked in column-major order}

\item{n}{The number of rows (columns) in the full square matrix.}
}
\value{
A matrix that is the scaled expansion of the upper triangular matrix.
}
\description{
Expands upper triangular to full matrix.
}
