% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/267_zzz_R_specific_exports.R
\name{tv}
\alias{tv}
\title{Total variation (deprecated alias)}
\usage{
tv(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[=total_variation]{total_variation()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link[=total_variation]{total_variation()}} instead.
}
\seealso{
\code{\link[=total_variation]{total_variation()}}
}
