% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{UnaryOperator-class}
\alias{UnaryOperator-class}
\alias{UnaryOperator}
\alias{name,UnaryOperator-method}
\alias{to_numeric,UnaryOperator-method}
\title{The UnaryOperator class.}
\usage{
\S4method{name}{UnaryOperator}(x)

\S4method{to_numeric}{UnaryOperator}(object, values)
}
\arguments{
\item{x, object}{A \linkS4class{UnaryOperator} object.}

\item{values}{A list of arguments to the atom.}
}
\description{
This base class represents expressions involving unary operators.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: Returns the expression in string form.

\item \code{to_numeric}: Applies the unary operator to the value.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{The \linkS4class{Expression} that is being operated upon.}

\item{\code{op_name}}{A \code{character} string indicating the unary operation.}
}}

