% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{validate_args}
\alias{validate_args}
\title{Validate Arguments}
\usage{
validate_args(object)
}
\arguments{
\item{object}{An \linkS4class{Atom} object.}
}
\description{
Validate an atom's arguments, returning an error if any are invalid.
}
