% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate Arguments to an Atom}
\usage{
validate_arguments(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
Invisible \code{NULL} (or error if invalid).
}
\description{
Validate Arguments to an Atom
}
\keyword{internal}
