% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{validate_val}
\alias{validate_val}
\title{Validate Value}
\usage{
validate_val(object, val)
}
\arguments{
\item{object}{A \linkS4class{Leaf} object.}

\item{val}{The assigned value.}
}
\value{
The value converted to proper matrix type.
}
\description{
Check that the value satisfies a \linkS4class{Leaf}'s symbolic attributes.
}
