% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/184_valinvec2mixedint.R
\docType{class}
\name{Valinvec2MixedInt-class}
\alias{Valinvec2MixedInt-class}
\alias{.Valinvec2MixedInt}
\alias{accepts,Valinvec2MixedInt,Problem-method}
\title{The Valinvec2MixedInt class.}
\usage{
\S4method{accepts}{Valinvec2MixedInt,Problem}(object, problem)
}
\arguments{
\item{object}{A \linkS4class{Valinvec2MixedInt} object.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
This class represents a discrete to mixed-integer reduction.
}
\section{Methods (by generic)}{
\itemize{
\item \code{accepts(object = Valinvec2MixedInt, problem = Problem)}: Checks whether or not the problem involves any finite set constraints.

}}
