% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{value}
\alias{value}
\title{Get the Numeric Value of an Expression}
\usage{
value(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
A numeric matrix, or \code{NULL} if no value has been set.
}
\description{
Returns the numeric value of a CVXR expression, variable, or constant.
For variables, the value is set after solving a problem.
}
