% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{variable_of_provenance}
\alias{variable_of_provenance}
\title{Return a variable with attributes from which this variable was generated}
\usage{
variable_of_provenance(object, variable)
}
\arguments{
\item{object}{A \linkS4class{Variable} object.}
}
\value{
a variable of provenance attributes
}
\description{
Return a variable with attributes from which this variable was generated
}
