% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{VecLength-class}
\alias{VecLength-class}
\alias{.VecLength}
\alias{VecLength}
\alias{to_numeric,VecLength-method}
\alias{dim_from_args,VecLength-method}
\alias{sign_from_args,VecLength-method}
\alias{is_atom_convex,VecLength-method}
\alias{is_atom_concave,VecLength-method}
\alias{is_atom_quasiconvex,VecLength-method}
\alias{is_atom_quasiconcave,VecLength-method}
\alias{is_incr,VecLength-method}
\alias{is_decr,VecLength-method}
\alias{.grad,VecLength-method}
\title{The VecLength class.}
\usage{
VecLength(x)

\S4method{to_numeric}{VecLength}(object, values)

\S4method{dim_from_args}{VecLength}(object)

\S4method{sign_from_args}{VecLength}(object)

\S4method{is_atom_convex}{VecLength}(object)

\S4method{is_atom_concave}{VecLength}(object)

\S4method{is_atom_quasiconvex}{VecLength}(object)

\S4method{is_atom_quasiconcave}{VecLength}(object)

\S4method{is_incr}{VecLength}(object, idx)

\S4method{is_decr}{VecLength}(object, idx)

\S4method{.grad}{VecLength}(object, values)
}
\arguments{
\item{x}{An \linkS4class{Expression} representing a vector.}

\item{object}{A \linkS4class{VecLength} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents the length of a vector (index of last nonzero, ones-based).
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(VecLength)}: The length of the vector.

\item \code{dim_from_args(VecLength)}: The dimensions of the atom determined from its arguments.

\item \code{sign_from_args(VecLength)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(VecLength)}: Is the atom convex?

\item \code{is_atom_concave(VecLength)}: Is the atom concave?

\item \code{is_atom_quasiconvex(VecLength)}: Is the atom quasiconvex?

\item \code{is_atom_quasiconcave(VecLength)}: Is the atom quasiconcave?

\item \code{is_incr(VecLength)}: Is the atom weakly increasing in the index?

\item \code{is_decr(VecLength)}: Is the atom weakly decreasing in the index?

\item \code{.grad(VecLength)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} representing a vector.}
}}

