% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{vectorized_lower_tri_to_mat}
\alias{vectorized_lower_tri_to_mat}
\title{Turns symmetric 2D array into a lower triangular matrix}
\usage{
vectorized_lower_tri_to_mat(v, dim)
}
\arguments{
\item{v}{A list of length (dim * (dim + 1) / 2).}

\item{dim}{The number of rows (equivalently, columns) in the output array.}
}
\value{
Return the symmetric 2D array defined by taking "v" to specify its
lower triangular matrix.
}
\description{
Turns symmetric 2D array into a lower triangular matrix
}
