% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{vectorized_lower_tri_to_triples}
\alias{vectorized_lower_tri_to_triples}
\title{Return the triplet form of the symmetric 2D array defined by taking
a vector specifying its lower triangular entries in column-major
order}
\usage{
vectorized_lower_tri_to_triples(v, dim)
}
\arguments{
\item{v}{A list of length (dim * (dim + 1) / 2).}

\item{dim}{The number of rows (equivalently, columns) in the output
array.}
}
\value{
Return the symmetric 2D array defined by taking `v` to
  specify its lower triangular matrix.
}
\description{
Return the triplet form of the symmetric 2D array defined by taking
a vector specifying its lower triangular entries in column-major
order
}
