% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{Version-class}
\alias{Version-class}
\alias{.Version}
\alias{Version}
\alias{==,Version,Version-method}
\alias{!=,Version,Version-method}
\alias{<=,Version,Version-method}
\alias{<,Version,Version-method}
\alias{>=,Version,Version-method}
\alias{>,Version,Version-method}
\alias{as.character,Version-method}
\title{The Version class.}
\usage{
\S4method{==}{Version,Version}(e1, e2)

\S4method{!=}{Version,Version}(e1, e2)

\S4method{<=}{Version,Version}(e1, e2)

\S4method{<}{Version,Version}(e1, e2)

\S4method{>=}{Version,Version}(e1, e2)

\S4method{>}{Version,Version}(e1, e2)

\S4method{as.character}{Version}(x)
}
\arguments{
\item{e1, e2}{The \linkS4class{Version} objects or numeric constants to compare.}
}
\description{
The class represents the version of CVXR.
}
\details{
It parses strings of the form 'x.y.z+[STUFF]' and tuples of the form (x, y, z).
In the former case, '+[STUFF]' refers to a local version identifier in the
sense of https://www.python.org/dev/peps/pep-0440/#local-version-identifiers.
We don't actually store the local version identifier for comparisons.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.character(Version)}: The string representation of the version.

}}
