% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/005_zzz_R_specific_generics.R
\name{violation}
\alias{violation}
\title{Get the Violation of a Constraint}
\usage{
violation(x, ...)
}
\arguments{
\item{x}{A constraint object.}

\item{...}{Not used.}
}
\value{
Numeric scalar.
}
\description{
Returns the scalar violation (distance to feasibility) of a constraint.
}
