% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/112_von_neumann_entr.R
\docType{class}
\name{VonNeumannEntr-class}
\alias{VonNeumannEntr-class}
\alias{.VonNeumannEntr}
\alias{VonNeumannEntr}
\alias{to_numeric,VonNeumannEntr-method}
\alias{validate_args,VonNeumannEntr-method}
\alias{sign_from_args,VonNeumannEntr-method}
\alias{dim_from_args,VonNeumannEntr-method}
\alias{is_atom_convex,VonNeumannEntr-method}
\alias{is_atom_concave,VonNeumannEntr-method}
\alias{is_incr,VonNeumannEntr-method}
\alias{is_decr,VonNeumannEntr-method}
\alias{get_data,VonNeumannEntr-method}
\alias{.grad,VonNeumannEntr-method}
\alias{.domain,VonNeumannEntr-method}
\title{The VonNeumannEntr atom.}
\usage{
VonNeumannEntr(X, quad_approx)

\S4method{to_numeric}{VonNeumannEntr}(object, values)

\S4method{validate_args}{VonNeumannEntr}(object)

\S4method{sign_from_args}{VonNeumannEntr}(object)

\S4method{dim_from_args}{VonNeumannEntr}(object)

\S4method{is_atom_convex}{VonNeumannEntr}(object)

\S4method{is_atom_concave}{VonNeumannEntr}(object)

\S4method{is_incr}{VonNeumannEntr}(object, idx)

\S4method{is_decr}{VonNeumannEntr}(object, idx)

\S4method{get_data}{VonNeumannEntr}(object)

\S4method{.grad}{VonNeumannEntr}(object, values)

\S4method{.domain}{VonNeumannEntr}(object)
}
\arguments{
\item{X}{An \linkS4class{Expression} or numeric matrix.}

\item{object}{A \linkS4class{VonNeumannEntr} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
The von Neumann entropy of the positive definite matrix \eqn{X \in \mathbb{S}_{+}^n}, \eqn{tr(X logm(X))},
where tr() is the trace and logm() is the matrix logarithm.
}
\details{
May alternatively be expressed as \eqn{von_neumann_entr(X) = -\sum_{i=1}^n \lambda_i \log \lambda_i},
where \eqn{\lambda_i} are the eigenvalues of \eqn{X}.

This atom does not enforce \eqn{tr(X) = 1} as is expected in applications from quantum mechanics.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(VonNeumannEntr)}: The von Neumann entropy of the matrix.

\item \code{validate_args(VonNeumannEntr)}: Check that the matrix is positive semidefinite.

\item \code{sign_from_args(VonNeumannEntr)}: The (is positive, is negative) sign of the atom.

\item \code{dim_from_args(VonNeumannEntr)}: The dimensions of the atom determined from its arguments.

\item \code{is_atom_convex(VonNeumannEntr)}: Is the atom convex?

\item \code{is_atom_concave(VonNeumannEntr)}: Is the atom concave?

\item \code{is_incr(VonNeumannEntr)}: Is the atom weakly increasing in the index?

\item \code{is_decr(VonNeumannEntr)}: Is the atom weakly decreasing in the index?

\item \code{get_data(VonNeumannEntr)}: Returns the parameter \code{quad_approx}.

\item \code{.grad(VonNeumannEntr)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.domain(VonNeumannEntr)}: Returns constraints describing the domain of the node.

}}
\section{Slots}{

\describe{
\item{\code{X}}{An \linkS4class{Expression} representing a positive semidefinite matrix.}
}}

