% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{VStack-class}
\alias{VStack-class}
\alias{.VStack}
\alias{VStack}
\alias{to_numeric,VStack-method}
\alias{validate_args,VStack-method}
\alias{dim_from_args,VStack-method}
\alias{is_atom_log_log_convex,VStack-method}
\alias{is_atom_log_log_concave,VStack-method}
\alias{graph_implementation,VStack-method}
\title{The VStack class.}
\usage{
VStack(...)

\S4method{to_numeric}{VStack}(object, values)

\S4method{validate_args}{VStack}(object)

\S4method{dim_from_args}{VStack}(object)

\S4method{is_atom_log_log_convex}{VStack}(object)

\S4method{is_atom_log_log_concave}{VStack}(object)

\S4method{graph_implementation}{VStack}(object, arg_objs, dim, data = NA_real_)
}
\arguments{
\item{...}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of columns.}

\item{object}{A \linkS4class{VStack} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
Vertical concatenation of values.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(VStack)}: Vertically concatenate the values using \code{rbind}.

\item \code{validate_args(VStack)}: Check all arguments have the same width.

\item \code{dim_from_args(VStack)}: The dimensions of the atom.

\item \code{is_atom_log_log_convex(VStack)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(VStack)}: Is the atom log-log concave?

\item \code{graph_implementation(VStack)}: The graph implementation of the atom.

}}
\section{Slots}{

\describe{
\item{\code{...}}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of columns.}
}}

