% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/052_atoms_affine_vstack.R
\name{vstack}
\alias{vstack}
\title{Vertical concatenation of expressions}
\usage{
vstack(...)
}
\arguments{
\item{...}{Expressions (same number of columns)}
}
\value{
A VStack atom
}
\description{
Vertical concatenation of expressions
}
