% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_utilities_scopes.R
\name{with_quad_form_dpp_scope}
\alias{with_quad_form_dpp_scope}
\title{Execute Expression Within a quad_form DPP Scope}
\usage{
with_quad_form_dpp_scope(expr)
}
\arguments{
\item{expr}{An R expression to evaluate within the quad_form DPP scope.}
}
\value{
The result of evaluating \code{expr}.
}
\description{
Execute Expression Within a quad_form DPP Scope
}
\keyword{internal}
