% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{XExp-class}
\alias{XExp-class}
\alias{.XExp}
\alias{XExp}
\alias{to_numeric,XExp-method}
\alias{sign_from_args,XExp-method}
\alias{is_atom_convex,XExp-method}
\alias{is_atom_concave,XExp-method}
\alias{is_atom_log_log_convex,XExp-method}
\alias{is_atom_log_log_concave,XExp-method}
\alias{is_incr,XExp-method}
\alias{is_decr,XExp-method}
\alias{.grad,XExp-method}
\alias{.domain,XExp-method}
\title{The XExp class.}
\usage{
XExp(x)

\S4method{to_numeric}{XExp}(object, values)

\S4method{sign_from_args}{XExp}(object)

\S4method{is_atom_convex}{XExp}(object)

\S4method{is_atom_concave}{XExp}(object)

\S4method{is_atom_log_log_convex}{XExp}(object)

\S4method{is_atom_log_log_concave}{XExp}(object)

\S4method{is_incr}{XExp}(object, idx)

\S4method{is_decr}{XExp}(object, idx)

\S4method{.grad}{XExp}(object, values)

\S4method{.domain}{XExp}(object)
}
\arguments{
\item{x}{An \linkS4class{Expression}.}

\item{object}{A \linkS4class{XExp} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}
}
\description{
This class represents elementwise \eqn{x\exp^x}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(XExp)}: The numeric value of the expression.

\item \code{sign_from_args(XExp)}: The (is positive, is negative) sign of the atom.

\item \code{is_atom_convex(XExp)}: Is the atom convex?

\item \code{is_atom_concave(XExp)}: Is the atom concave?

\item \code{is_atom_log_log_convex(XExp)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(XExp)}: Is the atom log-log concave?

\item \code{is_incr(XExp)}: Is the atom weakly increasing in the index?

\item \code{is_decr(XExp)}: Is the atom weakly decreasing in the index?

\item \code{.grad(XExp)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.domain(XExp)}: Returns constraints describing the domain of the node

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression}.}
}}

