% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/147_xpress_problem.R
\name{XpressProblem-arith}
\alias{XpressProblem-arith}
\alias{-,XpressProblem,missing-method}
\alias{*,XpressProblem,numeric-method}
\alias{+,XpressProblem,numeric-method}
\alias{+,numeric,XpressProblem-method}
\alias{+,XpressProblem,XpressProblem-method}
\alias{-,XpressProblem,numeric-method}
\alias{-,numeric,XpressProblem-method}
\alias{-,XpressProblem,XpressProblem-method}
\alias{*,numeric,XpressProblem-method}
\alias{/,Problem,numeric-method}
\title{Arithmetic Operations on XpressProblems}
\usage{
\S4method{-}{XpressProblem,missing}(e1, e2)

\S4method{*}{XpressProblem,numeric}(e1, e2)

\S4method{+}{XpressProblem,numeric}(e1, e2)

\S4method{+}{numeric,XpressProblem}(e1, e2)

\S4method{+}{XpressProblem,XpressProblem}(e1, e2)

\S4method{-}{XpressProblem,numeric}(e1, e2)

\S4method{-}{numeric,XpressProblem}(e1, e2)

\S4method{-}{XpressProblem,XpressProblem}(e1, e2)

\S4method{*}{numeric,XpressProblem}(e1, e2)

\S4method{/}{Problem,numeric}(e1, e2)
}
\arguments{
\item{e1}{The left-hand \linkS4class{XpressProblem} object.}

\item{e2}{The right-hand \linkS4class{XpressProblem} object.}
}
\value{
A \linkS4class{XpressProblem} object.
}
\description{
Add, subtract, multiply, or divide DCP optimization problems.
}
