% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\docType{class}
\name{XpressProblem-class}
\alias{XpressProblem-class}
\alias{.XpressProblem}
\alias{XpressProblem}
\title{The XpressProblem class.}
\usage{
XpressProblem(objective, constraints = list())
}
\arguments{
\item{objective}{A \linkS4class{Minimize} or \linkS4class{Maximize} object representing the optimization objective.}

\item{constraints}{(Optional) A list of \linkS4class{Constraint} objects representing constraints on the optimization variables.}
}
\description{
This class represents a convex optimization problem associated with the Xpress Optimizer.
}
\section{Slots}{

\describe{
\item{\code{objective}}{A \linkS4class{Minimize} or \linkS4class{Maximize} object representing the optimization objective.}

\item{\code{constraints}}{A list of \linkS4class{Constraint} objects representing constraints on the optimization variables.}
}}

