% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{AffAtom-class}
\alias{AffAtom-class}
\alias{AffAtom}
\alias{allow_complex,AffAtom-method}
\alias{sign_from_args,AffAtom-method}
\alias{is_imag,AffAtom-method}
\alias{is_complex,AffAtom-method}
\alias{is_atom_convex,AffAtom-method}
\alias{is_atom_concave,AffAtom-method}
\alias{is_incr,AffAtom-method}
\alias{is_decr,AffAtom-method}
\alias{is_quadratic,AffAtom-method}
\alias{has_quadratic_term,AffAtom-method}
\alias{is_qpwa,AffAtom-method}
\alias{is_pwl,AffAtom-method}
\alias{is_psd,AffAtom-method}
\alias{is_nsd,AffAtom-method}
\alias{.grad,AffAtom-method}
\title{The AffAtom class.}
\usage{
\S4method{allow_complex}{AffAtom}(object)

\S4method{sign_from_args}{AffAtom}(object)

\S4method{is_imag}{AffAtom}(object)

\S4method{is_complex}{AffAtom}(object)

\S4method{is_atom_convex}{AffAtom}(object)

\S4method{is_atom_concave}{AffAtom}(object)

\S4method{is_incr}{AffAtom}(object, idx)

\S4method{is_decr}{AffAtom}(object, idx)

\S4method{is_quadratic}{AffAtom}(object)

\S4method{has_quadratic_term}{AffAtom}(object)

\S4method{is_qpwa}{AffAtom}(object)

\S4method{is_pwl}{AffAtom}(object)

\S4method{is_psd}{AffAtom}(object)

\S4method{is_nsd}{AffAtom}(object)

\S4method{.grad}{AffAtom}(object, values)
}
\arguments{
\item{object}{An \linkS4class{AffAtom} object.}

\item{idx}{An index into the atom.}

\item{values}{A list of numeric values for the arguments}
}
\description{
This virtual class represents an affine atomic expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{allow_complex(AffAtom)}: Does the atom handle complex numbers?

\item \code{sign_from_args(AffAtom)}: The sign of the atom.

\item \code{is_imag(AffAtom)}: Is the atom imaginary?

\item \code{is_complex(AffAtom)}: Is the atom complex valued?

\item \code{is_atom_convex(AffAtom)}: The atom is convex.

\item \code{is_atom_concave(AffAtom)}: The atom is concave.

\item \code{is_incr(AffAtom)}: The atom is weakly increasing in every argument.

\item \code{is_decr(AffAtom)}: The atom is not weakly decreasing in any argument.

\item \code{is_quadratic(AffAtom)}: Is every argument quadratic?

\item \code{has_quadratic_term(AffAtom)}: Does the affine head of the expression contain a quadratic term? The affine head is all nodes with a path to the root node that does not pass through any non-affine atom. If the root node is non-affine, then the affine head is the root alone.

\item \code{is_qpwa(AffAtom)}: Is every argument quadratic of piecewise affine?

\item \code{is_pwl(AffAtom)}: Is every argument piecewise linear?

\item \code{is_psd(AffAtom)}: Is the atom a positive semidefinite matrix?

\item \code{is_nsd(AffAtom)}: Is the atom a negative semidefinite matrix?

\item \code{.grad(AffAtom)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

}}
