% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{BinaryOperator-class}
\alias{BinaryOperator-class}
\alias{BinaryOperator}
\alias{name,BinaryOperator-method}
\alias{to_numeric,BinaryOperator-method}
\alias{sign_from_args,BinaryOperator-method}
\alias{is_imag,BinaryOperator-method}
\alias{is_complex,BinaryOperator-method}
\title{The BinaryOperator class.}
\usage{
\S4method{name}{BinaryOperator}(x)

\S4method{to_numeric}{BinaryOperator}(object, values)

\S4method{sign_from_args}{BinaryOperator}(object)

\S4method{is_imag}{BinaryOperator}(object)

\S4method{is_complex}{BinaryOperator}(object)
}
\arguments{
\item{x, object}{A \linkS4class{BinaryOperator} object.}

\item{values}{A list of arguments to the atom.}
}
\description{
This base class represents expressions involving binary operators.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: Returns the name of the BinaryOperator object.

\item \code{to_numeric}: Apply the binary operator to the values.

\item \code{sign_from_args}: Default to rule for multiplication.

\item \code{is_imag}: Is the expression imaginary?

\item \code{is_complex}: Is the expression complex valued?
}}

\section{Slots}{

\describe{
\item{\code{lh_exp}}{The \linkS4class{Expression} on the left-hand side of the operator.}

\item{\code{rh_exp}}{The \linkS4class{Expression} on the right-hand side of the operator.}

\item{\code{op_name}}{A \code{character} string indicating the binary operation.}
}}

