% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/055_atoms_affine_bmat.R
\name{bmat}
\alias{bmat}
\title{Construct a Block Matrix}
\usage{
bmat(block_lists)
}
\arguments{
\item{block_lists}{A list of lists of Expression objects (or numerics).
Each inner list forms one block row.}
}
\value{
An Expression representing the block matrix.
}
\description{
Takes a list of lists. Each internal list is stacked horizontally.
The internal lists are stacked vertically.
}
