% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/039_atoms_affine_promote.R
\name{broadcast_args}
\alias{broadcast_args}
\title{Broadcast two expressions for binary operations}
\usage{
broadcast_args(lh_expr, rh_expr)
}
\arguments{
\item{lh_expr}{Left-hand expression}

\item{rh_expr}{Right-hand expression}
}
\value{
List of two expressions with compatible shapes
}
\description{
Broadcast two expressions for binary operations
}
\keyword{internal}
