% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/304_canonInterface.R
\name{build_lin_op_tree}
\alias{build_lin_op_tree}
\title{Construct C++ LinOp tree from Python LinOp tree.}
\usage{
build_lin_op_tree(root_linR, linR_to_linC)
}
\arguments{
\item{root_linR}{The root of the Python LinOp tree.}

\item{linR_to_linC}{A list for memoizing construction and storing
the C++ LinOps.}
}
\description{
Constructed C++ linOps are stored in the linR_to_linC list,
which maps Python linOps to their corresponding C++ linOps.
}
