% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant.R
\docType{class}
\name{CallbackParam-class}
\alias{CallbackParam-class}
\alias{.CallbackParam}
\alias{CallbackParam}
\alias{value,CallbackParam-method}
\title{The CallbackParam class.}
\usage{
CallbackParam(callback, dim = NULL, ...)

\S4method{value}{CallbackParam}(object)
}
\arguments{
\item{callback}{A callback function that generates the parameter value.}

\item{dim}{The dimensions of the parameter.}

\item{object}{A \linkS4class{CallbackParam} object.}

\item{name}{(Optional) A character string representing the name of the parameter.}

\item{sign}{A character string indicating the sign of the parameter. Must be "ZERO", "NONNEGATIVE", "NONPOSITIVE", or "UNKNOWN".}
}
\description{
This class represents a parameter whose value is obtained by evaluating a function.
}
\section{Slots}{

\describe{
\item{\code{callback}}{A numeric element, vector, matrix, or data.frame.}
}}

\examples{
x <- Variable(2)
y <- CallbackParam(value(x), dim(x), sign = "NONNEGATIVE")
get_data(y)
}
