% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{canonicalize}
\alias{canonicalize}
\title{Canonicalize an Expression}
\usage{
canonicalize(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
List with canonicalized expression and constraints.
}
\description{
Canonicalize an Expression
}
\keyword{internal}
