% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdiac}
\alias{cdiac}
\title{Global Monthly and Annual Temperature Anomalies (degrees C), 1850-2015
(Relative to the 1961-1990 Mean) (May 2016)}
\format{
A data frame with 166 rows and 14 variables:
\describe{
  \item{year}{Year}
  \item{jan}{Anomaly for month of January}
  \item{feb}{Anomaly for month of February}
  \item{mar}{Anomaly for month of March}
  \item{apr}{Anomaly for month of April}
  \item{may}{Anomaly for month of May}
  \item{jun}{Anomaly for month of June}
  \item{jul}{Anomaly for month of July}
  \item{aug}{Anomaly for month of August}
  \item{sep}{Anomaly for month of September}
  \item{oct}{Anomaly for month of October}
  \item{nov}{Anomaly for month of November}
  \item{dec}{Anomaly for month of December}
  \item{annual}{Annual anomaly for the year}
}
}
\source{
\url{https://ess-dive.lbl.gov/}
}
\usage{
cdiac
}
\description{
Global Monthly and Annual Temperature Anomalies (degrees C), 1850-2015
(Relative to the 1961-1990 Mean) (May 2016)
}
\references{
\url{https://ess-dive.lbl.gov/}
}
\keyword{data}
