% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{complex-methods}
\alias{complex-methods}
\alias{is_real}
\alias{is_imag}
\alias{is_complex}
\title{Complex Properties}
\usage{
is_real(object)

is_imag(object)

is_complex(object)
}
\arguments{
\item{object}{An \linkS4class{Expression} object.}
}
\value{
A logical value.
}
\description{
Determine if an expression is real, imaginary, or complex.
}
