% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex2real.R
\name{Complex2Real.abs_canon}
\alias{Complex2Real.abs_canon}
\title{Complex canonicalizer for the absolute value atom}
\usage{
Complex2Real.abs_canon(expr, real_args, imag_args, real2imag)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object}

\item{real_args}{A list of \linkS4class{Constraint} objects for the real part of the expression}

\item{imag_args}{A list of \linkS4class{Constraint} objects for the imaginary part of the expression}
}
\value{
A canonicalization of the absolute value atom of a complex expression, where the returned
variables are its real and imaginary components parsed out.
}
\description{
Complex canonicalizer for the absolute value atom
}
