% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/176_complex2real.R
\name{Complex2Real.canonicalize_tree}
\alias{Complex2Real.canonicalize_tree}
\title{Recursively Canonicalizes a Complex Expression.}
\usage{
Complex2Real.canonicalize_tree(expr, real2imag, leaf_map)
}
\arguments{
\item{expr}{An \linkS4class{Expression} object.}

\item{real2imag}{A list mapping the ID of the real part of a complex expression to the ID of its imaginary part.}

\item{leaf_map}{A map that consists of a tree representation of the expression.}
}
\value{
A list of the parsed out real and imaginary components of the
expression that was constructed by performing the canonicalization of each leaf
in the tree.
}
\description{
Recursively Canonicalizes a Complex Expression.
}
