% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/230_cone_matrix_stuffing.R
\docType{class}
\name{ConeDims-class}
\alias{ConeDims-class}
\alias{ConeDims}
\title{Summary of Cone Dimensions Present in Constraints}
\description{
Constraints must be formatted as dictionary that maps from
constraint type to a list of constraints of that type.
}
\section{Slots}{

\describe{
\item{\code{zero}}{The dimension of the zero cone.}

\item{\code{nonneg}}{The dimension of the non-negative cone.}

\item{\code{exp}}{The number of 3-dimensional exponential cones.}

\item{\code{soc}}{A vector of the second-order cone dimensions.}

\item{\code{psd}}{A vector of the positive semidefinite cone dimensions, where the dimension of the PSD cone of k by k matrices is k.}
}}

