% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/affine.R
\docType{methods}
\name{Conj.Expression}
\alias{Conj.Expression}
\alias{Conj}
\alias{Conj,Expression-method}
\alias{Conjugate-class}
\alias{.Conjugate}
\alias{Conjugate}
\alias{to_numeric,Conjugate-method}
\alias{dim_from_args,Conjugate-method}
\alias{is_incr,Conjugate-method}
\alias{is_decr,Conjugate-method}
\alias{is_symmetric,Conjugate-method}
\alias{is_hermitian,Conjugate-method}
\title{Conjugate Expression}
\usage{
\method{Conj}{Expression}(z)

\S4method{Conj}{Expression}(z)

Conjugate(expr)

\S4method{to_numeric}{Conjugate}(object, values)

\S4method{dim_from_args}{Conjugate}(object)

\S4method{is_incr}{Conjugate}(object, idx)

\S4method{is_decr}{Conjugate}(object, idx)

\S4method{is_symmetric}{Conjugate}(object)

\S4method{is_hermitian}{Conjugate}(object)
}
\arguments{
\item{z}{An \linkS4class{Expression}}

\item{expr}{An \linkS4class{Expression} or R numeric data.}

\item{object}{A \linkS4class{Conjugate} object.}

\item{values}{A list of arguments to the atom.}

\item{idx}{An index into the atom.}
}
\value{
An \linkS4class{Expression} representing the conjugate.
}
\description{
The conjugate of the expression.

This class represents the complex conjugate of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(Conjugate)}: Elementwise complex conjugate of the constant.

\item \code{dim_from_args(Conjugate)}: The (row, col) dimensions of the expression.

\item \code{is_incr(Conjugate)}: Is the composition weakly increasing in argument idx?

\item \code{is_decr(Conjugate)}: Is the composition weakly decreasing in argument idx?

\item \code{is_symmetric(Conjugate)}: Is the expression symmetric?

\item \code{is_hermitian(Conjugate)}: Is the expression hermitian?

}}
\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} or R numeric data.}
}}

\examples{
z <- Constant(1 + 2*1i)
Conj(z)
}
