% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{Conjugate-class}
\alias{Conjugate-class}
\alias{.Conjugate}
\alias{Conjugate}
\alias{to_numeric,Conjugate-method}
\alias{dim_from_args,Conjugate-method}
\alias{is_incr,Conjugate-method}
\alias{is_decr,Conjugate-method}
\alias{is_symmetric,Conjugate-method}
\alias{is_hermitian,Conjugate-method}
\title{The Conjugate class.}
\usage{
Conjugate(expr)

\S4method{to_numeric}{Conjugate}(object, values)

\S4method{dim_from_args}{Conjugate}(object)

\S4method{is_incr}{Conjugate}(object, idx)

\S4method{is_decr}{Conjugate}(object, idx)

\S4method{is_symmetric}{Conjugate}(object)

\S4method{is_hermitian}{Conjugate}(object)
}
\arguments{
\item{expr}{An \linkS4class{Expression} or R numeric data.}

\item{object}{A \linkS4class{Conjugate} object.}

\item{values}{A list of arguments to the atom.}

\item{idx}{An index into the atom.}
}
\description{
This class represents the complex conjugate of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: Elementwise complex conjugate of the constant.

\item \code{dim_from_args}: The (row, col) dimensions of the expression.

\item \code{is_incr}: Is the composition weakly increasing in argument idx?

\item \code{is_decr}: Is the composition weakly decreasing in argument idx?

\item \code{is_symmetric}: Is the expression symmetric?

\item \code{is_hermitian}: Is the expression hermitian?
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} or R numeric data.}
}}

