% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{construct_solving_chain}
\alias{construct_solving_chain}
\title{Build a reduction chain from a problem to an installed solver.}
\usage{
construct_solving_chain(problem, candidates)
}
\arguments{
\item{problem}{The problem for which to build a chain.}

\item{candidates}{A list of candidate solvers.}
}
\value{
A \linkS4class{SolvingChain} that can be used to solve the problem.
}
\description{
Build a reduction chain from a problem to an installed solver.
}
