% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/058_atoms_affine_conv.R
\name{conv}
\alias{conv}
\title{1D discrete convolution}
\usage{
conv(a, b)
}
\arguments{
\item{a}{An Expression (vector, one must be constant)}

\item{b}{An Expression (vector)}
}
\value{
A Convolve atom
}
\description{
1D discrete convolution
}
