% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053_atoms_affine_cumsum.R
\name{cumsum_axis}
\alias{cumsum_axis}
\title{Cumulative sum along an axis}
\usage{
cumsum_axis(x, axis = NULL)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (across rows), or 2 (down columns)}
}
\value{
A Cumsum atom
}
\description{
Cumulative sum along an axis
}
