% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/299_zzz_R_specific_exports.R
\name{curvature}
\alias{curvature}
\title{Get Expression Curvature}
\usage{
curvature(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
Character: \code{"CONSTANT"}, \code{"AFFINE"}, \code{"CONVEX"}, \code{"CONCAVE"},
or \code{"UNKNOWN"}.
}
\description{
Returns the DCP curvature of an expression as a string.
}
\seealso{
\code{\link[=is_convex]{is_convex()}}, \code{\link[=is_concave]{is_concave()}}, \code{\link[=is_affine]{is_affine()}}, \code{\link[=is_constant]{is_constant()}}
}
