% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/092_atoms_cvar.R
\name{cvar}
\alias{cvar}
\title{Conditional Value at Risk (CVaR)}
\usage{
cvar(x, beta)
}
\arguments{
\item{x}{An Expression (vector)}

\item{beta}{Confidence level in [0, 1)}
}
\value{
An Expression representing the CVaR
}
\description{
CVaR at confidence level beta: average of the (1-beta) largest values.
}
