% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001_zzz_R_specific_CVXR-package.R
\docType{package}
\name{CVXR-package}
\alias{CVXR}
\alias{CVXR-package}
\title{CVXR: Disciplined Convex Optimization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An object-oriented modeling language for disciplined convex programming (DCP) as described in Fu, Narasimhan, and Boyd (2020, \doi{10.18637/jss.v094.i14}). It allows the user to formulate convex optimization problems in a natural way following mathematical convention and DCP rules. The system analyzes the problem, verifies its convexity, converts it into a canonical form, and hands it off to an appropriate solver to obtain the solution. This version uses the S7 object system for improved performance and maintainability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cvxr.rbind.io}
  \item \url{https://www.cvxgrp.org/CVXR/}
  \item Report bugs at \url{https://github.com/cvxgrp/CVXR/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stat.stanford.edu}

Authors:
\itemize{
  \item Anqi Fu \email{anqif@alumni.stanford.edu}
  \item Steven Diamond
  \item John Miller
}

Other contributors:
\itemize{
  \item Stephen Boyd \email{boyd@stanford.edu} [contributor]
}

}
\keyword{internal}
