% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/039_atoms_affine_promote.R
\name{cvxr_promote}
\alias{cvxr_promote}
\title{Promote a scalar expression to the given shape}
\usage{
cvxr_promote(expr, shape)
}
\arguments{
\item{expr}{An expression}

\item{shape}{Target shape}
}
\value{
The expression (unchanged if already the right shape) or a Promote atom
}
\description{
Promote a scalar expression to the given shape
}
\keyword{internal}
