% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{DiagMat-class}
\alias{DiagMat-class}
\alias{.DiagMat}
\alias{DiagMat}
\alias{to_numeric,DiagMat-method}
\alias{dim_from_args,DiagMat-method}
\alias{is_atom_log_log_convex,DiagMat-method}
\alias{is_atom_log_log_concave,DiagMat-method}
\alias{is_nonneg,DiagMat-method}
\alias{graph_implementation,DiagMat-method}
\title{The DiagMat class.}
\usage{
DiagMat(expr)

\S4method{to_numeric}{DiagMat}(object, values)

\S4method{dim_from_args}{DiagMat}(object)

\S4method{is_atom_log_log_convex}{DiagMat}(object)

\S4method{is_atom_log_log_concave}{DiagMat}(object)

\S4method{is_nonneg}{DiagMat}(object)

\S4method{graph_implementation}{DiagMat}(object, arg_objs, dim, data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing the matrix whose diagonal we are interested in.}

\item{object}{A \linkS4class{DiagMat} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the extraction of the diagonal from a square matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(DiagMat)}: Extract the diagonal from a square matrix constant.

\item \code{dim_from_args(DiagMat)}: The size of the atom.

\item \code{is_atom_log_log_convex(DiagMat)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(DiagMat)}: Is the atom log-log concave?

\item \code{is_nonneg(DiagMat)}: A logical value indicating whether the atom is nonnegative.

\item \code{graph_implementation(DiagMat)}: The graph implementation of the atom.

}}
\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing the matrix whose diagonal we are interested in.}
}}

