% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{DiagVec-class}
\alias{DiagVec-class}
\alias{.DiagVec}
\alias{DiagVec}
\alias{to_numeric,DiagVec-method}
\alias{size_from_args,DiagVec-method}
\alias{graph_implementation,DiagVec-method}
\title{The DiagVec class.}
\usage{
DiagVec(expr)

\S4method{to_numeric}{DiagVec}(object, values)

\S4method{size_from_args}{DiagVec}(object)

\S4method{graph_implementation}{DiagVec}(object, arg_objs, size,
  data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing the vector to convert.}

\item{object}{A \linkS4class{DiagVec} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the conversion of a vector into a diagonal matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: Convert the vector constant into a diagonal matrix.

\item \code{size_from_args}: The size of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing the vector to convert.}
}}

