% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/119_atoms_one_minus_pos.R
\name{diff_pos}
\alias{diff_pos}
\title{The difference x - y with domain x > y > 0}
\usage{
diff_pos(x, y)
}
\arguments{
\item{x}{An Expression (positive)}

\item{y}{An Expression (positive, elementwise less than x)}
}
\value{
A product expression
}
\description{
Equivalent to x * one_minus_pos(y / x).
}
