% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/119_atoms_dist_ratio.R
\name{dist_ratio}
\alias{dist_ratio}
\title{Distance ratio}
\usage{
dist_ratio(x, a, b)
}
\arguments{
\item{x}{A vector expression}

\item{a}{A numeric constant vector}

\item{b}{A numeric constant vector}
}
\value{
An expression representing the distance ratio
}
\description{
Computes norm(x - a)_2 / norm(x - b)_2, where a and b are constants.
This is a quasiconvex atom.
}
