% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_atoms_dotsort.R
\name{dotsort}
\alias{dotsort}
\title{Weighted sorted dot product}
\usage{
dotsort(X, W)
}
\arguments{
\item{X}{An Expression or numeric value.}

\item{W}{A constant numeric vector or matrix.}
}
\value{
A scalar convex Expression.
}
\description{
Computes \verb{<sort(vec(X)), sort(vec(W))>} where X is an expression and W is a
constant. A generalization of \code{sum_largest} and \code{sum_smallest}.
}
