% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/003_utilities_scopes.R
\name{dpp_scope_active}
\alias{dpp_scope_active}
\title{Check if DPP Scope is Active}
\usage{
dpp_scope_active()
}
\value{
Logical; \code{TRUE} if a \code{with_dpp_scope} block is active.
}
\description{
Check if DPP Scope is Active
}
\keyword{internal}
