% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/125_scopes.R
\name{dpp_scope_active}
\alias{dpp_scope_active}
\title{Returns TRUE if a \code{dpp_scope} is active}
\usage{
dpp_scope_active()
}
\value{
TRUE OR FALSE
}
\description{
Returns TRUE if a \code{dpp_scope} is active
}
