% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dual_value}
\alias{dual_value}
\alias{dual_value<-}
\title{Get and set the value of the dual variable in a constraint.}
\usage{
dual_value(object)

dual_value(object) <- value
}
\description{
Get and set the value of the dual variable in a constraint.
}
