% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/060_atoms_elementwise_entr.R
\name{entr}
\alias{entr}
\title{Create an entropy atom -x * log(x)}
\usage{
entr(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
An Entr atom
}
\description{
Create an entropy atom -x * log(x)
}
