% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{entr}
\alias{entr}
\alias{entropy}
\title{Entropy Function}
\usage{
entr(x)
}
\arguments{
\item{x}{An \linkS4class{Expression}, vector, or matrix.}
}
\value{
An \linkS4class{Expression} representing the entropy of the input.
}
\description{
The elementwise entropy function, \eqn{-xlog(x)}.
}
\examples{
x <- Variable(5)
obj <- Maximize(sum(entr(x)))
prob <- Problem(obj, list(sum(x) == 1))
result <- solve(prob)
result$getValue(x)
}
