% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/constraints.R
\docType{methods}
\name{==,Expression,Expression-method}
\alias{==,Expression,Expression-method}
\alias{==,Expression,ConstVal-method}
\alias{==,ConstVal,Expression-method}
\alias{EqConstraint-class}
\alias{.EqConstraint}
\title{The EqConstraint class}
\usage{
\S4method{==}{Expression,Expression}(e1, e2)

\S4method{==}{Expression,ConstVal}(e1, e2)

\S4method{==}{ConstVal,Expression}(e1, e2)
}
\arguments{
\item{e1, e2}{The \linkS4class{Expression} objects or numeric constants to compare.}
}
\description{
The EqConstraint class
}
