% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{ExpCone-class}
\alias{ExpCone-class}
\alias{.ExpCone}
\alias{ExpCone}
\alias{as.character,ExpCone-method}
\alias{residual,ExpCone-method}
\alias{size,ExpCone-method}
\alias{num_cones,ExpCone-method}
\alias{cone_sizes,ExpCone-method}
\alias{is_dcp,ExpCone-method}
\alias{is_dgp,ExpCone-method}
\alias{is_dqcp,ExpCone-method}
\alias{dual_value<-,ExpCone-method}
\title{The ExpCone class.}
\usage{
ExpCone(x, y, z, constr_id = NA_integer_)

\S4method{as.character}{ExpCone}(x)

\S4method{residual}{ExpCone}(object)

\S4method{size}{ExpCone}(object)

\S4method{num_cones}{ExpCone}(object)

\S4method{cone_sizes}{ExpCone}(object)

\S4method{is_dcp}{ExpCone}(object, dpp = FALSE)

\S4method{is_dgp}{ExpCone}(object, dpp = FALSE)

\S4method{is_dqcp}{ExpCone}(object)

\S4method{dual_value}{ExpCone}(object) <- value
}
\arguments{
\item{x}{The variable \eqn{x} in the exponential cone.}

\item{y}{The variable \eqn{y} in the exponential cone.}

\item{z}{The variable \eqn{z} in the exponential cone.}

\item{constr_id}{(Optional) A numeric value representing the constraint ID.}

\item{object}{A \linkS4class{ExpCone} object.}

\item{value}{A numeric scalar, vector, or matrix.}
}
\description{
This class represents a reformulated exponential cone constraint operating elementwise on \eqn{a, b, c}.
}
\details{
Original cone:
\deqn{
K = \{(x,y,z) | y > 0, ye^{x/y} \leq z\} \cup \{(x,y,z) | x \leq 0, y = 0, z \geq 0\}
}
Reformulated cone:
\deqn{
K = \{(x,y,z) | y, z > 0, y\log(y) + x \leq y\log(z)\} \cup \{(x,y,z) | x \leq 0, y = 0, z \geq 0\}
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{residual(ExpCone)}: The size of the \code{x} argument.

\item \code{size(ExpCone)}: The number of entries in the combined cones.

\item \code{num_cones(ExpCone)}: The number of elementwise cones.

\item \code{cone_sizes(ExpCone)}: The dimensions of the exponential cones.

\item \code{is_dcp(ExpCone)}: An exponential constraint is DCP if each argument is affine.

\item \code{is_dgp(ExpCone)}: Is the constraint DGP?

\item \code{is_dqcp(ExpCone)}: Is the constraint DQCP?

\item \code{dual_value(ExpCone) <- value}: Replaces the dual values of an exponential cone constraint.

}}
\section{Slots}{

\describe{
\item{\code{x}}{The variable \eqn{x} in the exponential cone.}

\item{\code{y}}{The variable \eqn{y} in the exponential cone.}

\item{\code{z}}{The variable \eqn{z} in the exponential cone.}
}}

