% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{expr_copy}
\alias{expr_copy}
\title{Shallow Copy of an Expression Tree Node}
\usage{
expr_copy(x, args = NULL, id_objects = NULL)
}
\arguments{
\item{x}{A canonicalizable object.}

\item{args}{Optional replacement args.}

\item{id_objects}{Optional identity map for deduplication.}
}
\value{
A copy of the object.
}
\description{
Shallow Copy of an Expression Tree Node
}
\keyword{internal}
