% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/120_atoms_eye_minus_inv.R
\name{eye_minus_inv}
\alias{eye_minus_inv}
\title{Unity resolvent (I - X) inverse for positive square matrix X}
\usage{
eye_minus_inv(X)
}
\arguments{
\item{X}{An Expression (positive square matrix with spectral radius < 1)}
}
\value{
An EyeMinusInv atom
}
\description{
Log-log convex atom for DGP. Solve with
\code{psolve(problem, gp = TRUE)}.
}
\examples{
X <- Variable(c(2, 2), pos = TRUE)
prob <- Problem(Minimize(sum(eye_minus_inv(X))), list(X <= 0.4))
\dontrun{psolve(prob, gp = TRUE, solver = "SCS")}
}
