% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/097_atoms_geo_mean.R
\name{geo_mean}
\alias{geo_mean}
\title{(Weighted) geometric mean of a vector}
\usage{
geo_mean(x, p = NULL, max_denom = 1024L, approx = TRUE)
}
\arguments{
\item{x}{An Expression (vector)}

\item{p}{Numeric weight vector (default: uniform)}

\item{max_denom}{Maximum denominator for rational approximation}

\item{approx}{If TRUE (default), use SOC approximation. If FALSE, use exact power cone.}
}
\value{
A GeoMean or GeoMeanApprox atom
}
\description{
(Weighted) geometric mean of a vector
}
