% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{get_data}
\alias{get_data}
\title{Get Expression Data}
\usage{
get_data(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A list containing data.
}
\description{
Get information needed to reconstruct the expression aside from its arguments.
}
