% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction_solvers.R
\name{get_dual_values}
\alias{get_dual_values}
\title{Gets the values of the dual variables.}
\usage{
get_dual_values(result_vec, parse_func, constraints)
}
\arguments{
\item{result_vec}{A vector containing the dual variable values.}

\item{parse_func}{Function handle for the parser.}

\item{constraints}{A list of the constraints in the problem.}
}
\value{
A map of constraint ID to dual variable value.
}
\description{
Gets the values of the dual variables.
}
