% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{get_problem_data}
\alias{get_problem_data}
\title{Get Problem Data for a Solver (deprecated)}
\usage{
get_problem_data(x, solver = NULL, ...)
}
\arguments{
\item{x}{A \code{\link{Problem}} object.}

\item{solver}{Character string naming solver, or \code{NULL} for
automatic selection.}

\item{...}{Additional arguments.}
}
\value{
A list with components \code{data}, \code{chain}, and
\code{inverse_data}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link{problem_data}} instead.
}
\seealso{
\code{\link{problem_data}}
}
