% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/267_zzz_R_specific_exports.R
\name{\%>>\%}
\alias{\%>>\%}
\title{Positive Semidefinite Constraint Operator}
\usage{
e1 \%>>\% e2
}
\arguments{
\item{e1, e2}{CVXR expressions or numeric matrices.}
}
\value{
A \code{PSD} constraint object.
}
\description{
Creates a PSD constraint: \code{e1 - e2} is positive semidefinite.
This is the R equivalent of Python's \code{A >> B}.
}
\examples{
\dontrun{
X <- Variable(3, 3, symmetric = TRUE)
constr <- X \%>>\% diag(3)  # X - I is PSD
}
}
\seealso{
\code{\link[=PSD]{PSD()}}
}
