% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/299_zzz_R_specific_exports.R
\name{\%<<\%}
\alias{\%<<\%}
\title{Negative Semidefinite Constraint Operator}
\usage{
e1 \%<<\% e2
}
\arguments{
\item{e1, e2}{CVXR expressions or numeric matrices.}
}
\value{
A \code{PSD} constraint object.
}
\description{
Creates an NSD constraint: \code{e2 - e1} is positive semidefinite,
i.e., \code{e1} is NSD relative to \code{e2}.
This is the R equivalent of Python's \code{A << B}.
}
\examples{
\dontrun{
X <- Variable(c(3, 3), symmetric = TRUE)
constr <- X \%<<\% diag(3)  # I - X is PSD (X is NSD relative to I)
}
}
\seealso{
\code{\link[=PSD]{PSD()}}
}
