% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{graph_implementation}
\alias{graph_implementation}
\title{Graph Implementation}
\usage{
graph_implementation(object, arg_objs, dim, data)
}
\arguments{
\item{object}{An \linkS4class{Expression} object.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\value{
A list of \code{list(LinOp for objective, list of constraints)}, where LinOp is a list representing the linear operator.
}
\description{
Reduces the atom to an affine expression and list of constraints.
}
