% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reductions.R
\name{group_constraints}
\alias{group_constraints}
\title{Organize the constraints into a list keyed by constraint names.}
\usage{
group_constraints(constraints)
}
\arguments{
\item{constraints}{A list of \linkS4class{Constraint} objects}
}
\value{
A list keyed by constraint types where list[[cone_type]] maps to a list of exactly those constraints that are of type cone_type.
}
\description{
Organize the constraints into a list keyed by constraint names.
}
