% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/081_atoms_harmonic_mean.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Harmonic mean: n / sum(1/x_i)}
\usage{
harmonic_mean(x)
}
\arguments{
\item{x}{An Expression (must be positive for DCP)}
}
\value{
An Expression representing the harmonic mean
}
\description{
Harmonic mean: n / sum(1/x_i)
}
