% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{has_quadratic_term}
\alias{has_quadratic_term}
\title{Check if Expression Has a Quadratic Term}
\usage{
has_quadratic_term(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression Has a Quadratic Term
}
\keyword{internal}
