% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{has_quadratic_term}
\alias{has_quadratic_term}
\title{Determine if an expression has a quadratic term}
\usage{
has_quadratic_term(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the expression has a quadratic term or not
}
\description{
Determine if an expression has a quadratic term
}
\examples{
y <- Variable(pos = TRUE)
has_quadratic_term(3*y + y^2)
}
