% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/053_atoms_affine_hstack.R
\name{hstack}
\alias{hstack}
\title{Horizontal concatenation of expressions}
\usage{
hstack(...)
}
\arguments{
\item{...}{Expressions (same number of rows)}
}
\value{
An HStack atom
}
\description{
Horizontal concatenation of expressions
}
