% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/061_atoms_elementwise_huber.R
\name{huber}
\alias{huber}
\title{Create a Huber loss atom}
\usage{
huber(x, M = 1)
}
\arguments{
\item{x}{An Expression}

\item{M}{Numeric threshold (default 1)}
}
\value{
A Huber atom
}
\description{
Create a Huber loss atom
}
