% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_zzz_R_specific_exports.R
\name{id}
\alias{id}
\title{Get Expression ID}
\usage{
id(x)
}
\arguments{
\item{x}{A CVXR expression, variable, parameter, or constraint.}
}
\value{
An integer.
}
\description{
Returns the unique integer identifier for a CVXR object.
}
