% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{id}
\alias{id}
\title{Identification Number}
\usage{
id(object)
}
\arguments{
\item{object}{A \linkS4class{Variable} or \linkS4class{Constraint} object.}
}
\value{
A non-negative integer identifier.
}
\description{
A unique identification number used internally to keep track of variables and constraints. Should not be modified by the user.
}
\examples{
x <- Variable()
constr <- (x >= 5)
id(x)
id(constr)
}
\seealso{
\code{\link[CVXR]{get_id}} \code{\link[CVXR]{setIdCounter}}
}
