% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/077_atoms_elementwise_logic.R
\name{iff}
\alias{iff}
\title{Logical Biconditional}
\usage{
iff(x, y)
}
\arguments{
\item{x, y}{Boolean \link{Variable}s or logic expressions.}
}
\value{
A \link{Not} expression wrapping \link{Xor}.
}
\description{
Logical biconditional: x <=> y.
Returns 1 if and only if x and y have the same value.
Equivalent to \code{Not(Xor(x, y))}.
}
\examples{
\dontrun{
x <- Variable(boolean = TRUE)
y <- Variable(boolean = TRUE)
expr <- iff(x, y)
}
}
\seealso{
\code{\link[=implies]{implies()}}, \code{\link[=Not]{Not()}}, \code{\link[=And]{And()}}, \code{\link[=Or]{Or()}}, \code{\link[=Xor]{Xor()}}
}
