% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{import_solver}
\alias{import_solver}
\title{Import Solver}
\usage{
import_solver(solver)
}
\arguments{
\item{solver}{A \linkS4class{Solver} object.}
}
\description{
Import the R library that interfaces with the specified solver.
}
\examples{
import_solver(ECOS())
import_solver(SCS())
}
