% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R, R/constraints.R
\docType{class}
\name{<=,Expression,Expression-method}
\alias{<=,Expression,Expression-method}
\alias{<=,Expression,ConstVal-method}
\alias{<=,ConstVal,Expression-method}
\alias{<,Expression,Expression-method}
\alias{<,Expression,ConstVal-method}
\alias{<,ConstVal,Expression-method}
\alias{>=,Expression,Expression-method}
\alias{>=,Expression,ConstVal-method}
\alias{>=,ConstVal,Expression-method}
\alias{>,Expression,Expression-method}
\alias{>,Expression,ConstVal-method}
\alias{>,ConstVal,Expression-method}
\alias{IneqConstraint-class}
\alias{.IneqConstraint}
\alias{name,IneqConstraint-method}
\alias{dim,IneqConstraint-method}
\alias{size,IneqConstraint-method}
\alias{expr,IneqConstraint-method}
\alias{is_dcp,IneqConstraint-method}
\alias{is_dgp,IneqConstraint-method}
\alias{residual,IneqConstraint-method}
\title{The IneqConstraint class}
\usage{
\S4method{<=}{Expression,Expression}(e1, e2)

\S4method{<=}{Expression,ConstVal}(e1, e2)

\S4method{<=}{ConstVal,Expression}(e1, e2)

\S4method{<}{Expression,Expression}(e1, e2)

\S4method{<}{Expression,ConstVal}(e1, e2)

\S4method{<}{ConstVal,Expression}(e1, e2)

\S4method{>=}{Expression,Expression}(e1, e2)

\S4method{>=}{Expression,ConstVal}(e1, e2)

\S4method{>=}{ConstVal,Expression}(e1, e2)

\S4method{>}{Expression,Expression}(e1, e2)

\S4method{>}{Expression,ConstVal}(e1, e2)

\S4method{>}{ConstVal,Expression}(e1, e2)

\S4method{name}{IneqConstraint}(x)

\S4method{dim}{IneqConstraint}(x)

\S4method{size}{IneqConstraint}(object)

\S4method{expr}{IneqConstraint}(object)

\S4method{is_dcp}{IneqConstraint}(object)

\S4method{is_dgp}{IneqConstraint}(object)

\S4method{residual}{IneqConstraint}(object)
}
\arguments{
\item{e1, e2}{The \linkS4class{Expression} objects or numeric constants to compare.}

\item{x, object}{A \linkS4class{IneqConstraint} object.}
}
\description{
The IneqConstraint class
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: The string representation of the constraint.

\item \code{dim}: The dimensions of the constrained expression.

\item \code{size}: The size of the constrained expression.

\item \code{expr}: The expression to constrain.

\item \code{is_dcp}: A non-positive constraint is DCP if its argument is convex.

\item \code{is_dgp}: Is the constraint DGP?

\item \code{residual}: The residual of the constraint.
}}

