% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/259_zzz_R_specific_exports.R
\name{installed_solvers}
\alias{installed_solvers}
\title{List installed solvers}
\usage{
installed_solvers()
}
\value{
A character vector of solver names.
}
\description{
Returns the names of solvers whose R packages are available.
}
