% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/303_defines.R
\name{add_to_solver_blacklist}
\alias{add_to_solver_blacklist}
\alias{remove_from_solver_blacklist}
\alias{set_solver_blacklist}
\alias{installed_solvers}
\title{List installed solvers, taking blacklisted solvers into account}
\usage{
add_to_solver_blacklist(solvers)

remove_from_solver_blacklist(solvers)

set_solver_blacklist(solvers)

installed_solvers()
}
\arguments{
\item{solvers}{a character vector of solver names}
}
\value{
The names of all the installed solvers as a character vector
}
\description{
List installed solvers, taking blacklisted solvers into account
}
\section{Functions}{
\itemize{
\item \code{add_to_solver_blacklist()}: Add to solver blacklist, useful for temporarily disabling a solver

\item \code{remove_from_solver_blacklist()}: Remove solvers from blacklist

\item \code{set_solver_blacklist()}: Set solver blacklist to a value

}}
