% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/024_interface_matrix_utilities.R
\name{intf_is_psd}
\alias{intf_is_psd}
\title{Check if a symmetric matrix is PSD within tolerance}
\usage{
intf_is_psd(val, tol = EIGVAL_TOL)
}
\arguments{
\item{val}{A symmetric numeric matrix}

\item{tol}{Eigenvalue tolerance}
}
\value{
Logical
}
\description{
Check if a symmetric matrix is PSD within tolerance
}
\keyword{internal}
