% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/023_interface_matrix_utilities.R
\name{intf_is_sparse}
\alias{intf_is_sparse}
\title{Check if a value is a sparse matrix}
\usage{
intf_is_sparse(val)
}
\arguments{
\item{val}{A value}
}
\value{
Logical
}
\description{
Check if a value is a sparse matrix
}
\keyword{internal}
