% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/019_interface_matrix_utilities.R
\name{intf_shape}
\alias{intf_shape}
\title{Get the shape of a value as an integer vector c(nrow, ncol)}
\usage{
intf_shape(val)
}
\arguments{
\item{val}{A numeric scalar, vector, matrix, or Matrix object}
}
\value{
Integer vector of length 2
}
\description{
Get the shape of a value as an integer vector c(nrow, ncol)
}
\keyword{internal}
