% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/306_exports.R
\docType{methods}
\name{inv_pos}
\alias{inv_pos}
\title{Reciprocal Function}
\usage{
inv_pos(x)
}
\arguments{
\item{x}{An \linkS4class{Expression}, vector, or matrix.}
}
\value{
An \linkS4class{Expression} representing the reciprocal of the input.
}
\description{
The elementwise reciprocal function, \eqn{\frac{1}{x}}
}
\examples{
A <- Variable(2,2)
val <- cbind(c(1,2), c(3,4))
prob <- Problem(Minimize(inv_pos(A)[1,2]), list(A == val))
result <- solve(prob)
result$value
}
