% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_affine}
\alias{is_affine}
\title{Check if an Expression is Affine}
\usage{
is_affine(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if an Expression is Affine
}
