% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{is_atom_quasiconcave}
\alias{is_atom_quasiconcave}
\title{Determine if an atom is quasiconcave}
\usage{
is_atom_quasiconcave(object)
}
\arguments{
\item{object}{A \linkS4class{Atom} object.}
}
\value{
A logical value indicating whether the expression is quasiconcave or not
}
\description{
Determine if an atom is quasiconcave
}
