% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{is_atom_quasiconvex}
\alias{is_atom_quasiconvex}
\title{Determine if an atom is quasiconvex}
\usage{
is_atom_quasiconvex(object)
}
\arguments{
\item{object}{A \linkS4class{Atom} object.}
}
\value{
A logical value indicating whether the expression is quasiconvex or not
}
\description{
Determine if an atom is quasiconvex
}
