% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_concave}
\alias{is_concave}
\title{Check if an Expression is Concave}
\usage{
is_concave(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if an Expression is Concave
}
