% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_convex}
\alias{is_convex}
\title{Check if an Expression is Convex}
\usage{
is_convex(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if an Expression is Convex
}
