% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_dgp}
\alias{is_dgp}
\title{Check if a Constraint is DGP-Compliant}
\usage{
is_dgp(x, ...)
}
\arguments{
\item{x}{A constraint object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if a Constraint is DGP-Compliant
}
