% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_dpp}
\alias{is_dpp}
\title{Check DPP Compliance}
\usage{
is_dpp(x, ...)
}
\arguments{
\item{x}{An expression, constraint, or problem object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Determines whether an expression or problem satisfies the rules of
Disciplined Parameterized Programming (DPP). A DPP-compliant problem
enables caching the compilation across parameter value changes.
}
