% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_dpp}
\alias{is_dpp}
\title{DPP Compliance}
\usage{
is_dpp(object, context = c("dcp", "dgp"))
}
\arguments{
\item{object}{A \linkS4class{Problem} or \linkS4class{Expression} object.}

\item{context}{should be one of 'dcp' (default) or 'dgp'}
}
\value{
A logical value indicating whether the problem or expression is DPP compliant, i.e. uses parameters
}
\description{
Determine if a problem or expression complies with the disciplined parametric programming rules.
}
\examples{
p <- Parameter(pos = TRUE, value = 5)
y <- Variable(pos = TRUE)
prob <- Problem(Minimize(p*y), list(x >= 5, y >= 5))
is_dpp(prob)
}
