% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_dqcp}
\alias{is_dqcp}
\title{DQCP Compliance}
\usage{
is_dqcp(object)
}
\arguments{
\item{object}{A \linkS4class{Problem} or \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the problem or expression is DQCP compliant
}
\description{
Determine if a problem or expression complies with the disciplined quasiconvex programming rules.
}
\examples{
x <- Variable(pos = TRUE)
y <- Variable(pos = TRUE)
prob <- Problem(Minimize(x*y), list(x >= 5, y >= 5))
is_dqcp(prob)
}
