% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_incr}
\alias{is_incr}
\title{Check if Atom is Increasing in an Argument}
\usage{
is_incr(x, idx, ...)
}
\arguments{
\item{x}{An atom object.}

\item{idx}{Integer: argument index (1-based, R convention).}

\item{...}{Additional arguments.}
}
\value{
Logical scalar.
}
\description{
Check if Atom is Increasing in an Argument
}
\keyword{internal}
