% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/267_zzz_R_specific_exports.R
\name{is_matrix}
\alias{is_matrix}
\title{Is the Expression a Matrix?}
\usage{
is_matrix(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
Logical.
}
\description{
Returns \code{TRUE} if the expression has both dimensions greater than 1.
}
\seealso{
\code{\link[=size]{size()}}, \code{\link[=is_scalar]{is_scalar()}}, \code{\link[=is_vector]{is_vector()}}
}
