% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_mixed_integer}
\alias{is_mixed_integer}
\title{Is Problem Mixed Integer?}
\usage{
is_mixed_integer(object)
}
\arguments{
\item{object}{A \linkS4class{Problem} object.}
}
\value{
A logical value indicating whether the problem is a mixed-integer program
}
\description{
Determine if a problem is a mixed-integer program.
}
