% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_nsd}
\alias{is_nsd}
\title{Check if Expression is Negative Semidefinite}
\usage{
is_nsd(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Negative Semidefinite
}
