% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_pos}
\alias{is_pos}
\title{Check if Expression is Strictly Positive}
\usage{
is_pos(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Strictly Positive
}
\keyword{internal}
