% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{is_qp}
\alias{is_qp}
\title{Is Problem a QP?}
\usage{
is_qp(object)
}
\arguments{
\item{object}{A \linkS4class{Problem} object.}
}
\value{
A logical value indicating whether the problem is a quadratic program.
}
\description{
Determine if a problem is a quadratic program.
}
