% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_quasiconcave}
\alias{is_quasiconcave}
\title{Check if Expression is Quasiconcave}
\usage{
is_quasiconcave(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Quasiconcave
}
