% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_quasiconcave}
\alias{is_quasiconcave}
\title{Quasiconcave?}
\usage{
is_quasiconcave(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the expression is quasiconcave or not
}
\description{
Determine if an expression is quasiconcave
}
\examples{
y <- Variable(pos = TRUE)
is_quasiconcave(y)
}
