% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{is_quasiconvex}
\alias{is_quasiconvex}
\title{Check if Expression is Quasiconvex}
\usage{
is_quasiconvex(x, ...)
}
\arguments{
\item{x}{An expression object.}

\item{...}{Not used.}
}
\value{
Logical scalar.
}
\description{
Check if Expression is Quasiconvex
}
