% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_quasiconvex}
\alias{is_quasiconvex}
\title{Quasiconvex?}
\usage{
is_quasiconvex(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the expression is quasiconvex or not
}
\description{
Determine if an expression is quasiconvex
}
\examples{
y <- Variable(pos = TRUE)
is_quasiconvex(y)
}
