% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_quasilinear}
\alias{is_quasilinear}
\title{Quasilinear?}
\usage{
is_quasilinear(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the expression is quasilinear or not
}
\description{
Determine if an expression is quasilinear
}
\examples{
y <- Variable(pos = TRUE)
is_quasilinear(y)
}
