% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/297_zzz_R_specific_exports.R
\name{is_scalar}
\alias{is_scalar}
\title{Is the Expression a Scalar?}
\usage{
is_scalar(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
Logical.
}
\description{
Is the Expression a Scalar?
}
\seealso{
\code{\link[=size]{size()}}, \code{\link[=is_vector]{is_vector()}}, \code{\link[=is_matrix]{is_matrix()}}
}
