% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R
\docType{methods}
\name{is_skew_symmetric}
\alias{is_skew_symmetric}
\title{Determine if an expression is skew symmetric}
\usage{
is_skew_symmetric(object)
}
\arguments{
\item{object}{A \linkS4class{Expression} object.}
}
\value{
A logical value indicating whether the expression is skew symmetric or not
}
\description{
Determine if an expression is skew symmetric
}
\examples{
m <- matrix(0, nrow = 3, ncol = 3)
is_skew_symmetric(m)
}
