% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/267_zzz_R_specific_exports.R
\name{is_vector}
\alias{is_vector}
\title{Is the Expression a Vector?}
\usage{
is_vector(x)
}
\arguments{
\item{x}{A CVXR expression.}
}
\value{
Logical.
}
\description{
Returns \code{TRUE} if the expression has at most one dimension greater than 1.
}
\seealso{
\code{\link[=size]{size()}}, \code{\link[=is_scalar]{is_scalar()}}, \code{\link[=is_matrix]{is_matrix()}}
}
