% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_atoms_lambda_max.R
\name{lambda_max}
\alias{lambda_max}
\title{Maximum eigenvalue}
\usage{
lambda_max(A)
}
\arguments{
\item{A}{A square matrix expression}
}
\value{
An expression representing the maximum eigenvalue of A
}
\description{
Maximum eigenvalue
}
