% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_atoms_lambda_sum_largest.R
\name{lambda_sum_largest}
\alias{lambda_sum_largest}
\title{Sum of largest k eigenvalues}
\usage{
lambda_sum_largest(A, k)
}
\arguments{
\item{A}{A square matrix expression}

\item{k}{Number of largest eigenvalues to sum (positive integer)}
}
\value{
An expression representing the sum of the k largest eigenvalues
}
\description{
Sum of largest k eigenvalues
}
