% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/305_exports.R
\docType{methods}
\name{lambda_sum_largest}
\alias{lambda_sum_largest}
\title{Sum of Largest Eigenvalues}
\usage{
lambda_sum_largest(A, k)
}
\arguments{
\item{A}{An \linkS4class{Expression} or matrix.}

\item{k}{The number of eigenvalues to sum over.}
}
\value{
An \linkS4class{Expression} representing the sum of the largest \code{k} eigenvalues of the input.
}
\description{
The sum of the largest \eqn{k} eigenvalues of a matrix.
}
\examples{
C <- Variable(3,3)
val <- cbind(c(1,2,3), c(2,4,5), c(3,5,6))
prob <- Problem(Minimize(lambda_sum_largest(C,2)), list(C == val))
result <- solve(prob)
result$value
result$getValue(C)
}
