% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_atoms_lambda_sum_largest.R
\name{lambda_sum_smallest}
\alias{lambda_sum_smallest}
\title{Sum of smallest k eigenvalues}
\usage{
lambda_sum_smallest(A, k)
}
\arguments{
\item{A}{A square matrix expression}

\item{k}{Number of smallest eigenvalues to sum (positive integer)}
}
\value{
An expression representing the sum of the k smallest eigenvalues
}
\description{
Sum of smallest k eigenvalues
}
