% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/116_atoms_length.R
\name{length_expr}
\alias{length_expr}
\title{Length of a Vector (Last Nonzero Index)}
\usage{
length_expr(x)
}
\arguments{
\item{x}{A CVXR vector expression.}
}
\value{
A \code{Length} expression (scalar).
}
\description{
Returns the index of the last nonzero element of a vector (1-based).
This atom is quasiconvex but NOT convex, so it can only be used in
DQCP problems (solved with \code{qcp = TRUE}).
}
\seealso{
\code{\link{ceil_expr}}, \code{\link{floor_expr}}
}
