% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/016_lin_ops_LinOp.R
\name{linop_args_push_back}
\alias{linop_args_push_back}
\title{Append a child LinOp to the args list}
\usage{
linop_args_push_back(ptr, child_ptr)
}
\arguments{
\item{ptr}{External pointer to the parent LinOp.}

\item{child_ptr}{External pointer to the child LinOp.}
}
\description{
Append a child LinOp to the args list
}
\keyword{internal}
