% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/016_lin_ops_LinOp.R
\name{linop_new}
\alias{linop_new}
\title{Create a new C++ LinOp external pointer}
\usage{
linop_new()
}
\value{
An external pointer to a C++ LinOp object.
}
\description{
Create a new C++ LinOp external pointer
}
\keyword{internal}
