% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/013_lin_ops_LinOp.R
\name{linop_set_data_ndim}
\alias{linop_set_data_ndim}
\title{Set the data dimensionality on a LinOp}
\usage{
linop_set_data_ndim(ptr, value)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{value}{Integer dimensionality (0 = scalar, 2 = matrix).}
}
\description{
Set the data dimensionality on a LinOp
}
\keyword{internal}
