% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011_lin_ops_LinOp.R
\name{linop_set_dense_data}
\alias{linop_set_dense_data}
\title{Set dense data on a LinOp}
\usage{
linop_set_dense_data(ptr, value)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{value}{Dense matrix data.}
}
\description{
Set dense data on a LinOp
}
\keyword{internal}
