% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/016_lin_ops_LinOp.R
\name{linop_set_linop_data}
\alias{linop_set_linop_data}
\title{Set a LinOp data sub-tree on a LinOp}
\usage{
linop_set_linop_data(ptr, data_ptr)
}
\arguments{
\item{ptr}{External pointer to the LinOp.}

\item{data_ptr}{External pointer to the data LinOp sub-tree.}
}
\description{
Set a LinOp data sub-tree on a LinOp
}
\keyword{internal}
