% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011_lin_ops_LinOp.R
\name{linop_set_sparse_data}
\alias{linop_set_sparse_data}
\title{Set sparse data on a LinOp}
\usage{
linop_set_sparse_data(ptr, value)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{value}{Sparse matrix data (dgCMatrix or similar).}
}
\description{
Set sparse data on a LinOp
}
\keyword{internal}
