% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/016_lin_ops_LinOp.R
\name{linop_set_type}
\alias{linop_set_type}
\title{Set the type of a LinOp}
\usage{
linop_set_type(ptr, type)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{type}{Character string -- one of \code{.LINOP_TYPES}.}
}
\description{
Set the type of a LinOp
}
\keyword{internal}
