% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/015_lin_ops_LinOp.R
\name{linop_size_push_back}
\alias{linop_size_push_back}
\title{Push a dimension to the LinOp size vector}
\usage{
linop_size_push_back(ptr, value)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{value}{Integer dimension value.}
}
\description{
Push a dimension to the LinOp size vector
}
\keyword{internal}
