% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/011_lin_ops_LinOp.R
\name{linop_slice_push_back}
\alias{linop_slice_push_back}
\title{Push a slice (index vector) to a LinOp}
\usage{
linop_slice_push_back(ptr, int_vector)
}
\arguments{
\item{ptr}{External pointer to a LinOp.}

\item{int_vector}{Integer vector of indices (0-based for C++).}
}
\description{
Push a slice (index vector) to a LinOp
}
\keyword{internal}
