% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\docType{class}
\name{Log1p-class}
\alias{Log1p-class}
\alias{.Log1p}
\alias{Log1p}
\alias{to_numeric,Log1p-method}
\alias{sign_from_args,Log1p-method}
\alias{graph_implementation,Log1p-method}
\title{The Log1p class.}
\usage{
Log1p(x)

\S4method{to_numeric}{Log1p}(object, values)

\S4method{sign_from_args}{Log1p}(object)

\S4method{graph_implementation}{Log1p}(object, arg_objs, size, data = NA_real_)
}
\arguments{
\item{x}{An \linkS4class{Expression} or numeric constant.}

\item{object}{A \linkS4class{Log1p} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the elementwise operation \eqn{\log(1 + x)}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: The elementwise natural logarithm of one plus the input value.

\item \code{sign_from_args}: The sign of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} or numeric constant.}
}}

