% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/072_atoms_elementwise_log1p.R,
%   R/302_zzz_R_specific_aliases.R
\name{log1p_atom}
\alias{log1p_atom}
\alias{log1p_expr}
\title{Log(1 + x) -- elementwise}
\usage{
log1p_atom(x)

log1p_expr(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A Log1p atom
}
\description{
Log(1 + x) -- elementwise
}
