% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/066_atoms_elementwise_log1p.R
\name{log1p_atom}
\alias{log1p_atom}
\title{Log(1 + x) — elementwise}
\usage{
log1p_atom(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A Log1p atom
}
\description{
Log(1 + x) — elementwise
}
