% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/106_atoms_log_det.R
\name{log_det}
\alias{log_det}
\title{Log-determinant}
\usage{
log_det(A)
}
\arguments{
\item{A}{A square PSD matrix expression}
}
\value{
An expression representing log(det(A))
}
\description{
Computes log(det(A)) for PSD matrix A.
}
