% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/081_atoms_elementwise_log_normcdf.R
\name{log_normcdf}
\alias{log_normcdf}
\title{Elementwise log of the standard normal CDF}
\usage{
log_normcdf(x)
}
\arguments{
\item{x}{An Expression or numeric value.}
}
\value{
A concave Expression representing log(Phi(x)).
}
\description{
Quadratic approximation of \code{log(pnorm(x))} with modest accuracy over the range -4 to 4.
}
