% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/098_atoms_log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Log-sum-exp: log(sum(exp(x)))}
\usage{
log_sum_exp(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical: keep reduced dimensions?}
}
\value{
A LogSumExp atom
}
\description{
Log-sum-exp: log(sum(exp(x)))
}
