% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/079_atoms_elementwise_loggamma.R
\name{loggamma}
\alias{loggamma}
\title{Elementwise log of the gamma function}
\usage{
loggamma(x)
}
\arguments{
\item{x}{An Expression or numeric value (must be positive for DCP).}
}
\value{
A convex Expression representing log(gamma(x)).
}
\description{
Piecewise linear approximation of \code{log(gamma(x))}.
Has modest accuracy over the full range, approaching perfect accuracy
as x goes to infinity.
}
