% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/063_loggamma.R
\name{LogGamma}
\alias{LogGamma}
\title{The LogGamma atom.}
\usage{
LogGamma(x)
}
\arguments{
\item{x}{An \linkS4class{Expression} object.}
}
\value{
An expression representing the log-gamma of \code{x}.
}
\description{
The elementwise log of the gamma function.
}
\details{
This implementation has modest accuracy over the full range, approaching perfect
accuracy as \eqn{x} approaches infinity. For details on the nature of the approximation,
return to CVXPY GitHub Issue #228 https://github.com/cvxpy/cvxpy/issues/228#issuecomment-544281906.
}
