% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/309_sparse_utils.R
\name{make_sparse_diagonal_matrix}
\alias{make_sparse_diagonal_matrix}
\title{Make a CSC sparse diagonal matrix}
\usage{
make_sparse_diagonal_matrix(size, diagonal = NULL)
}
\arguments{
\item{size}{number of rows or columns}

\item{diagonal}{if specified, the diagonal values, in which case size is ignored}
}
\value{
a compressed sparse column diagonal matrix
}
\description{
Make a CSC sparse diagonal matrix
}
