% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/111_atoms_matrix_frac.R
\name{matrix_frac}
\alias{matrix_frac}
\title{Matrix fractional function}
\usage{
matrix_frac(X, P)
}
\arguments{
\item{X}{A matrix expression (n by m)}

\item{P}{A square matrix expression (n by n), must be PSD}
}
\value{
An expression representing \eqn{\mathrm{trace}(X^T P^{-1} X)}
}
\description{
Computes \eqn{\mathrm{trace}(X^T P^{-1} X)}. If P is a constant matrix, uses
a QuadForm shortcut for efficiency.
}
