% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/045_atoms_affine_trace.R
\name{matrix_trace}
\alias{matrix_trace}
\title{Trace of a square matrix expression}
\usage{
matrix_trace(x)
}
\arguments{
\item{x}{An Expression (square matrix)}
}
\value{
A Trace atom (scalar)
}
\description{
Trace of a square matrix expression
}
