% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/306_exports.R
\docType{methods}
\name{matrix_trace}
\alias{matrix_trace}
\alias{trace}
\alias{tr}
\title{Matrix Trace}
\usage{
matrix_trace(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} or matrix.}
}
\value{
An \linkS4class{Expression} representing the trace of the input.
}
\description{
The sum of the diagonal entries in a matrix.
}
\examples{
C <- Variable(3,3)
val <- cbind(3:5, 6:8, 9:11)
prob <- Problem(Maximize(matrix_trace(C)), list(C == val))
result <- solve(prob)
result$value
}
