% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/067_atoms_elementwise_maximum.R
\name{max_elemwise}
\alias{max_elemwise}
\title{Elementwise maximum of expressions}
\usage{
max_elemwise(...)
}
\arguments{
\item{...}{Expressions (at least 2)}
}
\value{
A Maximum atom
}
\description{
Elementwise maximum of expressions
}
