% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{max_elemwise}
\alias{max_elemwise}
\title{Elementwise Maximum}
\usage{
max_elemwise(arg1, arg2, ...)
}
\arguments{
\item{arg1}{An \linkS4class{Expression}, vector, or matrix.}

\item{arg2}{An \linkS4class{Expression}, vector, or matrix.}

\item{...}{Additional \linkS4class{Expression} objects, vectors, or matrices.}
}
\value{
An \linkS4class{Expression} representing the elementwise maximum of the inputs.
}
\description{
The elementwise maximum.
}
\examples{
c <- matrix(c(1,-1))
prob <- Problem(Minimize(max_elemwise(t(c), 2, 2 + t(c))[2]))
result <- solve(prob)
result$value
}
