% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/091_atoms_max.R
\name{max_entries}
\alias{max_entries}
\title{Maximum entry of an expression}
\usage{
max_entries(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical}
}
\value{
A MaxEntries atom
}
\description{
Maximum entry of an expression
}
