% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_problems_objective.R
\name{Maximize}
\alias{Maximize}
\title{Create a Maximization Objective}
\usage{
Maximize(expr)
}
\arguments{
\item{expr}{A CVXR expression or numeric value to maximize.}
}
\value{
A \code{Maximize} object.
}
\description{
Specifies that the objective expression should be maximized. The expression
must be concave and scalar for a DCP-compliant problem.
}
\examples{
x <- Variable()
obj <- Maximize(-x^2 + 1)

}
\seealso{
\code{\link{Minimize}}, \code{\link{Problem}}
}
