% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069_atoms_elementwise_minimum.R
\name{min_elemwise}
\alias{min_elemwise}
\title{Elementwise minimum of expressions}
\usage{
min_elemwise(...)
}
\arguments{
\item{...}{Expressions (at least 2)}
}
\value{
A Minimum atom
}
\description{
Elementwise minimum of expressions
}
