% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\docType{methods}
\name{min_elemwise}
\alias{min_elemwise}
\title{Elementwise Minimum}
\usage{
min_elemwise(arg1, arg2, ...)
}
\arguments{
\item{arg1}{An \linkS4class{Expression}, vector, or matrix.}

\item{arg2}{An \linkS4class{Expression}, vector, or matrix.}

\item{...}{Additional \linkS4class{Expression} objects, vectors, or matrices.}
}
\value{
An \linkS4class{Expression} representing the elementwise minimum of the inputs.
}
\description{
The elementwise minimum.
}
\examples{
a <- cbind(c(-5,2), c(-3,-1))
b <- cbind(c(5,4), c(-1,2))
prob <- Problem(Minimize(min_elemwise(a, 0, b)[1,2]))
result <- solve(prob)
result$value
}
