% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/100_atoms_min.R
\name{min_entries}
\alias{min_entries}
\title{Minimum entry of an expression}
\usage{
min_entries(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical}
}
\value{
A MinEntries atom
}
\description{
Minimum entry of an expression
}
