% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{MinEntries-class}
\alias{MinEntries-class}
\alias{.MinEntries}
\alias{MinEntries}
\alias{to_numeric,MinEntries-method}
\alias{sign_from_args,MinEntries-method}
\alias{is_atom_convex,MinEntries-method}
\alias{is_atom_concave,MinEntries-method}
\alias{is_atom_log_log_convex,MinEntries-method}
\alias{is_atom_log_log_concave,MinEntries-method}
\alias{is_incr,MinEntries-method}
\alias{is_decr,MinEntries-method}
\alias{is_pwl,MinEntries-method}
\alias{.grad,MinEntries-method}
\alias{.column_grad,MinEntries-method}
\title{The MinEntries class.}
\usage{
MinEntries(x, axis = NA_real_, keepdims = FALSE)

\S4method{to_numeric}{MinEntries}(object, values)

\S4method{sign_from_args}{MinEntries}(object)

\S4method{is_atom_convex}{MinEntries}(object)

\S4method{is_atom_concave}{MinEntries}(object)

\S4method{is_atom_log_log_convex}{MinEntries}(object)

\S4method{is_atom_log_log_concave}{MinEntries}(object)

\S4method{is_incr}{MinEntries}(object, idx)

\S4method{is_decr}{MinEntries}(object, idx)

\S4method{is_pwl}{MinEntries}(object)

\S4method{.grad}{MinEntries}(object, values)

\S4method{.column_grad}{MinEntries}(object, value)
}
\arguments{
\item{x}{An \linkS4class{Expression} representing a vector or matrix.}

\item{axis}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}

\item{keepdims}{(Optional) Should dimensions be maintained when applying the atom along an axis? If \code{FALSE}, result will be collapsed into an \eqn{n x 1} column vector. The default is \code{FALSE}.}

\item{object}{A \linkS4class{MinEntries} object.}

\item{values}{A list of numeric values for the arguments}

\item{idx}{An index into the atom.}

\item{value}{A numeric value}
}
\description{
The minimum of an expression.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(MinEntries)}: The largest entry in \code{x}.

\item \code{sign_from_args(MinEntries)}: The sign of the atom.

\item \code{is_atom_convex(MinEntries)}: The atom is not convex.

\item \code{is_atom_concave(MinEntries)}: The atom is concave.

\item \code{is_atom_log_log_convex(MinEntries)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(MinEntries)}: Is the atom log-log concave?

\item \code{is_incr(MinEntries)}: The atom is weakly increasing in every argument.

\item \code{is_decr(MinEntries)}: The atom is not weakly decreasing in any argument.

\item \code{is_pwl(MinEntries)}: Is \code{x} piecewise linear?

\item \code{.grad(MinEntries)}: Gives the (sub/super)gradient of the atom w.r.t. each variable

\item \code{.column_grad(MinEntries)}: Gives the (sub/super)gradient of the atom w.r.t. each column variable

}}
\section{Slots}{

\describe{
\item{\code{x}}{An \linkS4class{Expression} representing a vector or matrix.}

\item{\code{axis}}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}

\item{\code{keepdims}}{(Optional) Should dimensions be maintained when applying the atom along an axis? If \code{FALSE}, result will be collapsed into an \eqn{n x 1} column vector. The default is \code{FALSE}.}
}}

