% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\docType{class}
\name{Minimize-class}
\alias{Minimize-class}
\alias{.Minimize}
\alias{Minimize}
\alias{Minimize}
\alias{canonicalize,Minimize-method}
\alias{variables,Minimize-method}
\alias{parameters,Minimize-method}
\alias{constants,Minimize-method}
\alias{is_dcp,Minimize-method}
\alias{value,Minimize-method}
\title{The Minimize class.}
\usage{
Minimize(expr)

\S4method{canonicalize}{Minimize}(object)

\S4method{variables}{Minimize}(object)

\S4method{parameters}{Minimize}(object)

\S4method{constants}{Minimize}(object)

\S4method{is_dcp}{Minimize}(object)

\S4method{value}{Minimize}(object)
}
\arguments{
\item{expr}{A scalar \linkS4class{Expression} to minimize.}

\item{object}{A \linkS4class{Minimize} object.}
}
\description{
This class represents an optimization objective for minimization.
}
\section{Methods (by generic)}{
\itemize{
\item \code{canonicalize}: Pass on the target expression's objective and constraints.

\item \code{variables}: List of \linkS4class{Variable} objects in the objective.

\item \code{parameters}: List of \linkS4class{Parameter} objects in the objective.

\item \code{constants}: List of \linkS4class{Constant} objects in the objective.

\item \code{is_dcp}: A logical value indicating whether the objective is convex.

\item \code{value}: The value of the objective expression.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{A scalar \linkS4class{Expression} to minimize.}
}}

