% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_problems_objective.R
\name{Minimize}
\alias{Minimize}
\title{Create a Minimization Objective}
\usage{
Minimize(expr)
}
\arguments{
\item{expr}{A CVXR expression or numeric value to minimize.}
}
\value{
A \code{Minimize} object.
}
\description{
Specifies that the objective expression should be minimized. The expression
must be convex and scalar for a DCP-compliant problem.
}
\examples{
x <- Variable()
obj <- Minimize(x^2 + 1)

}
\seealso{
\code{\link{Maximize}}, \code{\link{Problem}}
}
