% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/082_atoms_mixed_norm.R
\name{mixed_norm}
\alias{mixed_norm}
\title{Mixed norm (\eqn{L_{p,q}} norm): column-wise p-norm, then q-norm}
\usage{
mixed_norm(X, p = 2, q = 1)
}
\arguments{
\item{X}{An Expression (matrix)}

\item{p}{Inner norm parameter (default 2)}

\item{q}{Outer norm parameter (default 1)}
}
\value{
An Expression representing the mixed norm
}
\description{
Mixed norm (\eqn{L_{p,q}} norm): column-wise p-norm, then q-norm
}
