% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/294_mosek_conif.R
\docType{class}
\name{MOSEK-class}
\alias{MOSEK-class}
\alias{MOSEK}
\alias{import_solver,MOSEK-method}
\alias{name,MOSEK-method}
\alias{accepts,MOSEK,Problem-method}
\alias{perform,MOSEK,Problem-method}
\alias{solve_via_data,MOSEK-method}
\alias{invert,MOSEK,ANY,ANY-method}
\title{An interface for the MOSEK solver.}
\usage{
\S4method{import_solver}{MOSEK}(solver)

\S4method{name}{MOSEK}(x)

\S4method{accepts}{MOSEK,Problem}(object, problem)

\S4method{perform}{MOSEK,Problem}(object, problem)

\S4method{solve_via_data}{MOSEK}(
  object,
  data,
  warm_start,
  verbose,
  feastol,
  reltol,
  abstol,
  num_iter,
  solver_opts,
  solver_cache
)

\S4method{invert}{MOSEK,ANY,ANY}(object, solution, inverse_data)
}
\arguments{
\item{object, x}{A \linkS4class{MOSEK} object.}

\item{problem}{A \linkS4class{Problem} object.}

\item{data}{Data generated via an apply call.}

\item{warm_start}{A boolean of whether to warm start the solver.}

\item{verbose}{A boolean of whether to enable solver verbosity.}

\item{feastol}{The feasible tolerance.}

\item{reltol}{The relative tolerance.}

\item{abstol}{The absolute tolerance.}

\item{num_iter}{The maximum number of iterations.}

\item{solver_opts}{A list of Solver specific options}

\item{solver_cache}{Cache for the solver.}

\item{solution}{The raw solution returned by the solver.}

\item{inverse_data}{A list containing data necessary for the inversion.}
}
\description{
An interface for the MOSEK solver.
}
\section{Methods (by generic)}{
\itemize{
\item \code{import_solver(MOSEK)}: Imports the solver.

\item \code{name(MOSEK)}: Returns the name of the solver.

\item \code{accepts(object = MOSEK, problem = Problem)}: Can MOSEK solve the problem?

\item \code{perform(object = MOSEK, problem = Problem)}: Returns a new problem and data for inverting the new solution.

\item \code{solve_via_data(MOSEK)}: Solve a problem represented by data returned from apply.

\item \code{invert(object = MOSEK, solution = ANY, inverse_data = ANY)}: Returns the solution to the original problem given the inverse_data.

}}
