% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressions.R
\docType{methods}
\name{*,Expression,Expression-method}
\alias{*,Expression,Expression-method}
\alias{*,Expression,ConstVal-method}
\alias{*,ConstVal,Expression-method}
\title{Elementwise multiplication operator}
\usage{
\S4method{*}{Expression,Expression}(e1, e2)

\S4method{*}{Expression,ConstVal}(e1, e2)

\S4method{*}{ConstVal,Expression}(e1, e2)
}
\arguments{
\item{e1, e2}{The \linkS4class{Expression} objects or numeric constants to multiply elementwise.}
}
\description{
Elementwise multiplication operator
}
