% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/299_zzz_R_specific_exports.R
\name{multiply}
\alias{multiply}
\title{Elementwise multiplication (deprecated)}
\usage{
multiply(x, y)
}
\arguments{
\item{x, y}{Expressions or numeric values.}
}
\value{
An Expression representing the elementwise product.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use the \code{*} operator instead: \code{x * y}.
}
