% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/261_zzz_R_specific_exports.R
\name{name}
\alias{name}
\title{Get Expression Name}
\usage{
name(x)
}
\arguments{
\item{x}{A CVXR expression, variable, parameter, constant, or constraint.}
}
\value{
A character string.
}
\description{
Returns a human-readable string representation of a CVXR expression,
variable, or constraint.
}
\examples{
x <- Variable(2, name = "x")
name(x)  # "x"
name(x + 1)  # "x + 1"

}
