% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R, R/019_atom.R
\docType{methods}
\name{name}
\alias{name}
\alias{name,Atom-method}
\title{Variable, Parameter, or Expression Name}
\usage{
name(x)

\S4method{name}{Atom}(x)
}
\arguments{
\item{x}{A \linkS4class{Variable}, \linkS4class{Parameter}, or \linkS4class{Expression} object.}
}
\value{
For \linkS4class{Variable} or \linkS4class{Parameter} objects, the value in the name slot. For \linkS4class{Expression} objects, a string indicating the nested atoms and their respective arguments.

For \linkS4class{Variable} or \linkS4class{Parameter} objects, the value in the name slot. For \linkS4class{Expression} objects, a string indicating the nested atoms and their respective arguments.
}
\description{
The string representation of a variable, parameter, or expression.

The string representation of a variable, parameter, or expression.
}
\examples{
x <- Variable()
y <- Variable(3, name = "yVar")

name(x)
name(y)
x <- Variable()
y <- Variable(3, name = "yVar")

name(x)
name(y)
}
