% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{NonlinearConstraint-class}
\alias{NonlinearConstraint-class}
\alias{.NonlinearConstraint}
\alias{NonlinearConstraint}
\title{The NonlinearConstraint class.}
\usage{
NonlinearConstraint(f, vars_, id = NA_integer_)
}
\arguments{
\item{f}{A nonlinear function.}

\item{vars_}{A list of variables involved in the function.}

\item{id}{(Optional) An integer representing the unique ID of the contraint.}
}
\description{
This class represents a nonlinear inequality constraint, \eqn{f(x) \leq 0} where \eqn{f} is twice-differentiable.
}
\section{Slots}{

\describe{
\item{\code{f}}{A nonlinear function.}

\item{\code{vars_}}{A list of variables involved in the function.}

\item{\code{.x_dim}}{(Internal) The dimensions of a column vector with number of elements equal to the total elements in all the variables.}
}}

