% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\docType{class}
\name{NonPosConstraint-class}
\alias{NonPosConstraint-class}
\alias{.NonPosConstraint}
\alias{name,NonPosConstraint-method}
\alias{is_dcp,NonPosConstraint-method}
\alias{is_dgp,NonPosConstraint-method}
\alias{canonicalize,NonPosConstraint-method}
\alias{residual,NonPosConstraint-method}
\title{The NonPosConstraint class}
\usage{
\S4method{name}{NonPosConstraint}(x)

\S4method{is_dcp}{NonPosConstraint}(object)

\S4method{is_dgp}{NonPosConstraint}(object)

\S4method{canonicalize}{NonPosConstraint}(object)

\S4method{residual}{NonPosConstraint}(object)
}
\arguments{
\item{x, object}{A \linkS4class{NonPosConstraint} object.}
}
\description{
The NonPosConstraint class
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: The string representation of the constraint.

\item \code{is_dcp}: Is the constraint DCP?

\item \code{is_dgp}: Is the constraint DGP?

\item \code{canonicalize}: The graph implementation of the object.

\item \code{residual}: The residual of the constraint.
}}

