% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/088_atoms_norm1.R
\name{norm1}
\alias{norm1}
\title{L1 norm of an expression}
\usage{
norm1(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical: keep reduced dimensions?}
}
\value{
A Norm1 atom
}
\description{
L1 norm of an expression
}
