% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\name{Norm2}
\alias{Norm2}
\title{The Norm2 atom.}
\usage{
Norm2(x, axis = NA_real_, keepdims = FALSE)
}
\arguments{
\item{x}{An \linkS4class{Expression} object.}

\item{axis}{(Optional) The dimension across which to apply the function: \code{1} indicates rows, \code{2} indicates columns, and \code{NA} indicates rows and columns. The default is \code{NA}.}

\item{keepdims}{(Optional) Should dimensions be maintained when applying the atom along an axis? If \code{FALSE}, result will be collapsed into an \eqn{n x 1} column vector. The default is \code{FALSE}.}
}
\value{
Returns the 2-norm of x.
}
\description{
The 2-norm of an expression.
}
