% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/089_atoms_norm_inf.R
\name{norm_inf}
\alias{norm_inf}
\title{L-infinity norm of an expression}
\usage{
norm_inf(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical: keep reduced dimensions?}
}
\value{
A NormInf atom
}
\description{
L-infinity norm of an expression
}
