% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/105_atoms_norm_nuc.R
\name{norm_nuc}
\alias{norm_nuc}
\title{Nuclear norm (sum of singular values)}
\usage{
norm_nuc(A)
}
\arguments{
\item{A}{A matrix expression}
}
\value{
An expression representing the nuclear norm of A
}
\description{
Nuclear norm (sum of singular values)
}
