% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{numeric_value}
\alias{numeric_value}
\title{Compute the Numeric Value of an Atom}
\usage{
numeric_value(x, values, ...)
}
\arguments{
\item{x}{An atom object.}

\item{values}{List of numeric values of the atom's arguments.}

\item{...}{Additional arguments.}
}
\value{
Numeric value.
}
\description{
Compute the Numeric Value of an Atom
}
\keyword{internal}
