% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/114_atoms_one_minus_pos.R
\name{one_minus_pos}
\alias{one_minus_pos}
\title{The difference 1 - x with domain (0, 1)}
\usage{
one_minus_pos(x)
}
\arguments{
\item{x}{An Expression (elementwise in (0, 1))}
}
\value{
A OneMInusPos atom
}
\description{
Log-log concave atom for DGP. Solve with
\code{psolve(problem, gp = TRUE)}.
}
\examples{
x <- Variable(pos = TRUE)
prob <- Problem(Maximize(one_minus_pos(x)), list(x >= 0.1, x <= 0.5))
\dontrun{psolve(prob, gp = TRUE)}
}
