% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{partial_optimize}
\alias{partial_optimize}
\title{Partially optimizes the given problem over the specified variables.}
\usage{
partial_optimize(
  prob,
  opt_vars = list(),
  dont_opt_vars = list(),
  solver = NA,
  ...
)
}
\arguments{
\item{prob}{The problem to partially optimize.}

\item{opt_vars}{The variables to optimize over.}

\item{dont_opt_vars}{The variables to not optimize over.}

\item{solver}{The default solver to use for value and grad.}

\item{...}{Additional solver specific keyword arguments.}
}
\value{
An expression representing the partial optimization.
Convex for minimization objectives and concave for maximization objectives.
}
\description{
Either opt_vars or dont_opt_vars must be given.
If both are given, they must contain all the variables in the problem.
}
\details{
Partial optimize is useful for two-stage optimization and graph implementations.
For example, we can write

\if{html}{\out{<div class="sourceCode">}}\preformatted{   x <- Variable(n)
   t <- Variable(n)
   abs_x <- partial_optimize(Problem(Minimize(sum(t)),
                             list(-t <= x, x <= t)), opt_vars = list(t))
}\if{html}{\out{</div>}}

to define the entrywise absolute value of x.
}
