% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementwise.R
\name{Pos}
\alias{Pos}
\title{An alias for MaxElemwise(x, 0)}
\usage{
Pos(x)
}
\arguments{
\item{x}{An R numeric value or \linkS4class{Expression}.}
}
\value{
An alias for MaxElemwise(x, 0)
}
\description{
An alias for MaxElemwise(x, 0)
}
