% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/132_problems_problem.R
\name{Problem}
\alias{Problem}
\title{Create an Optimization Problem}
\usage{
Problem(objective, constraints = list())
}
\arguments{
\item{objective}{A \code{\link{Minimize}} or \code{\link{Maximize}} object.}

\item{constraints}{A list of \code{Constraint} objects (e.g., created by
\code{==}, \code{<=}, \code{>=} operators on expressions). Defaults to
an empty list (unconstrained).}
}
\value{
A \code{Problem} object.
}
\description{
Constructs a convex optimization problem from an objective and a list of
constraints. Use \code{\link{psolve}} to solve the problem.
}
\section{Known limitations}{

\itemize{
\item Problems must contain at least one \code{\link{Variable}}.
Zero-variable problems (e.g., minimizing a constant) will cause an
internal error in the reduction pipeline.
}
}

\examples{
x <- Variable(2)
prob <- Problem(Minimize(sum_entries(x)), list(x >= 1))

}
