% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/145_problem.R
\docType{methods}
\name{psolve,Problem-method}
\alias{psolve,Problem-method}
\title{This internal method solves a DCP compliant optimization
problem. It saves the values of primal and dual variables in the
Variable and Constraint objects, respectively.}
\usage{
\S4method{psolve}{Problem}(
  object,
  solver = NA_character_,
  ignore_dcp = FALSE,
  warm_start = TRUE,
  verbose = FALSE,
  gp = FALSE,
  qcp = FALSE,
  requires_grad = FALSE,
  enforce_dpp = FALSE,
  ignore_dpp = FALSE,
  canon_backend = NA_character_,
  parallel = FALSE,
  feastol = NULL,
  reltol = NULL,
  abstol = NULL,
  num_iter = NULL,
  low = NA_real_,
  high = NA_real_,
  ...
)
}
\description{
This internal method solves a DCP compliant optimization
problem. It saves the values of primal and dual variables in the
Variable and Constraint objects, respectively.
}
