% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/141_problems_problem.R
\name{problem_solution}
\alias{problem_solution}
\title{Get the Raw Solution Object (deprecated)}
\usage{
problem_solution(x)
}
\arguments{
\item{x}{A \code{\link{Problem}} object.}
}
\value{
A \code{Solution} object, or \code{NULL} if the problem
has not been solved.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link{solution}} instead.
}
\seealso{
\code{\link{solution}}
}
