% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/141_problems_problem.R
\name{problem_status}
\alias{problem_status}
\title{Get the Solution Status of a Problem (deprecated)}
\usage{
problem_status(x)
}
\arguments{
\item{x}{A \code{\link{Problem}} object.}
}
\value{
Character string, or \code{NULL} if the problem has not been
solved.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Use \code{\link{status}} instead.
}
\seealso{
\code{\link{status}}
}
