% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/117_atoms_prod.R
\name{prod_entries}
\alias{prod_entries}
\title{Product of entries along an axis}
\usage{
prod_entries(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Whether to keep reduced dimensions}
}
\value{
A Prod atom
}
\description{
Used in DGP (geometric programming) context. Solve with
\code{psolve(problem, gp = TRUE)}.
}
\examples{
x <- Variable(3, pos = TRUE)
prob <- Problem(Minimize(prod_entries(x)), list(x >= 2))
\dontrun{psolve(prob, gp = TRUE)}
}
