% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{project}
\alias{project}
\title{Project a Value onto the Domain of a Leaf}
\usage{
project(x, val, ...)
}
\arguments{
\item{x}{A leaf expression object.}

\item{val}{The value to project.}

\item{...}{Additional arguments.}
}
\value{
The projected value.
}
\description{
Project a Value onto the Domain of a Leaf
}
\keyword{internal}
