% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp2quad_form.R
\docType{class}
\name{QpMatrixStuffing-class}
\alias{QpMatrixStuffing-class}
\alias{QpMatrixStuffing}
\title{The QpMatrixStuffing class.}
\description{
This class fills in numeric values for the problem instance and
outputs a DCP-compliant minimization problem with an objective
of the form
}
\details{
QuadForm(x, p) + t(q) %*% x

and Zero/NonPos constraints, both of which exclusively carry
affine arguments
}
