% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_solvers.R
\docType{class}
\name{QpSolver-class}
\alias{QpSolver-class}
\alias{accepts,QpSolver,Problem-method}
\alias{perform,QpSolver,Problem-method}
\title{A QP solver interface.}
\usage{
\S4method{accepts}{QpSolver,Problem}(object, problem)

\S4method{perform}{QpSolver,Problem}(object, problem)
}
\arguments{
\item{object}{A \linkS4class{QpSolver} object.}

\item{problem}{A \linkS4class{Problem} object.}
}
\description{
A QP solver interface.
}
\section{Methods (by generic)}{
\itemize{
\item \code{accepts}: Is this a QP problem?

\item \code{perform}: Constructs a QP problem data stored in a list
}}

