% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/097_atoms_quad_over_lin.R
\name{quad_over_lin}
\alias{quad_over_lin}
\title{Sum of squares divided by a scalar}
\usage{
quad_over_lin(x, y, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{y}{An Expression (scalar, positive)}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical: keep reduced dimensions?}
}
\value{
A QuadOverLin atom
}
\description{
Sum of squares divided by a scalar
}
