% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{reduction_accepts}
\alias{reduction_accepts}
\title{Check if a Reduction Accepts a Problem}
\usage{
reduction_accepts(x, problem, ...)
}
\arguments{
\item{x}{A Reduction object.}

\item{problem}{A Problem object.}

\item{...}{Additional arguments.}
}
\value{
Logical scalar.
}
\description{
Check if a Reduction Accepts a Problem
}
\keyword{internal}
