% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_atoms_elementwise_rel_entr.R
\name{rel_entr}
\alias{rel_entr}
\title{Relative Entropy: x*log(x/y)}
\usage{
rel_entr(x, y)
}
\arguments{
\item{x}{An Expression}

\item{y}{An Expression}
}
\value{
A RelEntr atom
}
\description{
Relative Entropy: x*log(x/y)
}
