% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{resetOptions}
\alias{resetOptions}
\title{Reset Options}
\usage{
resetOptions()
}
\value{
The default value of CVXR package global \code{.CVXR.options}.
}
\description{
Reset the global package variable \code{.CVXR.options}.
}
\examples{
\dontrun{
  resetOptions()
}
}
