% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/115_atoms_eye_minus_inv.R
\name{resolvent}
\alias{resolvent}
\title{Resolvent inverse(sI - X)}
\usage{
resolvent(X, s)
}
\arguments{
\item{X}{An Expression (positive square matrix)}

\item{s}{A positive scalar}
}
\value{
An expression for the resolvent
}
\description{
Equivalent to (1/s) * eye_minus_inv(X / s).
}
