% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/045_atoms_affine_binary_operators.R
\name{scalar_product}
\alias{scalar_product}
\title{Scalar product (alias for vdot)}
\usage{
scalar_product(x, y)
}
\arguments{
\item{x}{An Expression, numeric value, or nested list thereof. The
conjugate-linear argument.}

\item{y}{An Expression, numeric value, or nested list thereof. The linear
argument.}
}
\value{
A scalar Expression representing \code{sum(Conj(x) * y)}.
}
\description{
Scalar product (alias for vdot)
}
\seealso{
\code{\link[=vdot]{vdot()}}
}
