% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/079_atoms_elementwise_scalene.R
\name{scalene}
\alias{scalene}
\title{Scalene penalty: alpha * pos(x) + beta * neg(x)}
\usage{
scalene(x, alpha, beta)
}
\arguments{
\item{x}{An Expression}

\item{alpha}{Coefficient for the positive part}

\item{beta}{Coefficient for the negative part}
}
\value{
An Expression representing the scalene penalty
}
\description{
Scalene penalty: alpha * pos(x) + beta * neg(x)
}
