% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/287_scs_conif.R
\name{SCS.psdvec_to_psdmat}
\alias{SCS.psdvec_to_psdmat}
\title{Return \code{V} so that \emph{vec\link{indices} belongs to the SCS-standard PSD
cone} can be written in natural cvxpy syntax as \emph{V >> 0}.}
\usage{
SCS.psdvec_to_psdmat(vec, indices)
}
\arguments{
\item{vec}{list of affine expressions}

\item{indices}{contains nonnegative integers, which can index into
\code{vec}}
}
\description{
Return \code{V} so that \emph{vec\link{indices} belongs to the SCS-standard PSD
cone} can be written in natural cvxpy syntax as \emph{V >> 0}.
}
\details{
This function is similar to \code{tri_to_full}, which is also
found in this file. The difference is that this function works
without indexed assignment \code{mat[i,j] = expr}. Such indexed
assignment cannot be used, because this function builds a CVXR
expression, rather than a numeric matrix.
}
