% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atoms.R
\docType{class}
\name{SigmaMax-class}
\alias{SigmaMax-class}
\alias{.SigmaMax}
\alias{SigmaMax}
\alias{to_numeric,SigmaMax-method}
\alias{size_from_args,SigmaMax-method}
\alias{sign_from_args,SigmaMax-method}
\alias{is_atom_convex,SigmaMax-method}
\alias{is_atom_concave,SigmaMax-method}
\alias{is_incr,SigmaMax-method}
\alias{is_decr,SigmaMax-method}
\alias{graph_implementation,SigmaMax-method}
\title{The SigmaMax class.}
\usage{
SigmaMax(A = A)

\S4method{to_numeric}{SigmaMax}(object, values)

\S4method{size_from_args}{SigmaMax}(object)

\S4method{sign_from_args}{SigmaMax}(object)

\S4method{is_atom_convex}{SigmaMax}(object)

\S4method{is_atom_concave}{SigmaMax}(object)

\S4method{is_incr}{SigmaMax}(object, idx)

\S4method{is_decr}{SigmaMax}(object, idx)

\S4method{graph_implementation}{SigmaMax}(object, arg_objs, size, data = NA_real_)
}
\arguments{
\item{A}{An \linkS4class{Expression} or matrix.}

\item{object}{A \linkS4class{SigmaMax} object.}

\item{values}{A list of arguments to the atom.}

\item{idx}{An index into the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
The maximum singular value of a matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric}: The largest singular value of \code{A}.

\item \code{size_from_args}: The atom is a scalar.

\item \code{sign_from_args}: The atom is positive.

\item \code{is_atom_convex}: The atom is convex.

\item \code{is_atom_concave}: The atom is concave.

\item \code{is_incr}: The atom is not monotonic in any argument.

\item \code{is_decr}: The atom is not monotonic in any argument.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{A}}{An \linkS4class{Expression} or numeric matrix.}
}}

