% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{sign_from_args}
\alias{sign_from_args}
\title{Infer Sign from Arguments}
\usage{
sign_from_args(x, ...)
}
\arguments{
\item{x}{An atom object.}

\item{...}{Not used.}
}
\value{
Character string: sign constant.
}
\description{
Infer Sign from Arguments
}
\keyword{internal}
