% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/atoms.R
\name{sign_from_args}
\alias{sign_from_args}
\alias{sign_from_args,Atom-method}
\title{Atom Sign}
\usage{
sign_from_args(object)

\S4method{sign_from_args}{Atom}(object)
}
\arguments{
\item{object}{An \linkS4class{Atom} object.}
}
\value{
A logical vector \code{c(is positive, is negative)} indicating the sign of the atom.
}
\description{
Determine the sign of an atom based on its arguments.
}
