% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R, R/007_utility.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,ListORExpr-method}
\title{Size of Expression}
\usage{
size(object)

\S4method{size}{ListORExpr}(object)
}
\arguments{
\item{object}{An \linkS4class{Expression} object.}
}
\value{
A vector with two elements \code{c(row, col)} representing the dimensions of the expression.
}
\description{
The size of an expression.
}
\examples{
x <- Variable()
y <- Variable(3)
z <- Variable(3,2)

size(x)
size(y)
size(z)
size(x + y)
size(z - x)
}
