% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012_settings.R
\docType{data}
\name{solver-constants}
\alias{solver-constants}
\alias{SCS_SOLVER}
\alias{OSQP_SOLVER}
\alias{CLARABEL_SOLVER}
\alias{DIFFCP_SOLVER}
\alias{HIGHS_SOLVER}
\alias{MOSEK_SOLVER}
\alias{GUROBI_SOLVER}
\alias{GLPK_SOLVER}
\alias{GLPK_MI_SOLVER}
\alias{ECOS_SOLVER}
\alias{ECOS_BB_SOLVER}
\alias{CPLEX_SOLVER}
\alias{CVXOPT_SOLVER}
\alias{PIQP_SOLVER}
\alias{SCIP_SOLVER}
\alias{XPRESS_SOLVER}
\title{Solver Name Constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
SCS_SOLVER

OSQP_SOLVER

CLARABEL_SOLVER

DIFFCP_SOLVER

HIGHS_SOLVER

MOSEK_SOLVER

GUROBI_SOLVER

GLPK_SOLVER

GLPK_MI_SOLVER

ECOS_SOLVER

ECOS_BB_SOLVER

CPLEX_SOLVER

CVXOPT_SOLVER

PIQP_SOLVER

SCIP_SOLVER

XPRESS_SOLVER
}
\value{
A character string.
}
\description{
Character string constants identifying the available solvers.
}
\keyword{datasets}
