% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{solver_name}
\alias{solver_name}
\title{Get Solver Name}
\usage{
solver_name(x, ...)
}
\arguments{
\item{x}{A Solver object.}

\item{...}{Not used.}
}
\value{
Character string with solver name.
}
\description{
Get Solver Name
}
\keyword{internal}
