% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/075_atoms_elementwise_square.R
\name{square}
\alias{square}
\title{Square of an expression: x^2}
\usage{
square(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A \code{PowerApprox} atom with p=2, the same atom \code{power(x, 2)} builds
}
\description{
Square of an expression: x^2
}
