% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/069_atoms_elementwise_square.R
\name{square}
\alias{square}
\title{Square of an expression: x^2}
\usage{
square(x)
}
\arguments{
\item{x}{An Expression}
}
\value{
A Power atom with p=2
}
\description{
Square of an expression: x^2
}
