% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012_settings.R
\docType{data}
\name{status-constants}
\alias{status-constants}
\alias{OPTIMAL}
\alias{INFEASIBLE}
\alias{UNBOUNDED}
\alias{SOLVER_ERROR}
\alias{OPTIMAL_INACCURATE}
\alias{INFEASIBLE_INACCURATE}
\alias{UNBOUNDED_INACCURATE}
\alias{USER_LIMIT}
\alias{INFEASIBLE_OR_UNBOUNDED}
\title{Solution Status Constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
OPTIMAL

INFEASIBLE

UNBOUNDED

SOLVER_ERROR

OPTIMAL_INACCURATE

INFEASIBLE_INACCURATE

UNBOUNDED_INACCURATE

USER_LIMIT

INFEASIBLE_OR_UNBOUNDED
}
\value{
A character string.
}
\description{
Character string constants representing the possible solution statuses
returned by \code{\link{problem_status}}.
}
\seealso{
\code{\link{status}}, \code{\link{psolve}}
}
\keyword{datasets}
