% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/042_atoms_affine_sum.R
\name{sum_entries}
\alias{sum_entries}
\title{Sum the entries of an expression}
\usage{
sum_entries(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression or numeric value.}

\item{axis}{NULL (sum all), 1 (row-wise, like apply(X,1,sum)), or 2 (column-wise, like apply(X,2,sum)).}

\item{keepdims}{Logical: if TRUE, keep the reduced dimension as size 1.}
}
\value{
A SumEntries expression.
}
\description{
Sum the entries of an expression
}
