% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/099_atoms_sum_largest.R
\name{sum_largest}
\alias{sum_largest}
\title{Sum of k largest entries}
\usage{
sum_largest(x, k)
}
\arguments{
\item{x}{An Expression}

\item{k}{Number of largest entries to sum}
}
\value{
A SumLargest atom
}
\description{
Sum of k largest entries
}
