% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/089_atoms_sum_smallest.R
\name{sum_smallest}
\alias{sum_smallest}
\title{Sum of k smallest entries}
\usage{
sum_smallest(x, k, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{k}{Number of smallest entries to sum}

\item{axis}{NULL (all entries), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical; keep the reduced dimension as size 1}
}
\value{
An Expression equal to -SumLargest(-x, k)
}
\description{
Sum of k smallest entries
}
