% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/080_atoms_sum_smallest.R
\name{sum_smallest}
\alias{sum_smallest}
\title{Sum of k smallest entries}
\usage{
sum_smallest(x, k)
}
\arguments{
\item{x}{An Expression}

\item{k}{Number of smallest entries to sum}
}
\value{
An Expression equal to -SumLargest(-x, k)
}
\description{
Sum of k smallest entries
}
