% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/097_atoms_quad_over_lin.R
\name{sum_squares}
\alias{sum_squares}
\title{Sum of squares (= quad_over_lin(x, 1))}
\usage{
sum_squares(x, axis = NULL, keepdims = FALSE)
}
\arguments{
\item{x}{An Expression}

\item{axis}{NULL (all), 1 (row-wise), or 2 (column-wise)}

\item{keepdims}{Logical: keep reduced dimensions?}
}
\value{
A QuadOverLin atom
}
\description{
Sum of squares (= quad_over_lin(x, 1))
}
