% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conic_solvers.R
\name{name,SuperSCS-method}
\alias{name,SuperSCS-method}
\alias{solve_via_data,SuperSCS-method}
\title{The SuperSCS class}
\usage{
\S4method{name}{SuperSCS}(x)

\S4method{solve_via_data}{SuperSCS}(
  object,
  data,
  warm_start,
  verbose,
  solver_opts,
  solver_cache = list()
)
}
\arguments{
\item{data}{Data generated via an apply call.}

\item{warm_start}{An option for warm start.}

\item{verbose}{A boolean of whether to enable solver verbosity.}

\item{solver_opts}{A list of Solver specific options}

\item{solver_cache}{Cache for the solver.}
}
\description{
The SuperSCS class
}
\section{Methods (by generic)}{
\itemize{
\item \code{solve_via_data}: Solve a problem represented by data returned from apply.
}}

