% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/256_zzz_R_specific_to_latex.R
\name{to_latex}
\alias{to_latex}
\title{Convert CVXR Object to LaTeX}
\usage{
to_latex(x, ...)
}
\arguments{
\item{x}{A \code{Problem}, \code{Expression}, \code{Constraint}, or
\code{Objective}.}

\item{...}{Reserved for future options.}
}
\value{
A character string containing LaTeX code.
}
\description{
Renders a CVXR \code{Problem}, \code{Expression}, or \code{Constraint}
as a LaTeX string.  Problem-level output uses the \code{optidef} package
(\code{mini*}/\code{maxi*} environments) and atom macros from
\code{dcp.sty} (shipped in \code{inst/latex/dcp.sty}).
}
\examples{
x <- Variable(3, name = "x")
cat(to_latex(p_norm(x, 2)))
# \cvxnorm{x}_2

}
