% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087_atoms_total_variation.R
\name{total_variation}
\alias{total_variation}
\title{Total variation of a vector or matrix}
\usage{
total_variation(value, ...)
}
\arguments{
\item{value}{An Expression or numeric constant (vector or matrix)}

\item{...}{Additional matrix expressions extending the third dimension}
}
\value{
An Expression representing the total variation
}
\description{
Computes total variation using L1 norm of discrete gradients for vectors
and L2 norm of discrete gradients for matrices.
}
