% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/107_atoms_tr_inv.R
\name{tr_inv}
\alias{tr_inv}
\title{Trace of matrix inverse}
\usage{
tr_inv(X)
}
\arguments{
\item{X}{A square PSD matrix expression}
}
\value{
An expression representing \eqn{\mathrm{tr}(X^{-1})}
}
\description{
Computes \eqn{\mathrm{tr}(X^{-1})} for PSD matrix X.
}
