% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/044_trace.R
\docType{class}
\name{Trace-class}
\alias{Trace-class}
\alias{.Trace}
\alias{Trace}
\alias{to_numeric,Trace-method}
\alias{validate_args,Trace-method}
\alias{dim_from_args,Trace-method}
\alias{sign_from_args,Trace-method}
\alias{is_real,Trace-method}
\alias{is_complex,Trace-method}
\alias{is_atom_log_log_convex,Trace-method}
\alias{is_atom_log_log_concave,Trace-method}
\alias{graph_implementation,Trace-method}
\title{The Trace class.}
\usage{
Trace(expr)

\S4method{to_numeric}{Trace}(object, values)

\S4method{validate_args}{Trace}(object)

\S4method{dim_from_args}{Trace}(object)

\S4method{sign_from_args}{Trace}(object)

\S4method{is_real}{Trace}(object)

\S4method{is_complex}{Trace}(object)

\S4method{is_atom_log_log_convex}{Trace}(object)

\S4method{is_atom_log_log_concave}{Trace}(object)

\S4method{graph_implementation}{Trace}(object, arg_objs, dim, data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} representing a matrix.}

\item{object}{A \linkS4class{Trace} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{dim}{A vector representing the dimensions of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
This class represents the sum of the diagonal entries in a matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{to_numeric(Trace)}: Sum the diagonal entries.

\item \code{validate_args(Trace)}: Check the argument is a square matrix.

\item \code{dim_from_args(Trace)}: The atom is a scalar.

\item \code{sign_from_args(Trace)}: The (is positive, is negative) sign of the atom.

\item \code{is_real(Trace)}: A logical value indicating whether the atom is real.

\item \code{is_complex(Trace)}: A logical value indicating whether the atom is complex.

\item \code{is_atom_log_log_convex(Trace)}: Is the atom log-log convex?

\item \code{is_atom_log_log_concave(Trace)}: Is the atom log-log concave?

\item \code{graph_implementation(Trace)}: The graph implementation of the atom.

}}
\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} representing a matrix.}
}}

