% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{tree_copy}
\alias{tree_copy}
\title{Deep Copy of an Expression Tree}
\usage{
tree_copy(x, id_objects = NULL)
}
\arguments{
\item{x}{A canonicalizable object.}

\item{id_objects}{Optional identity map for deduplication.}
}
\value{
A deep copy of the entire expression tree.
}
\description{
Deep Copy of an Expression Tree
}
\keyword{internal}
