% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/287_scs_conif.R
\name{tri_to_full}
\alias{tri_to_full}
\title{Expands n*floor((n+1)/2) lower triangular to full matrix and scales
off-diagonal by 1/sqrt(2), as per the SCS specification.}
\usage{
tri_to_full(lower_tri, n)
}
\arguments{
\item{lower_tri}{A matrix representing the lower triangular part of
the matrix, stacked in column-major order}

\item{n}{The number of rows (columns) in the full square matrix.}
}
\value{
A matrix that is the scaled expansion of the lower
triangular matrix.
}
\description{
Expands n*floor((n+1)/2) lower triangular to full matrix and scales
off-diagonal by 1/sqrt(2), as per the SCS specification.
}
