% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{UnaryOperator-class}
\alias{UnaryOperator-class}
\alias{UnaryOperator}
\title{The UnaryOperator class.}
\description{
This base class represents expressions involving unary operators.
}
\section{Slots}{

\describe{
\item{\code{expr}}{The \linkS4class{Expression} that is being operated upon.}

\item{\code{op_name}}{A \code{character} string indicating the unary operation.}
}}

