% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{update_parameters}
\alias{update_parameters}
\title{Update Parameters for DPP Fast Path}
\usage{
update_parameters(x, problem, ...)
}
\arguments{
\item{x}{A Reduction object.}

\item{problem}{A Problem object.}

\item{...}{Additional arguments.}
}
\value{
NULL (called for side effects).
}
\description{
Update Parameters for DPP Fast Path
}
\keyword{internal}
