% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/051_atoms_affine_upper_tri.R
\name{upper_tri}
\alias{upper_tri}
\title{Extract strict upper triangle of a square matrix}
\usage{
upper_tri(x)
}
\arguments{
\item{x}{An Expression (square matrix)}
}
\value{
An UpperTri atom (column vector)
}
\description{
Extract strict upper triangle of a square matrix
}
