% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/306_exports.R
\docType{methods}
\name{upper_tri}
\alias{upper_tri}
\title{Upper Triangle of a Matrix}
\usage{
upper_tri(expr)
}
\arguments{
\item{expr}{An \linkS4class{Expression} or matrix.}
}
\value{
An \linkS4class{Expression} representing the upper triangle of the input.
}
\description{
The vectorized strictly upper triangular entries of a matrix.
}
\examples{
C <- Variable(3,3)
val <- cbind(3:5, 6:8, 9:11)
prob <- Problem(Maximize(upper_tri(C)[3,1]), list(C == val))
result <- solve(prob)
result$value
result$getValue(upper_tri(C))
}
