% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{UpperTri-class}
\alias{UpperTri-class}
\alias{.UpperTri}
\alias{UpperTri}
\alias{validate_args,UpperTri-method}
\alias{to_numeric,UpperTri-method}
\alias{size_from_args,UpperTri-method}
\alias{graph_implementation,UpperTri-method}
\title{The UpperTri class.}
\usage{
UpperTri(expr)

\S4method{validate_args}{UpperTri}(object)

\S4method{to_numeric}{UpperTri}(object, values)

\S4method{size_from_args}{UpperTri}(object)

\S4method{graph_implementation}{UpperTri}(object, arg_objs, size,
  data = NA_real_)
}
\arguments{
\item{expr}{An \linkS4class{Expression} or numeric matrix.}

\item{object}{An \linkS4class{UpperTri} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
The vectorized strictly upper triagonal entries of a matrix.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args}: Check the argument is a square matrix.

\item \code{to_numeric}: Vectorize the upper triagonal entries.

\item \code{size_from_args}: The size of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{expr}}{An \linkS4class{Expression} or numeric matrix.}
}}

