% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_generics.R, R/019_atom.R
\docType{methods}
\name{validate_args}
\alias{validate_args}
\alias{validate_args,Atom-method}
\title{Validate Arguments}
\usage{
validate_args(object)

\S4method{validate_args}{Atom}(object)
}
\arguments{
\item{object}{an \linkS4class{Atom} object}
}
\description{
Validate an atom's arguments, returning an error if any are invalid.

Validate an atom's arguments, returning an error if any are invalid.
}
