% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{value<-}
\alias{value<-}
\title{Set the Value of a Leaf Expression}
\usage{
value(x) <- value
}
\arguments{
\item{x}{A leaf expression object.}

\item{value}{The value to assign.}
}
\value{
The modified object (invisibly).
}
\description{
Assigns a numeric value to a \code{\link{Variable}} or
\code{\link{Parameter}}.
}
