% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/012_variable.R
\docType{class}
\name{Variable-class}
\alias{Variable-class}
\alias{.Variable}
\alias{Variable}
\alias{name,Variable-method}
\alias{grad,Variable-method}
\alias{variables,Variable-method}
\alias{canonicalize,Variable-method}
\alias{attributes_were_lowered,Variable-method}
\alias{set_variable_of_provenance,Variable,Variable-method}
\alias{variable_of_provenance,Variable-method}
\alias{as.character,Variable-method}
\title{The Variable class.}
\usage{
Variable(rows = 1, cols = 1, name = NA_character_, var_id = NA_integer_, ...)

\S4method{name}{Variable}(x)

\S4method{grad}{Variable}(object)

\S4method{variables}{Variable}(object)

\S4method{canonicalize}{Variable}(object)

\S4method{attributes_were_lowered}{Variable}(object)

\S4method{set_variable_of_provenance}{Variable,Variable}(object, variable)

\S4method{variable_of_provenance}{Variable}(object)

\S4method{as.character}{Variable}(x)
}
\arguments{
\item{rows}{The number of rows in the variable.}

\item{cols}{The number of columns in the variable.}

\item{name}{(Optional) A character string representing the name of the variable.}

\item{var_id}{(Optional) A unique ID for the variable.}

\item{...}{(Optional) Additional attribute arguments. See \linkS4class{Leaf} for details.}

\item{x}{A \linkS4class{Variable} object.}
}
\description{
This class represents an optimization variable.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name(Variable)}: The name of the variable

\item \code{grad(Variable)}: The sub/super-gradient of the variable represented as a sparse matrix

\item \code{variables(Variable)}: Returns itself as a variable

\item \code{canonicalize(Variable)}: The graph implementation of the object

\item \code{attributes_were_lowered(Variable)}: Returns TRUE iff variable generated when lowering a variable with attributes

\item \code{set_variable_of_provenance(object = Variable, variable = Variable)}: Returns the variable of provenance

\item \code{variable_of_provenance(Variable)}: Returns a variable with attributes from which this variable was generated

\item \code{as.character(Variable)}: Convert to a character representation

}}
\section{Slots}{

\describe{
\item{\code{dim}}{The dimensions of the variable.}

\item{\code{name}}{(Optional) A character string representing the name of the variable.}
}}

\examples{
x <- Variable(3, name = "x0") ## 3-int variable
y <- Variable(3, 3, name = "y0") # Matrix variable
as.character(y)
id(y)
is_nonneg(x)
is_nonpos(x)
size(y)
name(y)
value(y) <- matrix(1:9, nrow = 3)
value(y)
grad(y)
variables(y)
canonicalize(y)
}
