% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/004_zzz_R_specific_generics.R
\name{variables}
\alias{variables}
\title{Get the Variables in an Expression}
\usage{
variables(x, ...)
}
\arguments{
\item{x}{An expression or problem object.}

\item{...}{Not used.}
}
\value{
List of \code{\link{Variable}} objects.
}
\description{
Get the Variables in an Expression
}
