% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/087_atoms_affine_vec.R
\name{vec}
\alias{vec}
\title{Vectorize an expression (column vector)}
\usage{
vec(x)
}
\arguments{
\item{x}{An Expression or numeric value.}
}
\value{
A Reshape atom (column vector).
}
\description{
Reshapes an expression into a column vector of shape (n*m, 1).
}
