% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine.R
\docType{class}
\name{VStack-class}
\alias{VStack-class}
\alias{.VStack}
\alias{VStack}
\alias{validate_args,VStack-method}
\alias{to_numeric,VStack-method}
\alias{size_from_args,VStack-method}
\alias{graph_implementation,VStack-method}
\title{The VStack class.}
\usage{
VStack(...)

\S4method{validate_args}{VStack}(object)

\S4method{to_numeric}{VStack}(object, values)

\S4method{size_from_args}{VStack}(object)

\S4method{graph_implementation}{VStack}(object, arg_objs, size, data = NA_real_)
}
\arguments{
\item{...}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of columns.}

\item{object}{A \linkS4class{VStack} object.}

\item{values}{A list of arguments to the atom.}

\item{arg_objs}{A list of linear expressions for each argument.}

\item{size}{A vector with two elements representing the size of the resulting expression.}

\item{data}{A list of additional data required by the atom.}
}
\description{
Vertical concatenation of values.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_args}: Check all arguments have the same width.

\item \code{to_numeric}: Vertically concatenate the values using \code{rbind}.

\item \code{size_from_args}: The size of the atom.

\item \code{graph_implementation}: The graph implementation of the atom.
}}

\section{Slots}{

\describe{
\item{\code{...}}{\linkS4class{Expression} objects or matrices. All arguments must have the same number of columns.}
}}

